Bug 952127 - notify source map subscribers after pretty-printing; r=bgrins
Intercept the applySourceMap source map worker request, so that when a
source is pretty-printed, source map subscribers can be updated.

That this does not yet handle pretty-printing original sources.  This
isn't supported yet by the debugger, and since the plan is to handle it
by augmenting the existing source map, it should be easy to fix this
code when it is implemented.

The mochitest is included here for testing but I am going to land it
upstream as well.

MozReview-Commit-ID: 3Lp1ikO8IzZ

UltraBlame original commit: 05569ef2d175e7b3814141af3cd8085ce0898ecf