Bug 1396951 - 3. Move GeckoAppShell.viewSizeChanged() to GeckoLayerClient; r=rbarker
Let `GeckoLayerClient` directly deal with scroll-to-input on resize,
instead of going through `GeckoAppShell` and relying on
`getLayerView()`. This is a necessary fix to let us remove
`getLayerView()`, and in a follow-up bug we should actually fix
scroll-to-input to work on any GeckoView.

MozReview-Commit-ID: 1xsHh2vg08M

UltraBlame original commit: 155c1005c388da8a8a9b5f45d148905eba542a95