Bug 1393366 - Increase the marionette connect timeout in wptrunner, r=maja_zf
This switches to using a timeout of 120s * timeout_multplier, which is
twice the current value. The Marionette startup_timout value is
removed because it turns out that's only used when marionette starts
the Firefox instance and therefore is ignored in this case.

MozReview-Commit-ID: 4VA8yB6M1d5

UltraBlame original commit: 7f3cc80412c6d80d3d789e44dd6ccb8a4faa81c9