Bug 1403325 - Use a single --sandbox-read-whitelist argument in mochitest and reftest, r=Alex_Gaynor
The main bug this fixes is that on reftest, the objdir needs to be added to the
whitelist on Windows. However, this only happens when running on Linux for some
reason.

Changing the --work-path and --obj-path arguments to --sandbox-read-whitelist
was more of a drive-by cleanup than anything necessary.

MozReview-Commit-ID: Dq8ZLETMzeM

UltraBlame original commit: 35460a8966c7e32c45407cb9a2a333d20b120875