Bug 1404124 - Pickle account right after it's created r=eoger
This is a relative of Bug 988605, with an exception that instead of going the whole way
and ensuring pickled data is kept up-to-date as Nick proposed, this patch simply ensures that
we pickle as soon as possible, with a goal of eliminating pickle races. The end goal is to kill
off pickling entirely, and so the assumption here is that this workaround is good enough
in the meantime.

MozReview-Commit-ID: 7IjRH7KE2Z9

UltraBlame original commit: 6c5c7e9dd360c0f9aa2507f7faa0d1b155568aa3