Bug 1402170 - Drop invalid test for offsets; r=hiro
There is a test that assumes that an offset specified on a property-indexed
keyframe is applied to all generated keyframes but that behavior is not (yet)
specified.

This behavior will be specified in [1] but until that happens it seems invalid
to test for it. Furthermore, when that is specified we will need much more
thorough tests than this one.

[1] https://github.com/w3c/web-animations/issues/148

MozReview-Commit-ID: HUUw88dg2P7

UltraBlame original commit: 204dc1bf9d016d134868fdb718092151c1a53fa7