Bug 1405348 - eliminate extraneous AddRef when moving nsIdentifierMapEntry; r=mystor
Using GetKey() in the move constructor for nsIdentifierMapEntry meant
that we were Move'ing a `const AtomOrString&`, which gives a return type
of `const AtomOrString&&`, which is nonsensical.  But C++, in its
wisdom, selects the copy constructor to copy that type, rather than the
intended move constructor.

To avoid this, just eliminate GetKey() and perform the move directly.

UltraBlame original commit: 992cd5f41ca223535bde16f691cd997d54107e7b