Bug 1405195 - Prevent content to specify loadingprincipal on the node. r=baku
This is a follow-up for bug 1376971, bail out early if the node is
loaded by content.

This is to prevent content side could specify arbitrary principal on the
node. Originally bug 1376971 added a MOZ_ASSERT for this, but it's an
no-op in opt build.

UltraBlame original commit: 90b9ac6020e9301db7c064ce641ad961b8d1134e