Bug 1399886 - remove unnecessary invertable CSS classes on devtools icons;r=gl
Using fill instead of filter we don't need to define each icon as
invertable or not. If the icon is a SVG and supports fill="context-fill"
then it will be inverted/highlighted etc... as expected.

If not then it won't be impacted by DevTools themes.

MozReview-Commit-ID: CLFprKMuCt9

UltraBlame original commit: 534f2a0caa2c98d1eabb00e9a74352831bc2f2de