Bug 1406440 - don't inline nsSMILNullType::Singleton(); r=dholbert
Defining Singleton() in the declaration of nsSMILNullType implicitly
sticks an "inline" on the function, which is not what we want: inlining
it spreads around a lot of static initialization code.  Providing an
out-of-line definition is much better in terms of code size.

UltraBlame original commit: 7fd54747a3b754fd1dfad0be8fd561d78fa23b67