Bug 1402921 - Enhance browser_ext_tabs_readerMode to verify that correct isInReaderMode is reported for onUpdated, r=Gijs,mixedpuppy
All of the information needed for an extension to watch for tabs moving in and out
of reader mode is already available in the onUpdated event. This adds some asserts
to the existing test to check that this information is accurate.

MozReview-Commit-ID: 7OkR933MUPl

UltraBlame original commit: fb216e2b987078d7f79d2319072430292be3bc7e