Bug 1407864: Do not thrash the selector cache when parsing the same selector with different backend. r=heycam
Since the presence of an entry with a null selector is different for Gecko and
Servo, this seemed easier, and mimics nsLayoutStyleSheetCache.

Also, this is going away soon anyway as soon as I get to implement the rest of
the methods for stylo.

MozReview-Commit-ID: DtHJbw8C0GX

UltraBlame original commit: 06ca0845ca1eafeeca3e5495d1af613d4885b377