Bug 1406196 - Move GRADLE_USER_HOME in tree. r=maliu
The goal of this approach is to tell Gradle to not connect (or allow
it to connect) to the network when fetching dependencies.  No Android
automation tasks should fetch from the network, except the toolchain
tasks (which are specially intended to do so).

It's difficult to arrange this without including the `--offline` flag
everywhere.  It _should_ be possible to set offline using an
environment variable -- which would allow us to get rid of these
dotgradle-* files -- but offline isn't an option in
https://docs.gradle.org/4.2.1/userguide/build_environment.html#sec:gradle_configuration_properties
(and certainly not in earlier versions either).  Therefore,
environment variable that points to an init.gradle file in automation.

Before this patch, the files telling Gradle whether to start offline
were fetched from tooltool.  That's just a layer that doesn't need to
be there.

None of this impacts local developers.

MozReview-Commit-ID: LAXktbBu1Az

UltraBlame original commit: a49d59767cbd59dff4c4fad068c9c96348bff752