Bug 1361080 - update webextension tests to directly use devtools loader;r=rpl
In the end the webextensions tests that test devtools should be able to load devtools
apis in the same way as devtools tests. This allows to remove the gDevTools getter on
the devtools shim, which makes the paths that can potentially initialize devtools easier
to predict.

MozReview-Commit-ID: AHokH12ygKI

UltraBlame original commit: e53ec162daed8a43a54827a01945863355b42e2e