Bug 1405070 - Bypass React to clear console on destroy; r=bgrins
Previously we were using the regular clearOutput function, which
was doing too much when we only want to destroy the console. It was:
- dispatching an action so React did the update
- focusing the JSTerm.

By simply clearing the outputNode things are much snappier. Also, we tried
to not clear the console at all, but while destroying there was too much
repaint events happening that it would take the same time as doing it the
old way.

MozReview-Commit-ID: FrFqToG3nRg

UltraBlame original commit: 1cb834c1980659f46273fff15b0ca48923952e70