Bug 1408099 - Fix ExtensionPreferencesManager.getLevelOfControl to deal with undefined settings, r=aswan
Prior to this, if getLevelOfControl was called before the setting was defined to the EPM, an
exception was thrown. This fixes it by returning a level of control of "not_controllable"
if the setting has not yet been defined.

MozReview-Commit-ID: 96jXLoJlUI

UltraBlame original commit: 247bb54b615fd5a73353f7f8b186672fcbaf8fee