Bug 1404344 - Part 2: DOM Cache API schema 57-56 downgrade support. r=bkelly
Alter 57 and 58 DOM Cache API to use an on-disk schema version of 25 again,
as was used prior to the landing of bug 1290481 that bumped the disk schema
version to 26.  Patched versions will recognize this internally as schema 27
based on the presence of the column introduced by the 26 upgrade.

UltraBlame original commit: efe85265a82ad64e2e82128f245a6281222a9a28