Bug 1409040 - Add element.isDOMWindow. r=maja_zf
In addition to tell DOM- and XUL elements apart we will in the
future need to identify WindowProxies because WebDriver expects
there to be a separate "web window" serialisation for these.

Telling if an abritrary object is a WindowProxy is tricky but we
check if its node.toString is callable, that it returns "[object
Window"], and finally that its "self" attribute is itself.  This is
not perfect, but good enough for now.

MozReview-Commit-ID: 6FLOa3qYLPP

UltraBlame original commit: e156787ee23bfb2d2e1950920a45dcbfa1e93fd2