Bug 1409958 - reset stylo blocklist to empty string on Nightly for testing. r=heycam
The layout.css.stylo-blocklist.blocked_domains pref is set to "arewestyloyet.rs"
on Nightly, but empty string on Beta. Since we're going to QA the styloblocklist
add-on in Bug 1407911, we make the prefs aligned in both Beta and Nightly in this
patch.

MozReview-Commit-ID: HoleJvVYm7o

UltraBlame original commit: 229916481648ead2e21ec31aa14c8b54a766180d