Bug 1402450 - Ensure the test properly delays until the content process receives the touch events. r=botond
It appears that in this case just waiting for the APZ flush to complete was
insufficient, probably because the APZ flush and touch inputs go via different
IPC channels and therefore do not provide ordering guarantees. So instead of
waiting for the APZ flush, let's just wait until the touchstart is received
in the content process and use that to resume the test.

MozReview-Commit-ID: AcPRhox1Xkg

UltraBlame original commit: bf168e36be08b031d517eb10c649e5d327a41f00