Bug 1403653 - Part 1 - Refactor getDominantColor. r=nechen
We want to use the Palette library for getting a fallback accent colour for lightweight themes, however because of bug 1318667, we might have to continue using our own implementation of getDominantColor on x86 devices.

Therefore we move this into BitmapUtils, so we can have a central location from which to switch between our own and the Palette library implementation.

MozReview-Commit-ID: 52WsfZbW12x

UltraBlame original commit: 6869117e9c70f84604aaac444d0675f7934770c7