Bug 1410891 - Fix prompt handle assertions. r=maja_zf
The first assertion was never run because it was passed a tuple
of a response value and a boolean.  It now compares the response
value to the boolean using "is".

The second assertion mistakenly used the "==" (equals) operator
but should use "is" to get a better pytest error message.

MozReview-Commit-ID: HCCr5rExIYu

UltraBlame original commit: 2252020d28137221965982736cde24802a74d71a