Bug 1380014. Share fonts with WebRender. r=lsalzman
This changes the serialization format a little bit.
We now have an index at the end of the blob. This
is currently used to store a list of the used font keys.
In the future we'll add rects and can use it for invalidation.

UltraBlame original commit: 2a824a8a2277bc8f725b959be4b1962d8593b02f