Bug 1412431 - Remove support for MOZ_PREFLIGHT and MOZ_POSTFLIGHT; r=nalexander
These variables are unused in the repo. The %_ALL variation is used for
sccache, however.

The last use of these variables was removed when support for MacOS
universal builds was removed.

Someone somewhere may be relying on this functionality. But I don't
think it is worth continuing to support in client.mk. If someone
really needs it, we can reintroduce support for the feature elsewhere,
like in mach or via mozconfig support somehow.

MozReview-Commit-ID: 4KznCmK3JTN

UltraBlame original commit: 5e5ee518a4b9d43e72c879e9d124aa432fb06e1a