Bug 1413483 - add mPendingSeekOffset to store the seek offset initiated by the cache. r=gerald
Seek() can't create a new channel if the resource is still suspended. we need
to store the seek offset specified by the cache so we know where to seek when
the resource is resumed.

Otherwise the new channel will download data at the wrong position and the
cache will need to run Update() and call CacheClientSeek() again which is
a waste.

MozReview-Commit-ID: 9OJ3FRnKitT

UltraBlame original commit: 346ff0f73eaa8a492e2e808a6ee24b1f7017ccb2