Bug 1410796 - Add nodeType existence check. r=whimboo
If the node variable is an object that is not null, and nodeType is
not an own property of it, node.nodeType === node.ELEMENT_NODE might be
evaluated as undefined === undefined if ELEMENT_NODE is also missing.
This will pass the condition, which is an unintended consequence.

To remedy this, we first check that nodeType indeed exists on node
before proceding.

MozReview-Commit-ID: G0eaRvwm5mv

UltraBlame original commit: cefff31e8739bf321fc366470c7583f2973b9e43