Bug 1406505 - Fix element.isInView to use table cells instead of rows. r=ato
As long as bug 1413493 has not been fixed the <tr> nodes will not be part
of the element tree. To workaround this problem we could use the first
found cell of the given row to check for visibility.

MozReview-Commit-ID: 57nuvxYrSMg

UltraBlame original commit: a7d85e5e585fe19b082a3452f655e261084098a7