Bug 1406217 - Fix computation of scrollframe's offset to the reference frame. r=mstange
We need to get the offset across documents boundaries, since the scrollframe
might be inside a subdocument. In such cases we were previously computing
a (incorrect) zero offset for the scrollframe, which was throwing off the
margin for the sticky items inside the scrollframe.

MozReview-Commit-ID: AWeQ9ay2cmp

UltraBlame original commit: e12eb0aa5165ee6b053a18338ae8f86e10749292