Bug 1405396 - Post: Run geckoview tests on all but non-Gradle Android builds. r=gbrown
As we transition to Gradle (and only Gradle) Fennec builds, we're
turning the Gradle builds (Bg) into non-Gradle builds (Bng).  It's not
worth the effort to rename "gradle" to "non-gradle", so this is a
confusing looking patch that turns on GeckoView tests for everything
but Bng (which has kept the "gradle" name).

MozReview-Commit-ID: C1PlsehRwJf

UltraBlame original commit: ded315eea526ef0a9cff4d65b28000edac98551a