Bug 1375772 - don't evict the block which is in the current cached range. r=cpearce
Evicting the block will result in a gap in the current cached range starting
from mStreamOffset to mChannelOffset. Then we have
|GetCachedDataEndInternal(mStreamOffset) < mChannelOffset| and MediaCache
will open a new channel to fill the gap which is bad.

This is exactly the issue described in bug 1347174 where we limit the readahead
size to prevent the problem above from happening.

However bug 1347174 is indeed a workaround instead of a fix. It works around
the issue by suspending the download before the cache is full and has to evict
some blocks to allow new data to come in.

We should let MediaCache suspend the channel if it is full or evict played
blocks if possible.

MozReview-Commit-ID: HuUsZLdHGuh

UltraBlame original commit: 3fcb7efd900132296ab90c8f46d18add2532e6f6