Bug 1412737. P2 - Read() should return only when enough bytes are read or EOS/error is encountered. r=bechen,gerald
This will remove the need to retry reading for the callers.

Note since data is usually downloaded faster than being consumed, we don't
benefit much in reading data from a partial block in the memory. Chances are
we still need to wait for the block to commit to the cache so the reader can
continue. So we change the code to always read data from the cache or from
the last block when it is completed (reaching EOS).

This change allows up to somewhat optimize NotifyDataReceived() which won't
have to wake up readers if no blocks are committed to the cache.

MozReview-Commit-ID: KwgNSOawuAE

UltraBlame original commit: 39f2cccbf9f73a3c0ea7dc6e2fe72a0216418b03