Bug 1416052 - Remove comment filtering; r=nalexander
The auto-generated make file that we include (and the `mach environment`
output that we included before that) should not contain comment lines.
I think it is safe to remove the code that filters them out.

It is possible a multi-line value in mozconfigs could contain lines
looking like comments and this may cause problems. I'm inclined to
believe that this scenario doesn't exist. If someone complains and
we need to bring back support, we could certainly do that.

MozReview-Commit-ID: 8kKw91HH4ms

UltraBlame original commit: 1cd169ff8a0dbb2ea7d50b4275472ca9298765b3