Bug 1419221 - Generate target pseudo element before calling getUnanimatedComputedStyle for the pseudo element. r=birtles
Before this patch, 'content: ""' was applied to the ::before element after
calling getUnanimatedComputedStyle() with '::before' argument so the function
was called for inexistent pseudo element.
Gecko generates ::before and ::after element even if the content property for
the element is none (bug 1387931), so Gecko is not affected by this patch at
all.  Whereas Stylo has returned the parent style as the pseudo style for the
inexistent pseudo, it will be fixed by bug 1418867 and a test case will be
added in that bug since SimpleTest does not have todo for doesThrow().

MozReview-Commit-ID: 7uJcCrtu9ke

UltraBlame original commit: b096c6518c9f8715f22178041eeb18e3a86a00d3