Bug 1406253 - Part 2: Implement nsIImageLoadingContent.currentRequestFinalURI. r=bz
ImageLoadingContent.currentURI returns the "URI" of currentRequest, which is
the URI used to start that request.  Some consumers need to know the final URI
of that request instead.

If the image request gets redirected on loading (e.g. an add-on intercepts the
request), currentRequestFinalURI will be the redirected URI, while currentURI
would be the original URI before redirect.

MozReview-Commit-ID: 9lX063uAIp1

UltraBlame original commit: 9f13acd8e9e0ee76094a9fba76adfa80670cf431