Bug 1419182 - Pass topsrcdir into compare-mozconfigs.py; r=mshal
We currently have the logic for compare-mozconfigs spread across
2 Python files. I'd like to move the logic into 1 file so we
can do nicer things more easily. In preparation for this, change
compare-mozconfigs.py to receive an argument that is the path to
the topsrcdir. Also, switch to argparse because it is more modern.

MozReview-Commit-ID: 7oFQzAdmcim

UltraBlame original commit: 489860a7e03a7113521434f2cd6699664232dcfe