Bug 1416319 - 7. Don't expose viewport metrics from LayerSession; r=rbarker
Provide a set of coordinates APIs in LayerSession instead of exposing
the raw viewport metrics, which are hard to use. This also lets us
remove ImmutableViewportMetrics completely.

The new APIs provide rectangular bounds in client or surface
coordinates, and matrices to transform those coordinates to screen
coordinates. This is done because the transformation to screen
coordinates could (in the future) involve skew, rotation, etc, so it's
up to the application to decide how to handle non-rectangular screen
bounds.

MozReview-Commit-ID: 8Yw8L63TmrQ

UltraBlame original commit: 22648ca7dae8e7def463423f38390d09a1ee357a