Bug 1420092 - Don't reject --enable-jemalloc on platforms where it's not the default. r=njn
The first step of making --enable-jemalloc the default everywhere is to
at least allow to build with it everywhere. Which currently probably
fails on a few platforms, but they're not going to be fixed if they're
explicitly rejected at configure time.

UltraBlame original commit: abab9d80f685f778550e98625c284a0f01526700