Bug 1419426 - Implement browserSettings.contextMenuShowEvent, r=kmag
This new API exposes the ui.context_menus.after_mouseup preference to extensions
so mouse gesture add-ons can modify it.

Note that this is not supported on Android, and also calling it with a value of
"mousedown" on Windows is a no-op.

MozReview-Commit-ID: AkhRmAyzuSp

UltraBlame original commit: d4779e0fdbbba3291ca8c858a051fc979c5b7a3f