Bug 1421476 - Make waitForAnimationFrames wait for the correct number of consecutive animation frames for the given count. r=birtles
As mentioned in the previous commit, with the conformant Promise handling there
are cases where a requestAnimationFrame callback can happen in the same tick of
the refresh driver, that is, the same animation frame.  We should only count
callbacks that occur in frames subsequent to the current one.

MozReview-Commit-ID: IEC7uFwGysS

UltraBlame original commit: f3ab006af5341eda6adcb51733b8656ce430abb4