Bug 1415913 - tabs.create without a windowId should target only normal windows, r=aswan
This introduces a new property to WindowTracker (on desktop only) called topNormalWindow
which returns the topmost window which is not a popup window. That property is then used
by tabs.create to identify the top "normal" window when a windowId is not passed to it.

Bug 1421709 has been filed to look for other opportunities to use topNormalWindow
in place of topWindow in our codebase.

MozReview-Commit-ID: DcV93wO6FoV

UltraBlame original commit: 100e07f0bd493323d3b2e63686f0342a3f9a840b