Bug 1419034 - rabaldr, factor out the stack frame logic. r=bbouvier
This patch does several interconnected things with Rabaldr's stack
frame.

First, it abstracts the stack frame logic from BaseCompiler into
BaseStackFrame; BaseCompiler now only knows very little about the
frame, and what it is allowed to know is documented (mostly that there
is a static part and a dynamic part and that masm.framePushed() is
always correct).  In particular, BaseCompiler must use BaseStackFrame
to push and pop values from its value stack, it can't use masm
directly.  This abstraction is important for ARM64, where we can't
freely push and pop, because SP must be aligned on a 16-byte boundary.
BaseStackFrame is also in charge of allocating locals to frame slots
and to access the locals.  Crucially, these abstractions do not
require us to use a frame pointer.

Second, it removes all uses of the StackPointer variable (not defined
on ARM64) and most uses of masm.getStackPointer() from BaseCompiler;
the few that are left are well-defined cross-platform and are not
sensitive to the layout of the frame, only its size.

Third, it makes BaseLocalIter the property of BaseStackFrame (but
defined outside the latter since the iter class is accessed
externally).

MozReview-Commit-ID: 5BWCTeK1NOR

UltraBlame original commit: b6938001319243e025fc8f3606a1c287360f0b3d