Bug 1377007 - Implementation of the Token Reader dedicated to testing;r=arai,jorendorff
This patch ports to SpiderMonkey the tokenizer currently implemented
in the external binjs-ref tool.

While this tokenizer will clearly not be the tokenizer eventually
shipped to end-users (whether in Firefox or in binjs-ref), the plan is
to keep it both in binjs-ref and in SpiderMonkey (Nightly only) as a
tool for helping test the higher layers of Binjs.

MozReview-Commit-ID: 1i6XnVIf8p5

UltraBlame original commit: 893e303e17ecb4d31e62f1ff5de741e94531c6d1