bug 1423014 - 5 locales per chunk. r=Callek
We used to be at 5 locales per chunk, but I reduced it to 2 because we were timing out on signing. That appears to have been due to network issues; let's go back to 5 to reduce the number of tasks in the graph.

MozReview-Commit-ID: Km7YD6RrPmv

UltraBlame original commit: f05c730d1de7f7d4c8f57bdc09b04d567870f5d9