Bug 1417232 - Part 0: Extract |mach android archive-geckoview|. r=ted.mielczarek
We've had good success with |mach android *| for invoking test suites
and other Gradle-related features; this was just an oversight.  The
real advantage is that almost all of the magic Gradle targets are now
in mobile/android/mach_commands.py and not scattered throughout the
tree.  That'll make it easier to move the actual Gradle configuration
choices forward in the future.

MozReview-Commit-ID: GI1pUHHfXeO

UltraBlame original commit: 8699e19b02e534610cc8b3d8eb78741669fa2d95