Bug 1419964: Remove AdjustParentFrameForAfterContent. r=mats
Instead, handle the display: contents case and the ::after case using
FindNextSibling.

The removal of line frames is moved up to not interfere with the insertion point
computation.

Also parentAfterFrame is renamed to nextSibling, which is more relevant to what
it was doing with display: contents.

MozReview-Commit-ID: 1cqclsGQlaw

UltraBlame original commit: ada567096efbc34a281bcd10767e78a9bc2e200a