Bug 1420547: Notify the pres shell before everything else on removals. r=bz
This allows the flattened tree to have a reasonable state before content
removals, allowing us to fix bugs like bug 1420533.

This used to be also the old setup, but was changed in bug 382376, since the
frame constructor used to reconstruct stuff synchronously. This used to break
all sorts of stylo invariants, and I fixed this in bug 1389743, so now the order
can be sane again.

MozReview-Commit-ID: K2pTweGKSq0

UltraBlame original commit: 04a9c149d299878293af6551518e0eb42df11f18