Bug 1426558 - Make autospider builds not rely on system libnspr. r=sfink
The only reason autospider builds succeed in running tests at the moment
is that there is a libnspr4 library installed at the system level on
Centos that is binary compatible with what the js shell requires.

While on the long run we should just avoid depending on libnspr4 at all,
in the short term, we should make the effort to make those tests use the
libnspr4 present in dist/bin.

For the tests executed from js/src/Makefile.in, it turns out there is
already a level of wrapping that does that, relying on run-mozilla.sh,
which is only installed on gecko builds. Installing it on standalone js
builds solve half the problem.

The other half is addressed by setting LD_LIBRARY_PATH before invoking
the js shell in the various places it's invoked.

UltraBlame original commit: 8e075bee2a527a028d847334ed4fbda6ec5d07e3