Bug 1359405 - Redesign HTMLEditor::IsVisTextNode() r=m_kato
Despite of its name, HTMLEditor::IsVisTextNode() returns true with its out
argument when the given text node is empty.  And although it returns nsresult,
it's almost always NS_OK because it returns error only when the editor isn't
usual condition.

So, it should return bool and true when the given text node is visible.

This patch separates the method.  One is for checking the node with frames,
called IsInVisibleTextFrames().  The other is for checking it only with
its text, called IsVisibleTextNode().

MozReview-Commit-ID: EdQmkOxfNxO

UltraBlame original commit: 2a84ee8f73414ef95ab094cf7fe928971289a3f1