Bug 1397426 - Add hasLayers to nsITabParent and use it in the async tab switcher. r=billm
This is necessary because sometimes the async tab switcher will instantiate when
there already exists some background tabs that are rendering via print preview. When
that happens, it's important for the state to be set correctly for them so that we
don't accidentally treat them as still loading, and wait (forever) for them to report
having finished loading.

MozReview-Commit-ID: 2dwo5WlXlgJ

UltraBlame original commit: 3fa35817d29d5fe75e59f92cb794dcd6adb4f9b4