Bug 1404853 - Enable browser_webconsole_highlighter_console_helper.js in new console frontend; r=Honza.
This changes the code of the test since it was relying on direct access to the content page.
In order to be able to interact with the element picker, the test-actor and test-actor-registery
files are loaded by head.js.

MozReview-Commit-ID: FyebIiyF31j

UltraBlame original commit: e14fbde5349138be557996ca3f55c687ebcfb7be