Bug 1429672 - ensure DeferredTask is defined r=jaws
Binding the Preferences object to a "this" object was supposed to enable userChangedValue to access the DeferredTask instance that XPCOMUtils.defineLazyModuleGetter defines on the "this" object.

I thought that worked, because tests passed, but it turned out that no tests are exercising the branch within userChangedValue that references DeferredTask, which is only followed for pref form controls that specify "delayprefsave."

Only font controls specify that attribute, and there weren't any tests for changes to those controls.  This patch adds such tests and then changes the way that DeferredTask is lazily retrieved and referenced to ensure it's defined as referenced in userChangedValue.

MozReview-Commit-ID: BPbGp55s9wC

UltraBlame original commit: 72823606d8b47ddbbfa7cb2dd975e0e52f4a60a3