Bug 1429961: Make sure to remove all the continuations with the same PostDestroyData. r=mats
If we remove the first continuation of the ::after content before the last,
before this patch unbind the generated content ASAP. Part of the generated
content (the second counter) in the test-case ends up in a continuation of that
frame.

That text node would already be unbound from the tree when destroying its
frame.

MozReview-Commit-ID: 4VtVRSoGxCT

UltraBlame original commit: 467cbbd4d1602a7fd04722e32eba9740929d4f97