Bug 1431169 - P2. Fix Apple AAC decoder on some files. r=rillian
It is necessary to provide the AAC's magic cookie to the decoder for some files (such as ambisonics or quad-stereo). We already do so for the Windows and FFmpeg decoder.
It's another case of how did it work before really.

MozReview-Commit-ID: sOtY1sPeBU

UltraBlame original commit: e97d7f1498adb10c3fb615189a2e7a9479184441