Bug 1431090 - Add FuzzingInterface support to JS engine. r=froydnj
This patch adds basic support for the fuzzing interface in the JS engine on top
of the last patch. This includes all the necessary code except for actual
targets (just an example target skeleton) and also makes sure that the fuzzing
code is packaged for the standalone release.

MozReview-Commit-ID: D6Tyebz3jZS

UltraBlame original commit: 2e98bda3f39735bbc557c4652a03a071044b03b4