Bug 1420229 - Run formautofill xpcshell tests in headless mode;r=bdahl
This is needed because DOMParser is invoked on some fixture HTML that contains an
<svg> element (Walmart/Payment.html). When this happens **and** there is
a declaration for `color: MenuText` in a UA stylesheet (which is added in the next
changeset), there is a graphics crash. This is fixed in the new headless mode,
so this changeset forces the directory to run in it.

MozReview-Commit-ID: 9T8eFOW4xoz

UltraBlame original commit: 23d4aff26260628ccaaeb649dcdd44499a504cc2