Bug 1432541 - Don't reuse clips across a stacking context with a nonzero position. r=mstange
In bug 1426386, I made it so that in some cases, part or all of the
transform component of a stacking context got moved to the bounds
top-left. However I neglected to update the ScrollingLayersHelper, which
was relying on the transform to know if it could reuse clips across a
stacking context boundary. It now also needs to check the bounds
top-left as a result of the aforementioned change.

MozReview-Commit-ID: 7mDRSYqZ8bF

UltraBlame original commit: a60df143441674d4c948d95d0e69d5b6f3ee61ca