Bug 1432390 - Allow to derive docker images from other in-tree images. r=dustin
So far, the best we've been able to do is to upload an image to the
docker hub, and point an image's Dockerfile's FROM to the version
uploaded onto the hub.

That is a cumbersome process, and makes the use of "layered" docker
images painful.

This change allows to declare a parent docker image in the
taskcluster/ci/docker-image/kind.yml definitions, which will be
automatically loaded before building the image. The Dockerfile can then
reference the image, using the DOCKER_IMAGE_PARENT argument, which will
contain the full image name:tag.

Some details are left off, for now, such as VOLUMEs. At this point,
VOLUMEs should all be defined in leaf docker images.

UltraBlame original commit: 4b04997348874684161a8caec0c3954018eb6302