Bug 1395973 - Remove index cache from nsContentIterator. r=masayuki
nsContentIterator used to maintain a stack of indices so that when it
finished iterating through a subtree it would know the position of the
next node. Maintaining this stack is expensive and unnecessary since we
have fast getters for next and previous siblings.

UltraBlame original commit: 1d057e2c36fe0fddb020243ea2acf7f062be7d87