Bug 1435022 - Inline a one-line function with a confusing name. r=sotaro
This function is exposed as public but only ever called from two places,
both of which are inside the class. It's confusing to have it as a
separate function because it just makes it harder to trace through the
code.

MozReview-Commit-ID: BC7mvOc9wo

UltraBlame original commit: e7f0a181d2eb2055d6a9091e325b311e663dc025