Bug 1435022 - Inline the SetNeedsComposite function. r=sotaro
SetNeedsComposite is only ever called from one place on the compositor
thread, but it has a bunch of generic boilerplate to handle being called
from any thread. If we inline it we don't need the extra boilerplate and
it's much simpler to follow the code.

MozReview-Commit-ID: E1AcMh80KsH

UltraBlame original commit: 89844912178c7602736faae777b0966a759f2e08