Bug 1427007 - Part 1: When a SessionWorker is restarted, we also need to re-initialize it with the correct SessionFile data. r=Yoric
We now know that worker restarts are rather frequent and we've had reports that
are certain to point at us forgetting to properly re-initialize the worker.
This takes care of this by factoring the init flow into its own method and setting
the flag when a failing worker is terminated.

MozReview-Commit-ID: G5jccjxkBsF

UltraBlame original commit: 986dab420f52a20c772a2c0d7009ae8ad2967595