Bug 1333899 - Part 2: Add nsDataHandler::ParsePathWithoutRef. r=valentin
This adds a version of |ParseURI| that only deals with the path portion of the
URI and expects the ref not to be present. It's mainly copy and pasted from
|ParseURI| but updated to use nsString methods rather than NSPR.

Additionaly it only returns a substring to the path passed in for the data
buffer if requested.

|ParseURI| is updated to advance past the scheme and trim the ref, then it
just uses |ParseURIWithoutRef|.

UltraBlame original commit: fc9b9ea20e7c40074a41c42a2242d6c34629c183