Bug 1435296 Address test failures caused by bumping timer precision to 2 ms r=baku
This affects several tests, and in all but one case merely disables timer precision
reduction on those tests.

In the other singular test, it no longer requires the end time of an operation to
be strictly greater than the start time, and allows it to be equal to it.

MozReview-Commit-ID: J59c7xQtZZJ

UltraBlame original commit: 89c121b45b300d99a4a853947f80db9b85feae6e