Bug 1426063 - Add PLATFORM built-in to Fluent in Gecko. r=Pike
In order to allow localizations to produce different string depending on the
platform, we need to add a Gecko-specific built-in function to the MessageContext.

I'm explicitly listing the variables which we pass, rather than just passing the value
in order to ensure we don't start returning values we didn't plan for in case
the AppConstants.platform gets updated.

MozReview-Commit-ID: 1KZ6bf6zIY2

UltraBlame original commit: 2b7d42d527af582a465e49187fe387442d524a7c