Bug 1426482 Report browser errors in Nightly to Mozilla. r=gijs
Errors are collected via nsIConsoleService, shaped to a Sentry-compatible
format, and sent off. Reporting is on by default, and can be disabled using a
checkbox added to the privacy prefs in about:preferences.

Collected errors are sampled to avoid overloading the collection service; the
sample rate was determined by a previous Shield study that measured the number
of errors occurring in Nightly.

The feature is hard-disabled outside of Nightly and local builds, and the
preference is disabled by default in local builds. It is intended as a prototype
that will be evaluated and replaced by a more robust collection system if it
proves helpful.

Differential Revision: https://phabricator.services.mozilla.com/D561

MozReview-Commit-ID: 6aqUatXyuYs

UltraBlame original commit: e491f731de97e88845c62e66aa4aa5df319a4c93