bug 1437120 - Ensure pings sent after net shutdown are persisted to disk r=Dexter
bug 1397293 introduced a mechanism by which we would bail out early in trying
to send a ping if we were trying to send it after the network had been torn
down.

Unfortunately, it did so indistinguishably from the case where we weren't
allowed to send pings, so we neglected to save the ping (as "pending") and
just archived it.

This change cleanly rejects the ping, and correctly tests that the rejected
ping is persisted, not just ephemerally hanging in memory for but a few moments
longer.

MozReview-Commit-ID: 2g8cpeBEzSE

UltraBlame original commit: a0a032ce94f89b2a738037aa4bd71cf9277948e9