Bug 1422160 - Merge cache and site data in about:preferences. r=jaws
This commit removes most of the cache section in about:preferences,
following the UX concept from bug 1421690. This is in the general
interest of de-cluttering privacy preferences and giving users controls
that are easier to understand and use.

The cache size is instead shown in the site data section and the cache
can be cleared using the "Clear Data" button in that same section.

MozReview-Commit-ID: 7PDTDgllFFI

UltraBlame original commit: e8d05c7218c66151e2e30dcff2fabdf9bccae40e