Bug 1430856 - Default to two default capabilities when a camera can handle anything. r=jib
If a camera returns no capabilities we interpret it as it being able to handle
any capability we throw at it. However, we also end up trying to start it with
the default capability of 0x00. This often works, but we can crash when
rescaling it to the chosen target capability 0x00.

With this patch we inject up to two default capabilities, one at 640x48030 and
one at 1280x72030. With constraints present we'll try to adjust these defaults
so they fit within the constraints while at the same time preserve the
aspect ratio given by prefs.

MozReview-Commit-ID: 3mr7Li5TTbV

UltraBlame original commit: 12b8e4ad2fdc00a75fb9852cacf7d4b72f2b8bfb