Bug 1439016: Clear servo data on slot changes too. r=bholley
Pretty much the same way as what we do on XBL insertion point changes, since the
data would be stale.

I think both of those are kind of a hack, btw, and that we could fix it doing
ClearServoDataFromSubtree properly before we start unbinding. That'd prevent all
these issues, and all the complexity that entails receiving restyle requests
mid-unbind (we'd guarantee the tree is always in a stable state).

That's a matter of a different bug though.

MozReview-Commit-ID: Ev6RvGuPGiv

UltraBlame original commit: 577a507f9ee1b970382008c46d5402970eb82632