Bug 1439358 - Part 3 - Always raise ViewShowing events and don't update the "current" property. r=Gijs
The ViewShowing event is now called earlier and unconditionally, since we don't set the "current" attribute and call showMainView while the panel is closing anymore.

It is already the case that the ViewShowing event handlers don't depend on the "current" property, so we don't need to keep track of it before ViewShown events are dispatched.

MozReview-Commit-ID: Ii4SN03KjwW

UltraBlame original commit: 525b36047e168f0a5853ba39984b11327a99549c