Bug 1439525 - Remove spurious "continue" statement. r=nical
The "continue" statement makes it look like the loop can run multiple
allocations in a row when that never actually happens. (And if it did,
the latter half of the loop would do the wrong thing). This patch
removes the spurious continue statement and replaces it with a more
useful assertion that documents the state at that point.

MozReview-Commit-ID: 3rpVZ1imX2y

UltraBlame original commit: cc82773113d1958ec5515b869ba17895f6ebb1ef