Bug 1439333 - Make AutoFlushICache::flush() do something on ARM64. r=jandem
AutoFlushIcache::flush() was ifdef'd out on ARM64 without any
mechanism to signal that, but that's wrong.  Reorganize ifdefs for
this function and also for AutoFlushICache::setInhibit() so that a new
port will not fail to detect that these functions do nothing unless
the new platform opts in to the code that does something.

UltraBlame original commit: d2ffffb30b31aaaaf4d1d5623eb77894b0455806