Bug 1438113 - Make it possible to have wasm baseline without ion. r=luke
For ARM64 it will be the case initially that we'll have baseline but
not ion.  Thus we need to be able to disable asm.js support in some
cases when wasm is supported (since baseline does not support asm.js)
and to make the logic that selects the compiler to use slightly more
subtle.

As a result the old logic of "just use Ion as a fallback" needs to go,
but we can now move that into HasCompilerSupport() and the result is
cleaner overall.

UltraBlame original commit: 6e46ff4ff446e5548326e9d3468485c15934e088