Bug 1255924 - Remove mobile/android/base moz.build definitions. r=jchen
My word, that's a great deal of configuration!  The key technical
change here is subtle: we hard-code the GENERATED_FILES in the
Makefile.in rather than fishing it from the preprocessed files of the
constants.jar.  Again, the final goal here is to use GENERATED_FILES
for the APK itself, which will allow us to capture the dependencies in
the moz.build file directly, and then this hard-coding can be
removed.  One step at a time.

MozReview-Commit-ID: FpgKV1IVu05

UltraBlame original commit: 3d0a9dfcf5dfecdf1e81a2c9e8c0654d1d8af5fa