Bug 1433958 - Make NS_MutateURI methods warn immediately after calling method that fails r=mayhemer
Calling NS_ENSURE_SUCCESS at the begining of the methods shows the error in the console, but what we really care about is where the error code comes from. So it is best to use if (NS_FAILED(mStatus) {} at the begining of the methods, and NS_ENSURE_SUCCESS right after calling the appropriate method on mMutator.

MozReview-Commit-ID: 5vVuHk3N4FU

UltraBlame original commit: 6c437b464d613af08b898df54fa9dcf056591d2d