Bug 1429904 - When a JSContext for a thread is about to go away, collect enough information about any JIT entries in the buffer so that the entire buffer can be streamed to JSON. r=njn
This changeset changes behavior.
If the profile is streamed before any JSContext has gone away, we now iterate
over the entire buffer twice (per thread): First, to collect information about
JIT frames, and then again when we build the JSON for the samples. The first
traversal stores small pieces of JSON for JIT fromes in individual strings, and
the second iteration splices those strings into the thread JSON's frame table.

When the JSContext for a thread goes away, we no longer build JSON for samples,
and we don't reset the profiler buffer. We now only build the JSON for JIT
frames. Once the complete profile is requested and we build samples for it, we
iterate over the entire buffer, and look up the cached JIT frame information for
JitReturnAddr entries from the correct range. Different parts of the buffer may
correspond to the life time of different JSContexts: For each JSContext we will
have one range in the JITFrameInfo, and we can look up the correct range based
on the buffer position of the JitReturnAddr entry that we're processing.

This new way of doing things has multiple advantages:
 - We no longer reset the buffer, so we no longer lose information about other
   threads.
 - All threads from a given process now always have sample data for the same
   time range. Before this change, the "partial profile" from a thread that
   lost its JSContext could extend further into the past than the other threads'
   profiles.
 - Requesting profiles multiple times now has more consistent results. Before
   this change, the first requested profile would include the partial profile,
   but then the partial profile was discarded. And the second requested profile
   would not contain any data for the time before the JSContext went away.
 - We now do less work when a thread's JSContext goes away. This should
   decrease the interruption time.

MozReview-Commit-ID: 3KhnPtBijna

UltraBlame original commit: 544e3884d895cf1be55c55e40e8a0c2be56f7aa6