Bug 1174003 part 6: [css-flexbox] Replace ComputedCrossSize() helper with a new API that uses logical axes internally. r=mats
This patch doesn't affect behavior.

It removes a helper-function that simply returned nsStylePosition::mWidth or
mHeight -- whichever was in the flex container's cross axis. This helper was
only used to answer the question "is the cross size 'auto'", at a single
callsite.  So, this patch replaces the helper with a new helper that more
directly answers that question.  The new helper's implementation uses logical
axes for its reasoning, too, whereas the removed one used physical axes (and in
particular, it relied on AxisOrientationTracker::IsCrossAxisHorizontal(), which
I'll be getting rid of later in this patch series).

MozReview-Commit-ID: EJ8MObTauZH

UltraBlame original commit: 88d72b7833f55d731132e60ae6a4641452ceadcd