Bug 1437335 - Manual proxy port configuration should disable when its option is not selected, r=jaws
This is a regression from bug 1429593. The code that was added to enable/disable controls on the page
depending on whether an extension is in control was being run when the page opened, which overrode
the code which disables specific controls based on the current proxy setting. All that is needed is
to call that code (which is in gConnectionsDialog.proxyTypeChanged()) if an extension is not in control
after the new code that was added. This patch fixes that issue, and includes a test verifying that
the manual controls are not enabled when the screen first opens.

MozReview-Commit-ID: 7jpp1wewZ2k

UltraBlame original commit: e78fb987e0b00b420477cea6156f81a6551a08cb