Bug 1437382 - Part 4 - Make sure that flushing private tabs data reaches GeckoApp in time. r=jchen
There are two parts to this:
1. While the file writing itself by the session store can be done either
   synchronously or asynchronously, the session store's _writeFile function as
   a whole always behaves asynchronously. In addition, just writing the file
   (which will be done synchronously when flushing tabs) takes time which we
   don't have, as we should send the private tabs data as fast as possible to
   Java in order to avoid hanging the UI or missing a possible timeout.
2. Sending the data to Java needs to happen synchronously as well, so we need
   to listen for "PrivateBrowsing:Data" on the Gecko thread now. This in turn
   means that some sychronisation is now required between the UI thread handling
   onSaveInstanceState and the Gecko thread that is actually receiving the data.

To avoid hanging the UI and causing ANRs, we only wait a limited amount of time
for Gecko to respond with fresh private tabs data, though.
As this still leaves a certain possibility of outdated private browsing data
being saved and possibly restored after an OOM-kill, we're also going to speed
up the processing of TabClosed events by the session store in the following
parts.

MozReview-Commit-ID: EkNFre5RhQW

UltraBlame original commit: 22531e0df45684032848918d23cb812eb71395df