Bug 1441766 - Fix search keywords for site data and cookies in about:preferences. r=nhnt11
When moving cookies to the site data section we forgot about some
searchkeywords which were left on the cookies exceptions dialog.

We also forgot to remove the status column label when removing the
status column.

Further, the search tooltip is not displayed correctly on the three
vertically aligned dialog buttons (it's always on the last button),
because we did not put each of them in its own hbox (which is just way
too easy to get wrong and needs a better mid term solution).

MozReview-Commit-ID: L6ZaocWsjYU

UltraBlame original commit: 161a8e168596d92afa8c7f0c3bc46e185e2ee540