Bug 1441872 - Wait for next tick in BrowserTestUtils.tabRemoved. r=arai
Ideally we shouldn't listen "SessionStore:update" to know the tab was removed
since closed tab information is updated in the listener of "SessionStore:update"
in browser.  So it might be possible to that the listener in BrowserTestUtils
is invoked before the listener in browser.

MozReview-Commit-ID: A3S28Bmyvtw

UltraBlame original commit: 2a6481c5aabbaa7a4c9cf657a20c16c4084a6b2d