Bug 1442187: Wait until next tick after popupshown before resolving show(). r=Gijs
With the migration of Promise microtasks to the main microtask queue, promise
resolution handlers are now called immediately after any JS event listener
returns. The result of this is that in the case of the show() method, the
resolution handlers now run before the view's own popupshown handlers run and
the view is set active, which breaks handlers which expect it to act more
sensibly.

By deferring the resolution until the next tick, the resolution handlers are
called when the view is visible and in a consistent state.

MozReview-Commit-ID: C7697CBoHFt

UltraBlame original commit: 2a775fb2d85e9794364569b1bff761459b270c61