Bug 1338039 - Add extremely permissive fuzzy annotation for box-sizing-replaced-001.xht to prevent failures due to paint delay reduction. r=dholbert
The test that we are updating the fuzzy annotation for is the W3C copy. There is an *existing* local copy of this test that
was created as part of Bug 1295466 (copy of test is here->https://searchfox.org/mozilla-central/source/layout/reftests/bugs/1295466-1.xhtml)
and this copy is the stricter modified version which makes it ok for us to have this extremely permissive annotation on the W3C copy.

MozReview-Commit-ID: 2qXSbVONl64

UltraBlame original commit: b9d430803080aef235ed2d064eebf1d79f7ff738