Bug 1438394 - Do SysV IPC check before bailing out early. r=jld
The SandboxLaunchPrepare currently bails out early if it detects a
lack of user namespaces. Hoist the check for drivers needing SysV
IPC up so it's done before that early exit, and the required env
variables get correctly set.

With this we no longer fail with a SIGSYS sandbox error, though
in a debug build we still crash because many assumptions in the
graphics stack get broken when that fails to initialize the driver
for WebGL.

MozReview-Commit-ID: 8n3Hx6VSjTF

UltraBlame original commit: 9e92145bb5726a3a23255140fe995edc724cac63