Bug 1437295 - Remove unnecessary change from bug 1374166. r=botond
The change that was made in bug 1374166 was attempting to fix the
problem fixed by the previous patch, but didn't actually succeed (it
just made it less likely to occur). Now that we have the proper fix we
can revert that botched attempt to speed up the test a little bit.

MozReview-Commit-ID: 3hWZ6bFTdxb

UltraBlame original commit: 5a973861da6a9d3ecb4e7ba10b744958f98c40b6