Bug 1445386 - Use GreD for artifact build scalars. r=chutten
On macOS, the paths for the special directories `GreD` (Resources) vs. `GreBinD`
(MacOS) are different.  On other platforms, they are the same.

With scalar loading for artifact builds, we need to load a file from
`Resources`, so `GreD` should be used for this case.

MozReview-Commit-ID: 91JFwOISQCk

UltraBlame original commit: 3df355881aa04907482a7a96a84430c4ce732bd0