Bug 1445624 - [mozlog] Various improvements and fixes to the mach formatter, r=gbrown
This consolidates the printing of status logs, which was previously handled
differently in 3-4 places. This also fixes a few of the annoyances listed in
the bug 1445624 description. Finally this also fixes a few edge cases that I
noticed when writing the tests.

MozReview-Commit-ID: APudT8yBqVS

UltraBlame original commit: 61a14c8bcfcf095c16caf4564fdc7d879386e2ca