Bug 1447506: Make Devtools' TableWidget tests not rely on XBL bindings getting applied synchronously when XUL is mixed with HTML. r=bgrins
So TableWidget's columns does something where they mix HTML and XUL:

  https://searchfox.org/mozilla-central/rev/b29daa46443b30612415c35be0a3c9c13b9dc5f6/devtools/client/shared/widgets/TableWidget.js#1055

The label's .value is an XBL binding property:

  https://searchfox.org/mozilla-central/rev/b29daa46443b30612415c35be0a3c9c13b9dc5f6/toolkit/content/widgets/text.xml#19

This tests expects the XBL binding for the <label> to be applied synchronously,
which the other patch in this bug prevents.

Is there a reason that is a XUL label btw? We don't seem to use the binding
methods at all (except in this test, lol).

Anyway, just use getAttribute("label") in the test instead.

MozReview-Commit-ID: 3DOI0Lff6AW

UltraBlame original commit: c094743582b2cef801700cdf0085bf176a5619fd