Bug 1278581 - Part 4 - Ensure the progress bar doesn't get stuck after displaying the error page. r=esawin
For a normal load, we get a request start, the various location change/load
events and then a stop. When encountering an error (e.g. server not found), the
usual event order is start, stop and then the location change/load events while
the error page is loading (compare bug 976426).

When displaying the "unknown protocol" error page from within the Content-
DispatchChooser however, we never receive any onStateChange events from the
WebProgressListener. When this happens within an existing tab, this isn't a
problem - the UI never thinks that the tab is starting a page load and thanks
to bug 976426 never displays the progress bar.

When opening a link in a new tab, though, the UI speculatively creates the tab
as STATE_LOADING and shows the progress bar. Because we never receive any state
change events if the "unknown protocol" error was triggered, especially not a
"Stop" event, the progress bar then never gets hidden because the UI assumes
that the tab is still loading.

To work around this, we make use of the error type that we were still trans-
mitting during DOMContentLoaded - if an error page was detected, we reset the
tab's load state in the UI and stop the progress bar.

We also take this opportunity to change the handling of Content:LoadError
messages. Originally, this message was intended to reset the URL bar if an
attempt to load an URL resulted in an immediate error in browser.js. For un-
known reasons, the rewrite from loading throbber to progress bar changed the
meaning of the event to set the progress value of the progress bar to 80 %, just
like DOMContentLoaded.

With this patch, a Content:LoadError message will immediately stop and hide the
progress bar as well.

MozReview-Commit-ID: 8AXICB7mEfK

UltraBlame original commit: cb9e04f4c32d8d6c77e7e6f714d9c88d9a246b6d