Bug 1448200 - Handle the creation of the spellchecker more gracefully. r=Felipe
If the user right-clicks on an input field and that click causes us to create
a spellchecker, then it won't be initialized in time to show the context menu.
Handle that case more gracefully by passing that state up to the parent.

MozReview-Commit-ID: 1hjxloq2OLp

UltraBlame original commit: 8fbafbc46da259a054dd00feb93abc20e3140c56