Bug 1450307 - Round scaled paint size instead of paint bounds in GenerateFallbackData. r=jrmuizel
A component of dtSize could be zero because RoundedToInt(rect).Size() is not the same as
RoundedToInt(rect.Size()). The size after rounding might be zero, even if the rounded size
is not zero. For example with x0=0.5, x1=1.1.

MozReview-Commit-ID: 3NeBMAVD5ub

UltraBlame original commit: aa8bcc37d8884c02d34d547308c355f4a7f72624