Bug 1437004 - Introducing BinSource parser generator;r=froydnj,jorendorff
This crate contains a parser generator as a Rust crate. The parser generator is used to generate
BinSource-auto.h, BinSource-auto.cpp, BinToken.h. As of this changeset, to limit yak shaving,
the parser generator is not part of the build system. Making it part of the build system
is delegated to bug 1439645.

MozReview-Commit-ID: 1lODDSIsz8W

UltraBlame original commit: 646bdf14bf8e51b07f75643839ee807972b7feba