bug 1440673 - Allow changing the max number of keys per-keyed-scalar r=Dexter
I considered a few different ways to do this, but storing the limit inside the
KeyedScalar itself was the cleanest. The limit check is too deep to pass it as
an argument, and making it conditional on the keyed scalar being called
"telemetry.event_counts" was too fragile.

MozReview-Commit-ID: AyfEKB40Abb

UltraBlame original commit: 66fa06d26606b6e75a8e5905cd81a7cc444df33e