Bug 1451748: Always remove unloaded views from views list. r=mixedpuppy
Contexts for active extension views are kept in a Set on the owning extension.
That list is meant to be kept current, with views added and removed as they're
created and unloaded. A refactoring at some point in the past, though, changed
that so that we only cleaned up parent views at extension shutdown, not at
view shutdown.

MozReview-Commit-ID: FW8KHPOD9qc

UltraBlame original commit: 19e0f2b68a3c77d2591cef2bc2ecd0a8feabf201