Bug 1443943 - Ensure redirect SJS' serve the correct content types. r=jya
dynamic_redirect.sjs and midflight-redirect.sjs are serving files
with "Content-Type: video/ogg", which is incorrect and could lead
to problems given that we're not always asking it to serve Ogg
files. So include the type be to served as a query parameter.

MozReview-Commit-ID: 5f0PXy8lL3G

UltraBlame original commit: 53d478be27969bb9a00c67b085d6dddbfd0b79b6