Bug 1452592 - Check mutability when importing globals. r=bbouvier
The mutability of the imported value must match the mutability of the
import declaration.

If an imported value is provided as a number constant and not a
WebAssembly.Global, then it is implicitly immutable.

Drive-by fix: We test for non-number imports before testing for i64.

UltraBlame original commit: 8a735eb98349815f732f712f4a67bcdbd0024b4b