Bug 1448674 - avoid closing extension ports while detaching tabs r=aswan
Extension ports are automatically closed when the message manager of
the source is destroyed. When a tab is detached from a window, its
frameloader is moved to the new window and the original message
manager is destroyed.

Bug 1445537 started listening for SwapDocShells events, but that only
works for the first swap (e.g. detaching a tab once). To avoid early
disconnection of the port, we should continue to subscribe to
SwapDocShells events.

MozReview-Commit-ID: G2ZYAhNyHIL

UltraBlame original commit: b35a42daaf1e9407d4835813cf3de5d1d8f11b66