Bug 1427726 - move test for bug 1063538 away from jar URIs, r=baku
Bug 1063538 was reported with a testcase that accesses a .jsp, which
in turn sleeps "forever". The testcase seems to be simulating this by
using a jar:http: URI for a really big file contained in some other
directory.

Instead, we can just use a .sjs file that does a similar thing to the
original jsp file in the reporter's testcase, which conveniently also
allows us to remove dependencies on support files in other directories.

MozReview-Commit-ID: 2JCOS9VLgVv

UltraBlame original commit: 753ece6c9f1bf415676b97ebc89208909814f2a5