Bug 1452487 - Remove directInject functionality in evaluate.sandbox. r=maja_zf
The evaluate.sandbox function accepts a directInject argument,
which is a relic from Marionette's B2G past when it did not support
evaluating scripts with lasting side-effects.

The API documentation in GeckoDriver#execute_ mentions directInject as a
valid parameter, but it is not picked up or passed on to evaluate.sandbox.
This effectively means the directInject functionality is unused.

MozReview-Commit-ID: 3rYjRQ2R5GV

UltraBlame original commit: feda10ec46db61602a3f39d0cd0f701d3bbccb35