Bug 1452483 - Set correct line number for JS evaluation. r=maja_zf
Marionette incorrectly sets the JavaScript context line number to 0.
The line number is provided to us in in the input, so we should
use this.  The default fallback if line is not provided is 0 as before.

MozReview-Commit-ID: 8gOt9r4awee

UltraBlame original commit: e73d9a52e2e47a4e2db55fd3a73bbbf3948ccb0f