Bug 1448553 - Part 1: Defines common functions for getting a Unicode URL or a Unicode URL component r=nchevobbe
This patch defines some common functions for getting a Unicode URL or a Unicode
URL component.

These functions are not only intended to be used by future patches for this bug,
but also intended to be the common functions for any future patches which need
to reuse such functionalities.

MozReview-Commit-ID: DZBYoKv9Fmx

UltraBlame original commit: 348535562dc3d2acaa659c2983b49c9931ec303e