Bug 1445272: Add gcTypesEnabled to wasm::ModuleEnvironment and Metadata; r=luke
This is going to be used in subsequent patches, to know if AnyRef is accepted
as a ValType in misc places.

Note temporaryGcTypesEnabled should be removed once we've implemented proper GC
support instead of the hack planned in bug 1445277.

UltraBlame original commit: 40baddc5f0b6206f9a90bd694c596b0d8f8a9837