Bug 1452111 - Simplify dispatch-to-content regions as well. r=mattwoodrow
We already do this for the maybe-hit regions because on some pages we can get
oodles and oodles of regions and unioning them all takes a long time.
Simplifying the regions speeds this up massively. It should be functionally
correct since the dispatch-to-content region is allowed to overestimate the
actual dispatch-to-content region.

MozReview-Commit-ID: 6Wl5nuVXB7w

UltraBlame original commit: 0c7db264c68e1f3074462cc17c488047a16f85c4