Bug 1453846 - Remove DevTools RDP events. r=Honza
A few years ago in bug 1126274, we added RDP events to the DevTools transport so
that the add-on RDP Inspector could display packets as they flow by.

The add-on no longer works as it is not a WebExtension.  Maybe someday we'll
revisit this, but for now this removes some dead code.

MozReview-Commit-ID: AvgQhYWwBUA

UltraBlame original commit: 4012129c0b4023a46e345b8f99e76b31dcd3514f