Bug 1451469 - Set the sample time on APZSampler. r=botond,nical
When sampling APZ transforms from rust code, we will need a timestamp at
which to sample the transforms. It's not obvious what the right
timestamp is to use here, and this will almost certainly be revisited
when we are hooking up OMTA in bug 1453360. For now we just stash the
most recent composite timestamp on the APZSampler and use that when
sampling. This seems to work fine.

MozReview-Commit-ID: KinsXO9tEJH

UltraBlame original commit: 61724e1bb2ad7d0642d4d4cd7c05a396f44b7591