Bug 1450199 - Cookie is not synced across tabs r=Ehsan
This patch reverts parts of changeset e87e706def11 (bug 1425031).
The problem in bug 1425031 was that when the content process set a cookie
a notification was sent to the parent process. This notification was then
forwarded to all the content processes, including the one it originated from.
The solution was to not forward cookies that originated from a content
process, but this causes the current bug.
The correct fix is to forward the cookie changes to all content processes
except the one they originated from.
The test for bug 1425031 remains, and should keep passing.


MozReview-Commit-ID: 1P6JwHQDy93

UltraBlame original commit: 5f3b781af6a705f77ea41b1f0dde160eb3968b38