Bug 1355143 - Size scrollbars with 'hover' Gtk+ state, r=jhorak
Ubuntu/Ambiance has tiny scrollbars when it's not hovered by mouse and large
ones when it's hovered/active. Our current Gtk+ toolkit code does not support such scrollbar
resize on the fly.

We use a workaround as we get size of active (hovered) scrollbar only and that
we pass to gecko. Then we draw scrolbar smaller when it's not hovered and full sized
when it's hovered.

MozReview-Commit-ID: mxt9q5Bcg9

UltraBlame original commit: 4405d1c4673b0b1eb415b528c5e252585094b150