Bug 1455226 - Load memory/initializer.js via browserRequire instead of a script tag;r=ochameau
"async" methods defined in scripts loaded via a <script> tag seem
to indirectly use the document's Promise, which is not usable after
the DOM of the document has been destroyed.

This is an issue if we invoke this code during the destroy of the
toolbox.

MozReview-Commit-ID: C8juQqJlVDN

UltraBlame original commit: ff1831623026f2f3b7e208c63810fee500b7e565