Bug 1456262: Optimize addAddonsToCrashReporter at startup. r=aswan
This consistently shows up in profiles as taking about 6-10ms during startup.

Part of the problem is that it loads TelemetrySession.jsm, which we can't fix.

Part of the problem is that the escaping and string manipulation it does is fairly expensive. We can cache the escaped values in addonStartup.jsm to avoid that.

A lot of the problem is that we call it every time we load a bootstrap scope, even though we already set the correct value before we start loading any bootstrapped scopes, and it doesn't change after that. We can avoid this by skipping the redundant calls during startup.

MozReview-Commit-ID: 68EWdnJdPvk

UltraBlame original commit: 31dd07dd000efefc62ae97317ac336aec085ef79