Bug 1456115 - Remove locking in AcmReceiver::GetAudio. r=dminor
This also causes a lot of dropouts. We don't need to lock here. NetEQ is thread
safe, and its created in the ctor. The rest of the members are made atomic or is
simply never accessed in multiple threads.

MozReview-Commit-ID: 2fRw5ZgxdpQ

UltraBlame original commit: 290879198eb4a7d2b7ecfd5ba07b38ea31710529