Bug 1448439 - Extract a FlingPhysics abstraction from GenericFlingAnimation. r=kats
The FlingPhysics controls the shape of the fling curve, including how fast
and far it goes, while leaving other logic (related to flywheel, handoff,
overscroll, etc.) centralized in GenericFlingAnimation.

FlingPhysics is a template parameter of GenericFlingAnimation because we
already have a dynamic dispatch on the type of the animation, we don't
need another one for the physics.

I called the specific class implementing the current physics
DesktopFlingPhysics because since the deprecation of B2G and Android
switching to StackScrollerFlingAnimation, it's only used in desktop touch
scenarios.

MozReview-Commit-ID: LhJ9rpPrnXC

UltraBlame original commit: da26dcf12417c29b3d1f11e251e088ffb83632c9