Bug 1454978 - Make OMTP feature detection not depend on tiling. r=bas
There's a circular dependency between `UsesTiling` and `InitOMTPConfig` because
we try to disable OMTP if we will be using tiling and edge padding is enabled.
Now that edge padding is disabled everywhere except android it should be safe
to assume that if edge padding is enabled then we'll be using tiling as well
and should disable OMTP.

This commit also removes a check on `UsesTiling` from CalculateWorkerCount
as it is redundant.

MozReview-Commit-ID: 1ruWPwXfLwO

UltraBlame original commit: 2ba500ed529be48c75e68b896ae1739f801fed80