Bug 1424505: Don't try to retrieve the accessible focus from remote documents on Windows 7. r=eeejay
Bug 1421144 fixed IAccessible::accFocus to work when focus is within a remote document.
Unfortunately, this causes mysterious intermittent crashes when called from a UIA client in Windows 7.
Ideally, we'd deal with the actual cause of the crashes, but they seem to be deep in Windows RPC code and all attempts at tracking this down have failed.
Clients don't seem to need this too often anyway (and it's a minor annoyance if it doesn't work when they do).

MozReview-Commit-ID: IxvbBGJ2wxA

UltraBlame original commit: 68bfd841265f54fdd4a214214e9ccc8297c64672