Bug 1458798 - Clear the previous timestamp when we skip calling SampleAnimations in CompositeToTarget(). r=kats
The purpose of the previous timestamp is to reduce the time difference between
the timestamps on the main-thread and the compositor.  Once we skipped a frame
to compose animations on the compositor, the timestamp is far behind from the
timestamp on the main-thread, so we should clear it.

MozReview-Commit-ID: EGiQfwejfsy

UltraBlame original commit: da163bfbc6c8a460d6b3c551456174153912f743