Bug 1459245 - Migrate <stringbundle> from a XBL binding to a Custom Element;r=mossop
This loads a JS file to register the Custom Element whenever XUL documents
get created. This is a pattern we plan to follow for future elements as well -
stringbundle is just an easy target to prove it out with.

MozReview-Commit-ID: 7PlisCknrKW

UltraBlame original commit: 80fc39b7c5ec8442c4f6666e291a241b60e965f0