Bug 1455568 - [wdspec] Use shared command function for each WebDriver command. r=ato
Tests for a specific command have to use a shared function
for each of the included tests of a module to avoid redefinition
of the same code.

In a follow-up patch those shared methods could even be
separated out as fixtures for each of the sub folders.

MozReview-Commit-ID: 11lAp7J8IoZ

UltraBlame original commit: a73b7f0983b04398dfa83ca1bc5ca9f200445f5b