Bug 1459737 - Make run-task more runnable on non-POSIX platforms; r=dustin
I want to make run-task work on Windows. The script is currently very
POSIX oriented, as it assumes the existence of the grp and pwd modules,
that user IDs are numeric, and that system calls like setresgid() and
setresuid() are available, etc.

This commit starts to make some of the POSIX-centric code conditional
on running on POSIX.

Code for uid/gid extraction has been moved to its own function. Some
error messages were tweaked slightly as part of the move. Otherwise,
the changes should be pretty straightforward.

There are still other parts of this file that won't work on e.g.
Windows. But this gets us a big step closer.

MozReview-Commit-ID: HNyytKcBbBo

UltraBlame original commit: acfbf6d8eb020a4bde835cc68380927dd540462b