Bug 1459697 Part 4: Change a WPT reftest to make failures visible in red. r=dbaron
MozReview-Commit-ID: A4UxNn4Njpi

This test applied a green color to the test area, when it should be
transparent, to allow the red to show through in case of test failure. Without
this change, this test won't catch cases where the float area is too small.

UltraBlame original commit: 731dfa211b3846b7c5aff134ecc097f83865a913