Bug 1453501 - Immediately create Compositor on Android r=jchen,rbarker
Currently we don't create the Compositor until we have a valid surface
to render into. This causes a race that can result in us not being able
to display anything at all once a surface is provided and the compositor
is started. It seems the easiest thing to do right now is to avoid the
race by starting the Compositor immediately.

MozReview-Commit-ID: HkdVL3LBNZB

UltraBlame original commit: c075885867014f8e48374746c18f09a5d3aacf21