Bug 1457972 - Part 9: Allow passing invalid values to xpidl arrays for compat reasons, r=mccr8
This one sucks. I found out when running tests that we exploit some edge
case behaviour in our tests where passing an invalid value to an xpidl
array parameter will be passed as an empty array...

I figured that just respecting this behaviour for now is the easiest
approach, but we will probably want to fix it in the future.

UltraBlame original commit: 18baccb9f4c61498a69b73d396902045bcd1a682