Bug 1389099 - Properly encode multi-byte translations in CSS r=kmag
The localization filter was not unicode-aware because convertToStream
assigns the output to a nsIStringInputStream (which takes 8-bit chars).
The input was read as a 8-bit string, but after localization it can
contain wide strings if a translation has a multi-byte character.

To fix this, the input stream is now first read as a UTF-8 string, then
localized, and finally exported via a nsIArrayBufferInputStream..

MozReview-Commit-ID: LjCxczIFKCR

UltraBlame original commit: c4caca02549844815d894a60acf340cade739cc1