Bug 1461001 - Implement frontend for console.timeLog; r=bgrins.
The patch handles timeLog packets we receive from the backend.
We add some stubs in order to have access to packets in mocha,
and some mocha tests to make sure we render those messages as expected,
be it when the timer exists or not.
The BASE_PATH constant now makes use of MOZ_DEVELOPER_REPO_DIR environment
variable in order to have an absolute path; without it the stubs couldn't
be written to local files.

MozReview-Commit-ID: Fx4DyRgOzcd

UltraBlame original commit: d441f5a05fbc076595d3502c6544271ec4c9497e