Bug 1462030 [wpt PR 11033] - [LayoutNG] Improve support for negative clearance., a=testonly
Automatic update from web-platform-tests[LayoutNG] Improve support for negative clearance.

When there are adjoining floats that need to be separated from a cleared
block, clearance is predetermined, and we have to force it on the block.
Any large top margin on the child will just be eaten by negative
clearance. There's nothing that can be done to prevent clearance in such
cases. So make sure that we don't try to determine whether to apply
clearance or not based on the clearance offset set in the constraint
space. When clearance has been predetermined (which may have triggered a
BFC offset resolution and what not), refusing to apply it to the child is
a bug.

This aligns the implementation with the "alternative" way of calculating
clearance in the spec [1]. Everyone but Edge does this. I think what
Edge does here is problematic, because it requires us to use a
hypothetical position that was calculated before clearance got applied
(clearance causes margins that would otherwise collapse to be
separated). We'd end up using a hypothetical position not based on the
actual layout situation.[2]

[1] https://www.w3.org/TR/CSS22/visuren.html#flow-control
[2] https://github.com/w3c/csswg-drafts/issues/2608

This fixes one existing test. Added a few new ones as well. Not all of
them failed prior to this CL, but they serve as regression tests for
things I found to lack coverage while working on this.

Cq-Include-Trybots: master.tryserver.chromium.linux:linux_layout_tests_layout_ng
Change-Id: Ia45e9048c75c338477eee4923ff16eea3245bc6a
Reviewed-on: https://chromium-review.googlesource.com/1061470
Commit-Queue: Morten Stenshorne <mstenshochromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrickchromium.org>
Cr-Commit-Position: refs/heads/master{#560208}

--

wpt-commits: 499bf362082f5ca8679ad981536cbb38a98fccb3
wpt-pr: 11033

UltraBlame original commit: d76e1688749096dbe55cd2867c696d7df364a9cc