Bug 1265584 - Move wptrunner marionette usage onto a single thread, r=ato
Running marionette on a background thread is problematic in the case
that a test times out. In this case the background thread is not
terminated. If we then call into marionette again on the main thread
we may race with something that happens on the runner thread. The
marionette client isn't threadsafe, so this leads to buggy behaviour.

The simplest fx for the problem is just to move all the marionette
calls onto the main thread and instead of waiting on the main thread,
spin up a thread with a timer.

MozReview-Commit-ID: 3vVlMcwPHSx

UltraBlame original commit: 9b8539a9a2a6c20220cdcb07e19b7bf3a91c7436