Bug 1265584 - Output asserts to wptreport.json, r=maja_zf
In the future we want to update metadata using the wptreport.json data, since
that's much much smaller and easier to parse. In addition, this is
required to determine if a test fully passed, so it makes sense to
store it here.

MozReview-Commit-ID: ErYT33BhqGg

UltraBlame original commit: b7716dedbf5a24e0e9a940a907225f7b3dd7d35c