Bug 1146316 - Leak fix, set up CC in XPCShell too. r=smaug.
XPCShell allows interaction with object that are cycle collected, but we never run the
cycle collector in it. In XPCShell we don't call nsJSContext::EnsureStatics.
nsJSContext::EnsureStatics is responsible for setting DOMGCSliceCallback as a GC callback,
which we need for running the cycle collector.

UltraBlame original commit: 5c390ce65c956c29c028f02cb122529233d13680