Bug 1461276 - Update telemetry probes for media. r=jya datareview=chutten
This makes drno and jya the recipient for the media alert emails, removes
cpearce, ajones, and also removes a few people that are not involved with the
project anymore (and the some emails were no longer valid anyways).

Also, bump a number of probes to 63, to get an updated picture of the data.

MozReview-Commit-ID: EmlHyihnxIw

UltraBlame original commit: 359724bd9ca78c1bed7cd885cec7cb6a8fb5b04b