Bug 1387678 - Update tests to not use shadow DOM like a frame; r=ato
The current element finding and interaction code expects the shadow DOM
to act like a frame when it comes to keeping references to elements found
or not found on the document. This expectation is not correct for the way
shadow DOM works in reality.

MozReview-Commit-ID: JFWvlbEylj4

UltraBlame original commit: 3c9c76ca14ad5303b36ac5e2974c15b835da27fb