Bug 1466031 - Tidy up test_animation-finished.html; r=hiro
This mostly just involves using promise_rejects to simplify some test code.

The change to the test description of the last test might seem like it is giving
it the opposite meaning but the original test name was "Test finished promise
changes when animationPlayState set to running"[1], and it is about testing
that there are NO changes but that got lost when the "Test" part of the
description was dropped.


[1] https://searchfox.org/mozilla-central/rev/cb846d13d3f9ec5d38ace93a74f749a18e9c67f5/dom/animation/test/css-animations/test_animation-player-finished.html#287

UltraBlame original commit: 865062de80fb7f7beb5c74fef93b0412147d090c