Bug 1465983 part 1 - Fix race in test_font_loading_api.html. rs=heycam
This patch changes 'TEST 3' so that instead of testing that all the
documents' document.fonts.ready Promises are resolved at a certain time, that
instead it just waits for them to resolve and then checks that they resolved
to the correct object.

The test previously assumed that calling SpecialPowers.pushPrefEnv on the
page's 'load' and waiting for pushPrefEnv's callback would be enough time
for document.fonts.ready to have been resolved for the top-level document and
all of its frames.  That is not necessarily the case.  Even the Promise in the
top-level document itself is not guaranteed to have resolved by that point.
The Promises will not have been resolved at least until style and layout has
been flushed, and in fact in Mozilla's implementation it frequently won't
happen until the first refresh driver tick after layout has finished.  The
result of this is that the CI machine 'Linux x64 QuantumRender opt' was
failing 'TEST 3' intermittently.

UltraBlame original commit: 94409f81ab0e75db962f24b29cb8a7f1bc3a71f2