Bug 1465935 - Handle another edge case with hit-testing inside fixed-pos items. r=mstange
Transforms are containing blocks for fixed-pos items, so if a fixed-pos
item is inside a scrolled transform, then it should use that scrollframe
as the scroll target for hit-testing. This patch handles this case for
WebRender by stashing the appropriate ASR on the nsDisplayFixedPosition
item and using it instead of the presShell's root scrollframe in this
scenario.

The patch also adds a mochitest (which is basically a mochitested
version of the reftest in fixed-pos-scrolled-clip-3.html, with a
hit-test check to ensure that it's hitting the right scrollframe).

MozReview-Commit-ID: 7YQAeOiMMuP

UltraBlame original commit: ada5a84764728f3d16d60f65052cf56f84aabd51