Bug 1465697 - Add menu open probes for Savant Shield study; r=jaws
These probes will register and record (for the duration of the study only):
* When the library menu opens.
* When the hamburger menu opens.
* When the dotdotdot menu opens.

These will be captured in existing and new windows.

Note: The library panel is dynamically added and removed from the document (see PanelUI.js for where the panel is created and removed), so a listener can't be added to it in advance. However, the library menu "ViewShowing" event bubbles up to the navBar in its default location. A separate listener is needed if it is moved to the overflow panel via Hamburger > Customize.

MozReview-Commit-ID: EBBBgXAQnxE

UltraBlame original commit: 58e58e8e3601985c88be1a279ef7813a6ba5995e