Bug 1467076 - Reflect the content of the style attribute in the inline element rule; r=gl
This makes the content of the inline element rule much simpler.
It now is always in line with what the style DOM attribute contains.
It does not attempt to remember old properties that used to exist but
were removed.
It also reflects the order in which properties appear in the attribute.

This is what Chrome DevTools also does.
It is also more correct because we used to try to remember old properties
and marked them as disabled, but trying to enable them would never work.

One test was therefore removed because it wasn't needed anymore after this
change.
One new test was added. It checks the order of properties specifically.


MozReview-Commit-ID: uzNezMdxBX

UltraBlame original commit: 6809bb66d5b8abea2c02cd5752c6a2043238367d