bug 1465562 - ensure succeededCertChain is set in TLS handshakes with session resumption r=fkiefer
When doing TLS session resumption, Firefox currently does not have enough
information to trivially reconstitute the original connection's security
information. Consequently, we have to rebuild the certificate chain in the
handshake callback. Before this patch, we determined the EV and CT status of the
connection but did not set the succeeded cert chain unless the certificate was
EV. This was insufficient. In this patch, we set the succeeded cert chain
regardless of if the certificate is EV or not (provided we found a valid chain).

MozReview-Commit-ID: AuKrlBwX1Qh

UltraBlame original commit: 5d47226b6b8c1d57afbb92ebc50bae2632611e0e