Bug 1466231 Part 5: Update a wpt reftest to remove its reliance on fitting inline-blocks into a too-narrow space. r=dbaron
This reftest currently uses a container that is 1 pixel narrower than needed
to fit two blocks side-by-side. That's useful for making the first two blocks
land on separate lines, but it makes the test rely on the browser forcing the
first inline block on a line to fit even if there is no space for it (in this
case, the block that should be pushed to the right).

MozReview-Commit-ID: IxtOR2yFNn6

UltraBlame original commit: 074014588d2f7d7afb821f68c3bd903f40d3152b