Bug 1468545 - Add a test for dragging the scrollbar on an inactive scrollframe. r=botond
I had originally planned to land this as part of bug 1348321, with it
exercising the case from bug 1326290. However it exposed the issue that
I fixed in this bug, so it seems reasonable to land this test as part of
this bug.

MozReview-Commit-ID: EL1N9EmMXaC

UltraBlame original commit: 79067eeb997ddddc8ab612ffab6c39a16b7cf424