Bug 1469712 - Don't update mtime for backend file during --dry-run; r=gps
When we do a './mach build-backend --dry-run', no files in the objdir
should be written or touched. Updating the mtime of the backend file
during dry-run causes a future './mach build' to think that the backend
is up-to-date when it is not, which causes an incorrect build.

MozReview-Commit-ID: 2NA0rcSGrvL

UltraBlame original commit: 4f001c506ff8ba7ec562a0576107496c9eeaddeb