Bug 1463911 - Store the WindowInfo objects on the heap. r=sotaro
I'm going to make this structure a little bigger in the next patch, so
to avoid the overhead of copying it around we can just store it on the
heap. I also switched to using a std::unordered_map since now the value
is just a pointer.

MozReview-Commit-ID: 2g1Eyu95W4Q

UltraBlame original commit: b74674d04a6f4e5d5fee169e667b157fbe4e078e