Bug 1471132 - Change how static xpcom components are initialized. r=froydnj
The sStaticModules list is, practically speaking, a copy of the list
of components we already have in libxul, augmented at runtime with
a few other components for tests (for gtest and xpcshell). We don't
actually need to keep that copy in memory. We can instead just use the
pointers in libxul directly to register them to the component manager,
and use a separate list, only for those extra components when they need
to be registered.

UltraBlame original commit: 32d574a02a2eb31ccee9259a210d67edaab0fb03