Bug 1472751 - Remove MAC_NON_BROWSER_WINDOW preprocessor directive;r=Gijs
The menuitem used to be only included in the markup for the non-browser window
case. This changes it to include it as a hidden <menuitem> for the browser window
case, and then unhide it for the non-browser window case.

MozReview-Commit-ID: 8tY3GiTFmqe

UltraBlame original commit: d7750a13f9b5dae0aa5779da04a3c62a9941abd8