Bug 1467148 - exclude .DS_Store file from list of subdirs to check for license info, r=ted
vendor_rust.py checks all dirs in third_party/rust/ for license information, but it wasn't distinguishing between subdirs and files when listing the contents of that directory, and the script earlier causes macOS to create a .DS_Store file in the directory, which the script then treats as a subdir, causing an IOError "not a directory". This change excludes the .DS_Store file (and other non-dirs) from the list of subdirs of third_party/rust/ to check for license info.

MozReview-Commit-ID: 4CrJH9tZLnG

UltraBlame original commit: 5a560404cf9e4c2e92ddc4892c6f36cc8cec7eed