Bug 1473305 - Move browser_jsterm_popup.js to devtools/client/shared/test ;r=Honza.
This test is directly calling Autocomplete popup methods and only uses the
console for the instance and the input it provides. This patch moves the
test to the shared component folder, where the autocompletion code lives
so it feels more natural.
A few test cases are added in the jsterm_accessibility test, this time in
direct relation with the console input.

MozReview-Commit-ID: Ih7rOJ3vHpb

UltraBlame original commit: 5bc3f36982ec7585fc785cdf5fc7bf93eeca32cd