Bug 1121705 - Look at window handles decrement when closing tab. r=whimboo
When Puppeteer opens a new tab using various strategies it relies
on the list of window handles increasing.  When performing the
reverse operation of closing a tab, it looks at the length of <tab>
chrome elements in the UI.

This changes the close operation to use the same mechanism as opening
a new tab to determine if the tab has been closed.  This seems to
be as reliable as looking at the number of <tab> elements.

As part of a forthcoming window tracking refactoring of Marionette
(https://bugzilla.mozilla.org/show_bug.cgi?id=marionette-window-tracking),
the list of window handles will be made even more reliable: a content
browser will not appear in the window handle list until both the tab
and it's linked content browser has been created and properly initialised.

MozReview-Commit-ID: FY5vGBpn64R

UltraBlame original commit: c5adc3b888b932065842854e09027cc981b8662d