Bug 1473525 - Add an API endpoint to wait for HCT to init. r=janerik,chutten
The HCT content library needs the chrome to broadcast the value of the
Telemetry state before it can report a reliable value. This patch adds
a way to wait for the broadcast to happen and singnal that it's ok
to read the value from canUpload.

Differential Revision: https://phabricator.services.mozilla.com/D2030

UltraBlame original commit: c42c7ac4fbac411c40dcc647338b845b4a3fdb07