Bug 1468951 [wpt PR 11526] - Update external/wpt/webrtc/RTCRtpParameters* tests., a=testonly
Automatic update from web-platform-testsUpdate external/wpt/webrtc/RTCRtpParameters* tests.

This is in preparation for get/setParameters[1] and
RTCRtpTransceiver[2].

Tests are updated to use promise_test and async/await to make them
easier to write and to execute sequentially.

Tests are updated to set up a call before getParameters() because many
parameters are defined as "that has been negotiated". If we want to test
what is returned before negotiation we should add that as separate
tests. For now, this is necessary for the tests to run on [1][2].

Some tests are removed because they no longer reflect the spec.

Because get/setParameters() and RTCRtpTransceiver has not landed yet,
the tests are still expected to fail.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/1102436
[2] https://chromium-review.googlesource.com/c/chromium/src/+/1025771

Bug: 803494, 777617
Change-Id: I3f14a37925837576a602d91bb67e48181c56a946
Reviewed-on: https://chromium-review.googlesource.com/1102502
Reviewed-by: Harald Alvestrand <htachromium.org>
Commit-Queue: Henrik Boström <hboschromium.org>
Cr-Commit-Position: refs/heads/master{#571430}

--

wpt-commits: 36fb033a4051e63a0ab4bd073fdaadf808953564
wpt-pr: 11526

UltraBlame original commit: 79e6840b7382fde7d034229ad88f981ec7a6bf32