Bug 1470447 - JPEG decoder should post an invalidation for each row. r=tnikkel
The JPEG decoder will currently only post an invalidation when it has
processed all of the rows it is able to. If it is has all the data, that
means it must fully decode before invalidating. This causes very large
JPEGs to appear in large chunks which feels janky compared to slowly
appearing row by row with the refresh tick. With WebRender, it also
allows us to upload less data per frame update which can be another
source of jank.

UltraBlame original commit: 853cf14dc4628cd431ed8ec3540a4afe6ca3f476