Bug 1468475 - Part 2. Display the zero graduation. r=daisuke
This patch will ensure that displaying the zero graduation into the timeline by
using shift the graduation with zeroBaseTime. If we shifted all of the
graduations, the graph might not have the first graduation. So this patch will
add the first graduation intentionally in this case.

MozReview-Commit-ID: HnLouXBja6L

UltraBlame original commit: b319a3547802fddfeac5347a27caae6f459c15e1