Bug 1445683 - Do not build aomstats unless examples are enabled; r=chmanchester
The aomstats library is only used in the code examples, but we assume that all
libraries should be linked into libxul, which leads to an unresolved external
dependency on fatal at link time. This adds a guard to only build aomstats if
we are building the examples.

MozReview-Commit-ID: 8CRK3klUPk7

UltraBlame original commit: 83f43f8c169f1e61f5080f99122891c1bff1cbcb