Bug 1445683 - Disable av1 on win32; r=chmanchester
We get intermittent OOMs building aom with MSVC on win32. The PGO builds are
definitely a problem, but it may affect other builds as well. The plan for now
is to stop supporting AV1 on win32 until we switch to clang, which hopefully
is not too far away.

UltraBlame original commit: b4ac908007b9af47907b57891fc537201ef489a4