Bug 1459750 - Remove unused principal variable r=rpl
Variable `principal` is only used when `this.isExtensionPage` is false.
So move this inside the branch where this condition is satisfied, and
remove the `else if (this.isExtensionPage) { ... }` block.

(This is a follow-up to bug 1214658, where the check was introduced that
 obsoleted the principal variable when `this.isExtensionPage` due to the
 use of `contentWindow` instead of `principal`).
(`principal = contentPrincipal` was already unused at its introduction
 in bug 1317697)

MozReview-Commit-ID: F7VX9vAH8MM

UltraBlame original commit: 11a2e09c259b41c8183b655bebd8b1979c7b7a31