Bug 1450250: Make <animate> target the light DOM parent, not the flattened tree parent. r=heycam
This matches Chrome:

  https://cs.chromium.org/chromium/src/third_party/blink/renderer/core/svg/animation/svg_smil_element.cc?l=260&rcl=325eafc631f396944801f17675a7092b216d60ca

And we rely on that to animate <use> with SMIL, like:

<use href="#foo">
  <animate ...>
</use>

There, the flattened tree parent of the <animate> element is null, because <use>
is a shadow root without slots.

Similarly we need to skip the GetCtx check which seems unneeded, and also is not
correct since <animate> can animate when it's out of the flattened tree.

Differential Revision: https://phabricator.services.mozilla.com/D2194

MozReview-Commit-ID: 13hRCOj500n

UltraBlame original commit: 91b097951606e1825f3cedc23303697cb1482e14