Bug 1472238: hook up alwaysConnect pref changes r=lina
The original code had a bug in that going from disabled to enabled
would always connect, whereas actually we want to only connect if
there are records (or alwaysConnect is on).

We maintain the existing behavior that if the user has set
dom.push.connection.enabled to false, we don't connect, figuring that
this is the way a privacy-conscious user might indicate that they
don't want to talk to "the mothership".

MozReview-Commit-ID: ClbhYhnHVog

UltraBlame original commit: bcb897d35e71db66f78a5f18fbda1963cff48684