Bug 1478917 - Apply upstream patch fixing ld64 crash with LTO. r=dmajor,froydnj
While fiddling with clang (upgrading it and applying some miscompilation
patches), my mac LTO builds started to fail because ld64 would crash
during configure.

It turns out, it was crashing trying to print a warning it shouldn't
even print out, about failure to create a cache path.

This, in turn, is due to a pointer not being initialized in the ld64
code. I sent this upstream, and this was promptly fixed:
https://github.com/tpoechtrager/cctools-port/pull/57

However, since our last update of cctools-port, upstream landed a change
that broke support for tbd files if you don't compile against the new
libtapi library. Doing so is more work than I'm ready to put here,
so we just cherry-pick the fix.

UltraBlame original commit: 91307a0650ac547c13aed381477178132255ad5a