Bug 1478616 - Generalize the wasm write barrier.  r=bbouvier
We need to generalize the barrier to handle not just globals, but also
fields in structures.  To do this we pass the location of the store
(ie the Cell**) to the C++ barrier machinery, not the global index.

UltraBlame original commit: e713a94f6e551b1e7bff74841faacbfdaae3057e