Bug 1479416 - Add Connect sidebar item to aboutdebugging-new;r=daisuke
This patch adds:
- a new action actions/ui:selectPage
  - takes a page id in argument
  - returns type PAGE_SELECTED, with selected page id as 'page'
- a new reducer reducers/ui-state
  - on PAGE_SELECTED, updates the state by setting 'page' in 'selectedPage'

The state of the application after this patch looks as follows:
{
  ui: {
    selectedPage: String
  }
}

As we add more devices the way we identify the selected page might change
(maybe use an index instead of an id) but that is good enough for now.

The "dispatch" helper is forwarded to the App component via connect.
From that point, components are responsible for forwarding the dispatch
property to all components who need to dispatch actions.

Differential Revision: https://phabricator.services.mozilla.com/D2549

UltraBlame original commit: 42a9a104150d73d6f00f9e64ca081bcffc576618