Bug 1349416 - Only delay removeEventListener when hiding after mouseup;r=birtles
The inplace-editor is programmatically hiding the autocomplete-popup
on TAB and notifies the caller. The caller (eg rule view) will trigger
a click on the next focusable element immediately after that.

If the HTMLTooltip is still consuming outside clicks at that point
the click will be swallowed and the new inplace-editor cannot be created.

Depends on D2661

Differential Revision: https://phabricator.services.mozilla.com/D2662

UltraBlame original commit: 61021fadb55e64afc860e88a36bad794ee335aab