Bug 1481691 - Unpack files only if the devtools-adb-extension versionis mismatch or there are no unpacked files. r=jdescottes
getManifestFromUnpacked() and isUnpacked() are mostly the same as they are in
binary-manager.js [1] in the adbhelper addon.  A big difference is that the
manifest.json copied from the extension is the manifest.json for the
devtools-adb-extension itself, not manifest.json(s) in a subdirectory for
each platforms (e.g. linux/manifest.json).

Due to the verion check in each xpcshell test needs to use a different version
for the test extension, otherwise subsequent tests will use unpacked files in
the first test case, thus the ADB mock won't be used at all.

[1] https://github.com/mozilla/adbhelper/blob/0821424b6b895645560089a134054b35b6649ff8/binary-manager.js

UltraBlame original commit: 32ad5fef292ac00aaf5049c4ceceb5d93f173809