Bug 1482042 - [wdspec] Make "test_negative_x_y" and "test_height_width" independent from system settings. r=jgraham
Both tests are failing due to differnt system settings:

"test_negative_x_y" makes the assumption that the menu
bar on MacOS has always a height of 21px. This is not
true for all systems, because it can vary due to system
settings (font, locale). To make this test work correctly
on all systems the minimum y position can be retrieved
with "window.screen.availTop`. While this CSSOM interface
is not standardized, all of the browsers we care about on
MacOS have it implemented.

"test_height_width" fails in combination with other test,
and that only if the window get moved by the browser to
fit into the screen when a resize would enlarge it outside
of the screen. To fix it a known to work start position
has to be set.

MozReview-Commit-ID: Eu2EN0XT39x

UltraBlame original commit: 83793017b08643c4ef917fcb477e46a936114e08