Bug 1467344 - Update tests in test_animation-ready.html; r=hiro
The test: 'A new ready promise is created each time play() is called the
animation property', is covered in:

  web-animations/interfaces/Animation/ready.html

by the 'A new ready promise is created when play()/pause() is called' test.
As a result, this patch removes that test.

UltraBlame original commit: a4a1c2383121e53bb913333eb04ac6193fa9133c