Bug 1467344 - Drop use of flushComputedStyle in css-transitions tests; r=hiro
flushComputedStyle is as follows:

  var cs = getComputedStyle(elem);
  cs.marginLeft;

That _probably_ flushes style in most engines, but it's possible some engine in
the future will optimize things in a way that it can flush marginLeft without
flushing, say, backgroundColor. Before moving these tests to wpt, it would be
better to explicitly flush the property (or at least one of the properties) we
are transitioning.

This patch also updates the "Setting zero combined duration" test from
test_animation-cancel.html since, in making this change, I realized the test was
wrong.

Specifically, it was testing that when we set a combined duration of zero that
a transition is canceled. However, nothing in the spec requires this. The spec
only requires a non-zero combined duration to _start_ a transition and only
cancels an existing transition if the combined duration becomes _zero_ if the
"the end value of the running transition is not equal to the value of the
property in the after-change style".

This test passed, however, because it changed the transition property to
margin-top, hence it was actually testing the same condition as the previous
test.

UltraBlame original commit: 5b8c402641ac0a302770b6661b5f01c77226c70e