Bug 1481951 part 2: Adjust reftest contain-layout-overflow-002.html to be clearer & have accurate expectations. r=dbaron
This test is basically a copy of its -001 variant, with some "float:left"
sprinkled around on contained descendants.

Before this patch, this test had an additional arbitrary sizing difference as
compared to the -001 version -- there's one element that arbitrarily has class
"outer" in the -002 test whereas it has class "inner-lg" in the -001 version.
These classes have different sizing characteristics, which makes a difference
to whether scrollbars show up, because this element is not contained (though it
is a layout container itself).

This patch undoes this arbitrary difference and also adds a "float" class to
make it easier to see which elements we've sprinkled float styling onto.

Differential Revision: https://phabricator.services.mozilla.com/D3826

UltraBlame original commit: 6c3db80981da2613b3e58c4904dec3504b7476d8