Bug 1467632 - Make AsmJSGlobal's pod field be POD. r=luke, r=waldo
The pod member needs to be POD but has members that have evolved no
longer to be POD - a ValType and a LitVal.  We work around the problem
locally by using ValType's representation type PackedTypeCode to
represent types, and by specializing LitVal as LitValPOD for use in
this structure.

UltraBlame original commit: a0883c99c897122286c6e1711cada83d9889e093