Bug 1470490 - Introduce Atomics.notify, make Atomics.wake an alias.  r=anba
TC39 has decided to rename Atomics.wake as the more conventional
Atomics.notify.  The webcompat fallout from this should be roughly
zero, as browser currently have disabled SAB as a result of the
Spectre kerfuffle.

This patch introduces Atomics.notify, and changes a number of names in
the code and in test cases as a result, but retains Atomics.wake as an
alias until we're happy that we don't need it.

A note on nomenclature used in the code: Though Atomics.notify
/notifies/ the target agent, we still say that the agent was /woken/
by the notification if it becomes schedulable.  An agent can be
notified without being woken in obscure implementation-specific
settings, normally having to do with waiting in interrupt handlers.

Also fixes a completely blatant and impossible-not-to-discover bug in
tests/shell/futex.js.  Do we never run shell tests marked "slow"?
This makes me nervous.

UltraBlame original commit: d9d4bf8a92b36bb8ebadbaecf951e4bfa268d877