Bug 1482025: Don't set aria-label on hbox page action nodes. r=Gijs
This change means that the #pocket-button-box hbox is pruned from the accessibility tree as it should be.

Normally, hbox elements aren't exposed to accessibility, as they are semantically meaningless.
However, this code was previously setting aria-label on hbox elements, which forces the hbox to be exposed to accessibility.

Differential Revision: https://phabricator.services.mozilla.com/D2987

UltraBlame original commit: d2c2abac6d01e7b78142bffecdf5bd3e3265a47c