Bug 1484373: Part 2a - Add Tuple ForEach helper function. r=froydnj
It's currently fairly difficult to perform some operation on each member of a
tuple. This is a particular issue in the context of adding a cycle collection
helper, where we need to perform traverse and unlink operations on each member
of a tuple, and don't have a way to do so without several layers of templates.

This patch adds a ForEach function which will call a function on each element
of the tuple. This would typically be used with a lambda function with a
single `auto&` argument.

UltraBlame original commit: 70b72c8bddc6dd189bf6cbc79828b790e032af3e