Bug 1473514: Display an empty space for the tab icon while waiting for the real favicon to appear. r=dao
This adds a simple empty box that is displayed when we're still loading an icon
but are no longer showing the throbber. Ideally I'd like to keep showing the
throbber and maintain the busy state but that seems more risky for now.

Differential Revision: https://phabricator.services.mozilla.com/D2364

UltraBlame original commit: 141d116489a7de2cd69ca7bd5eb494d1a1dfafbc