Pocket: Tag auto complete appears behind send to mobile section (Bug 1487157). r=thecount
In the Pocket dropdown, the Tag input field has an auto-complete. In the case
that the user is enrolled in a A/B experiment, a "Send to Phone" section will
be added to the panel. The tag auto-complete field, when active, appears
behind the experiment section.

Differential Revision: https://phabricator.services.mozilla.com/D4580

UltraBlame original commit: c7754dbab0b089cc24cd61a3270c598b03d0f6e9