Bug 1479391 - Fix crash after forced return from a generator. r=jimb
This patch switches from GeneratorObject::finalSuspend to setClosed in order to
dodge an assertion in finalSuspend that the Generator state machine is
transitioning along an expected edge. The way the Debugger manipulates
Generator state is decidedly unexpected, from the perspective of the normal
rules, and we've decided to accept that.

UltraBlame original commit: 55517d66b222c53a03d75e84d79d1a32f760274a