Bug 1488337 - Rename mozilla::dom::SelectionChangeListener to mozilla::SelectionChangeEventDispatcher r=smaug
SelectionChangeListener is too generic name but it just dispatches
selectionchange event when it's necessary.  So, it should be renamed to
SelectionChangeEventDispatcher. Additionally, it's in mozilla::dom namespace
but it does not represent any DOM object. So, it should be in mozilla namespace
instead.

Differential Revision: https://phabricator.services.mozilla.com/D4913

UltraBlame original commit: 59b228f7faf14eee64ae6c446a1bb6ee0c520646