Bug 1488692 - Remove old ini files when updating wpt metadata r=jdm
The easiest way to ensure that old metadata files are removed is
simply to check that all the existing metadata files still correspond
to a test file, and remove those that do not. This happens before
metadata update so that if the metadata update causes a patch to be
created, these changes are included in the patch.

Differential Revision: https://phabricator.services.mozilla.com/D5015

UltraBlame original commit: 403376728e2a6f023afe6ff2db8e576b062557fa