Bug 1487279 - Pref to control TLS downgrade check, r=keeler
Background here is that we are disabling a piece of the downgrade protection in TLS 1.3 and we want to turn it on.  We don't know if that is safe, so a pref (and an associated experiment) seems prudent.  This is that pref.

Differential Revision: https://phabricator.services.mozilla.com/D4629

UltraBlame original commit: 8cc7bb4477793283807b0e4ece8424e8974154ce