bug 1490333: marionette: move whenIdle to sync module; r=whimboo
This patch moves the private whenIdle function to sync so it can
be shared across JSMs.

It also changes its semantics somewhat, so that instead of taking
a callback function (suitable for DOM event callbacks) it returns
a promise that is resolved when the main thread becomes idle and
the window has completed an animation frame tick.

UltraBlame original commit: 3624871612dd4a23c604cadc5c67cdded584e881