Bug 1491744 - Give the More Actions button and popup menu in doorhanger notifications a better accessibility label, r=Gijs
When screen readers speak the button and subsequent menu when it pops up, the current full prompt is spoken by screen readers. This patch adds a localizable string that is spoken for both instances instead, but it not being displayed visually. This streamlines and improves the screen reader experience for this control and its menu popup a lot.

Differential Revision: https://phabricator.services.mozilla.com/D5999

UltraBlame original commit: 2b62c236e468302ff31b4991d627f6115bd4c00c