Bug 1488206 - TelemetryStopwatch: error requesting elapsed time for nonexisting stopwatch r=Honza
Because firefox can be closed with the toolbox in a half initialized state telemetry can throw if it expects a tool to be closing when no tools are fully initialized although this is completely expected. Here we simply choose not to report those errors because this is expected behaviour.

Differential Revision: https://phabricator.services.mozilla.com/D4869

UltraBlame original commit: 705f3c3d5d60ed20c255fbe33d79e0fc8f1e4526