Bug 1491117 - Do not add crypto info to unencrypted samples from encrypted WebM tracks. r=jya
WebMs with encrypted tracks may have unencrypted samples in these tracks.
Previously we would populate some of the crypto metadata on these samples. This
data was correct, but it was potentially misleading to include crypto metadata on
clear samples.

This changeset alters the behaviour so that we do not populate any such data for
unencrypted packets. This should not alter existing behaviour, notably the
Widevine CDM version 9 should continue to work. However, this change makes our
samples easier to feed to version 10 of the CDM. Without this change, we would
need to do extra conversion steps to appease the new CDM.

Differential Revision: https://phabricator.services.mozilla.com/D6183

UltraBlame original commit: 01b434fc56830fb16519d163db371dd555c113f3