Bug 1490948: Don't set RUST_TARGET to 0 in autospider; r=sfink
This argument is passed as rustc's --target, and it's obviously not a valid
target. It seems there was a good reason to set it as is, but it vanished over
time and the hack remained. This patch gets rids of it.

UltraBlame original commit: 9d57ac1bef3af5233c631ad2ba9bee323680d04f