Bug 1492341 - Bump mozrunner version to 7.0.2; r=tcsc
TPS is currently depending on mozrunner v7.0.1, which depends on an earlier version of mozprofile. For TPS to be able to use the latest version of mozprofile we need to release a new version of mozrunner with the recent mozprofile version identifier update.

Differential Revision: https://phabricator.services.mozilla.com/D6406

UltraBlame original commit: 483e333ad24b323c1326a8df7414500c08ae433b