Bug 1493403: AccGroupInfo::TotalItemCount: Don't assume a container has children. r=eeejay
The code does check for the case where a container has no children.
However, before it does so, it tries to check whether it needs to skip a text leaf child.
This check previously assumed there was a child to check, which crashed for a container with no children.

Differential Revision: https://phabricator.services.mozilla.com/D6588

UltraBlame original commit: 0648f196a589a3ced8f8b9a8509cb7b0211180f1