Bug 1489224 - Enable adb scanner only if addon status is installed;r=daisuke
Using await check() to see if ADB is running is causing aboutdebugging to
leak on debug windows test-verify.

Ultimately we should be able to call ADB scanner without worrying if
the addon is installed or not but for now we should keep things simple
and only start the scanner if the extension is available.

UltraBlame original commit: 26719c685fd34be70dc98012b393525036e83335