Bug 1410820 - Enhance webconsole test helpers; r=bgrins.
1. Allow passing a selector to waitForMessage helper function:

This will allow to wait for specific message (warning, error, result, …).
This patch also fixes 2 tests that were already passing an erroneous selector
to waitForMessage.

2. Add an executeAndWaitForMessage function:

This will make a lot of our tests much easier to read. The function takes
an input, an expected output and an optional selector, execute the input
and resolves when a message matching the provided text (and selector
if passed) is displayed in the output.

Depends on D6903

Differential Revision: https://phabricator.services.mozilla.com/D6925

UltraBlame original commit: a7edd08b7f6d0755191a8eadb000ed63bc10e082