Bug 1494308 - Use consistent logger in wpt commands r=ato
Before we were using a different logger for the manifest download and the actual test run.
This caused timestamps to get reset in a confusing way. Now create the logger early and
share it for all the subseteps.

Depends on D7171

Differential Revision: https://phabricator.services.mozilla.com/D7172

UltraBlame original commit: ace8fc8c112ec192ae853157875f1be6ea43a755