Bug 1493649: Fix IPDL test compilation (r=jld)
This patch allows IPDL tests to compile -- in other words, it fixes the build when the --enable-ipdl-tests flag is present.  Most of the fixes are simple API updates but we also needed to fix an issue with the constness of nsTArray<Shmem> parameters that are defined in IPDL structs.

This patch does not fix the tests themselves -- some IPDL tests will still fail when run.

Differential Revision: https://phabricator.services.mozilla.com/D6705

UltraBlame original commit: d7057bcc6e824a5cbdc5053c419874a9700af8ef