Bug 1486819 - support mozParams in webext search engines r=aswan,adw,mkaply
mkaply for overall search engine api changes
adw for searchservice changes.  note that a small part of it will be removed in favor of the fix from bug 1485508
aswan for webextension side.  note that I want to do better with the distribution signal, that can be in bug 1488517

Differential Revision: https://phabricator.services.mozilla.com/D4473

UltraBlame original commit: 1f268b2058c9e498547c2a902127fa45a93673f8