Bug 1494707 - Add a null check for a retrieved row accessible in HTMLTableAccessible::CellAt, r=surkov
When checking for an accessible if it is a table row instead of a table cell, when retrieving the actual row at the given index, null check it to make sure we don't pass an invalid accessible to the TableAccessible::CellInRowAt method. I accidentally omitted that null check in the updated patch for bug 1486668.

Differential Revision: https://phabricator.services.mozilla.com/D7162

UltraBlame original commit: b5a47b972169a6fea58589a0513ddce7b289c7b9