Bug 1470504 - Make layout.scroll.root-frame-containers a Live pref. r=mattwoodrow
All of its callers are in painting code, so changes will take effect on the
next paint, so there is no need for it to be a Once pref.

Making it Live allows us to selectively enable it in specific mochitests.

Differential Revision: https://phabricator.services.mozilla.com/D7340

UltraBlame original commit: a02b46f1567ed9dad5c294b862c9c5ec6d48bea8