Bug 1489327 - Start animation once after a MozReftestInvalidate event is received in continuation-opacity.html. r=kats
The animation in this reftests runs on the compositor.  In the mean time,
reftest harness waits for the state where there is no pending paint in the
initial phase (STATE_WAITING_TO_FIRE_INVALIDATE_EVENT, i.e. before sending
a MozReftestInvalidate event).  So if the animation starts running on the
compositor before a MozReftestInvalidate event is received, it means that
the reftest harness has to wait for the 'no pending paint' state until the
animation finishes because the reftest harness keeps flushing styles in the
initial phase which means the animation causes a paint on every flush.

To avoid above situation, we start the animation in question after we get a
MozReftestInvalidate event.

Differential Revision: https://phabricator.services.mozilla.com/D7681

UltraBlame original commit: a2ecbcdcec39d0d478213034aa0fd4cd20e32552