Bug 1492704 - 2. Keep track of the auto-fill focus root node; r=droeh
The crash is happening because we may not have a full node tree when
looking at a particular node; in that case we may not be able to find
the root node. This patch makes us keep track of the root node from the
beginning, so we always know what the root node is.

Differential Revision: https://phabricator.services.mozilla.com/D6758

UltraBlame original commit: 538da6bb518b6a794a81ea76584c190dd66e17e9