Bug 1494896 - Make hybrid content telemetry init reject on older versions of Frefox. r=chutten
This patch enables the HCT init promise to reject even for Firefox
versions (before 63) that do not have the required chrome code
to communicate back to the library the lack of privileges or that
HCT is disabled. It introduces a timeout that rejects the promise
after 3 seconds.

Differential Revision: https://phabricator.services.mozilla.com/D7693

UltraBlame original commit: 6a61b59ee9931e5868320c6fdda7a0f677c9b71a