Bug 1475139 part 3 - Add serialization support for nsTHashtable with nsUint64HashKey. r=froydnj
I'd like to use a nsTHashtable acting as a set in a struct that will be passed over IPDL, but
couldn't find any ParamTraits implementation for it. It'd be nice to make the implementation
generic, but I couldn't find an easy way to do it.

Differential Revision: https://phabricator.services.mozilla.com/D6782

UltraBlame original commit: ba9a0af62c4b60260cb90b7cc4673dec0e25fd7d