Bug 1494535 [wpt PR 13235] - Use adjusted relatedTarget on WindowEventContext, a=testonly
Automatic update from web-platform-testsUse adjusted relatedTarget on WindowEventContext

In event capturing phase, when the event is dispatched on window, the
related target was not set properly (it uses the state of the event
at the target, which means related target is the same as when it was
at the target). This might cause leaks when the event's relatedTarget
is inside a shadow tree. This change makes it so at window context,
the relatedTarget used is the relatedTarget of the last
NodeEventContext.

Bug: 817222
Change-Id: I1edea3977202b0a05e3e8ea46a5b8c1edb4b834d
Reviewed-on: https://chromium-review.googlesource.com/1248244
Reviewed-by: Hayato Ito <hayatochromium.org>
Commit-Queue: Rakina Zata Amni <rakinachromium.org>
Cr-Commit-Position: refs/heads/master{#594606}

--

wpt-commits: dfd2d032e8818f2536eb4cceae723eab9ec9ed03
wpt-pr: 13235

UltraBlame original commit: b8df96831cdbbcd1e47a7dfabd688b8a42dcc16e