Bug 1494644 [wpt PR 13240] - WPT: service worker: Fix potential race in navigation-redirect.https.html., a=testonly
Automatic update from web-platform-testsWPT: service worker: Fix potential race in navigation-redirect.https.html.

Before this CL, the test makes a frame which posts a message to the test
window. But if the test fails, we might not consume the message with
`await` and move on the next test. It's possible the next test gets an
unexpected message from the previous test.

I think a related issue caused the timeout after the Clients variant was
added to the test, since the message handler was created after the frame
was created and it was possible the message arrived after that.

Fix things by explicitly asking the frame for the message when we're
ready for it.

Bug: 876223, 778497
Change-Id: I82b8685df2977aeadea7475965ecccf69bd33111
Reviewed-on: https://chromium-review.googlesource.com/1249423
Reviewed-by: Ben Kelly <wanderviewchromium.org>
Commit-Queue: Matt Falkenhagen <falkenchromium.org>
Cr-Commit-Position: refs/heads/master{#594723}

--

wpt-commits: 8a07843f73f18ca87c2751938eb1f6f393aa7875
wpt-pr: 13240

UltraBlame original commit: bb2525c6a01b779647f126c6e2939dad3b37270a