Bug 1493779 - Gracefuly handle OOM when setting a buffer on ConvolverNode.  r=karlt
When OOMing when allocating the temporary buffer, we return an error from the
ctor via an output parameter, and make the ConvolverNode output silence.
Additionaly, a warning is issued each time we fail to set a buffer to the buffer
property of a ConvolverNode.

Differential Revision: https://phabricator.services.mozilla.com/D6936

UltraBlame original commit: 0db994314ad3fe2a9cd2478ef0df0f121bed3374