Bug 1426245 - Allow partial removal of places oberver listeners r=mak
This was a bug in the original design that just hadn't shown up yet.
If I were to listen to, say "notification-a", and then try to remove
my listener for "notification-a" and "notification-b", we would end
up with a dead listener staying in the list. This fixes that.

MozReview-Commit-ID: KhYQSJaBDF9

Differential Revision: https://phabricator.services.mozilla.com/D4605

UltraBlame original commit: 01f7327625291933b8886befa8fb561ced401b3d