Bug 1357561 - ConstraintError should mention which constraint is not being satisfied. r=tt
The creation of a new index or an object store may trigger a
ConstraintError if any of those elements already exists. Unfortunately,
the information provided by the ConstraintError is not enough to
determine the conditions that triggered the error.

This patch overrides the generic ConstraintError message with more
specific information about the duplicated index or object store.

UltraBlame original commit: d0c617e2d35101856d1f26431520a65187f29c78