Bug 1497757 [wpt PR 13445] - Allow posting a SharedArrayBuffer to AudioWorklet, a=testonly
Automatic update from web-platform-testsAllow posting a SharedArrayBuffer to AudioWorklet

The HTML spec only allows passing SharedArrayBuffers to an agent in the
same agent cluster. Worklets currently do not belong to any agent
cluster, so a SharedArrayBuffer can not be shared with a worklet.

However, it is intended that this is possible; see, for example,
https://github.com/w3c/css-houdini-drafts/issues/380 and the
AudioWorklet article here:
https://developers.google.com/web/updates/2018/06/audio-worklet-design-pattern#webaudio_powerhouse_audio_worklet_and_sharedarraybuffer

This change funnels the agent cluster ID through when creating a
ThreadedWorklet, so a SharedArrayBuffer can be shared as long as the
creator of the Worklet's thread is in the same agent cluster. It is not
clear that this is the behavior that will be specified, however.

Bug: chromium:892067
Change-Id: If1a2187ae38da41f2389538c07e7b04921c6128f
Reviewed-on: https://chromium-review.googlesource.com/c/1262932
Commit-Queue: Ben Smith <binjichromium.org>
Reviewed-by: Hongchan Choi <hongchanchromium.org>
Reviewed-by: Hiroki Nakagawa <nhirokichromium.org>
Cr-Commit-Position: refs/heads/master{#598619}

--

wpt-commits: 2c89bbecfab9a69190906abd7610c3bc62303dd4
wpt-pr: 13445

UltraBlame original commit: a9f42748483700ec4e6d4f79bcc8ffef9f0b4824