Bug 1497446 - Follow-up to repair incorrect vendoring of lalrpop-snap. r=froydnj
This (large) file exists in the lalrpop-snap crate and should be part of
the vendoring of that crate. However it seems to have been accidentally
removed in bug 1497446. This patch adds it back by running
  ./mach vendor rust --build-peers-said-large-imports-were-ok
on a clean m-c tree.

Differential Revision: https://phabricator.services.mozilla.com/D8863

UltraBlame original commit: 3361fcc40ea294df9eed70a218c0eafe11bf60a7