Bug 1494069 - [mozlint] Properly exclude files in LineType linters, r=rwood
This fixes a latent bug that is currently not being hit (by sheer luck). Basically
the 'ignore' argument of a FileFinder object needs all paths to be relative to the
base. As luck would have it, most of the time it worked out that way if you were
running |mach lint| from the root of the repo.

However there are edge cases where this will cause an 'exclude' directive to get
missed. Plus this bug is about to be exposed 100% of the time in the next commit :).

Depends on D8842

Differential Revision: https://phabricator.services.mozilla.com/D8843

UltraBlame original commit: 3cbd82a01bc27518e21aae6668292c0b6f25802b