Bug 1356427 - P1. Remove redundant code from PendingLookup::LookupNext. r=francois
When mBlocklistCount is set, we will call PendingLookup::OnComplete() to
block this download. The |if(mBlocklistCount)| in PendingLookup::LookupNext
will never be true.

mBlocklistCount member variable is also useless after removing the code.
But I think we can keep it since we have mAllowlistCount and we might also
need it in the future.

Differential Revision: https://phabricator.services.mozilla.com/D8682

UltraBlame original commit: ced08ec47ace994bf5c5f1e4052052e7cb9bab30