Bug 1496814 - Show only search suggestion results when only an engine alias and query are typed in the address bar r=mak
This is based on the patch in bug 1496811.

This patch looks a little worse than it probably is. (Maybe not by much.) Some of this is indentation changes, moving code around, renaming, and adding jsdocs. It looks like we missed the boat on uplifting this (and the other couple of bugs) to 63, so there's not a super-pressing need to keep the patch minimal.

PlacesSearchAutocompleteProvider assumes you're fetching suggestions from the current engine, so I had to modify it to take an engine. While I was doing that, I got a little frustrated with some of its implementation, naming, and interface. It seems like it was written to be a little more generic than it ended up being? There doesn't seem to be any need for it to return generic "match" objects instead of simply engines and `{ suggestion, historical }` objects, for example. The "defaultMatch" concept also doesn't make much sense IMO, especially with the aforementioned changes. So I made some improvements, hopefully, and I also added some jsdocs.

Differential Revision: https://phabricator.services.mozilla.com/D8818

UltraBlame original commit: 665fecc387403fd9336daaaa6bfa619dc5b7baed