Bug 1500475 - Modify browser_jsterm_content_defined_helpers.js to find cause of intermittent; r=bgrins.
The log we currently receive are not really
actionable (we are using ok, so we don't know
what the received value actually is). Switching
to is should give us what the value is when the
test fails, which might help us fix the failure.

Differential Revision: https://phabricator.services.mozilla.com/D9270

UltraBlame original commit: a159b99e2a0b8516b852e47e257c8752fa3b55b6