Bug 1498420 Convert talos tart extension to a webextension r=mconley
The biggest change here is that the tart.html page that drives the
test (not to be confused with the pages loaded in new tabs during the
test) moves from a chrome: page inside the extension to a regular http:
page.  That also required revamping the communication between tart.html
and the extension.  The rest of the changes are just the packaging and
startup mechanics for the test extension.

Differential Revision: https://phabricator.services.mozilla.com/D8807

UltraBlame original commit: 3133f6ab1bd8ea2aa261f8e9024ab3fb8eb0ddcb