Bug 1498531 - Localize Webconsole WebReplay Jump icon; r=loganfsmyth,Honza.
This patch adds localization for the WebReplay Jump icon, and uses
the same terminology as the one used in the context menu that triggers
the same action.
The Jump button was used in-place of the existing level icons (Error, Warning, …),
and was only displayed when the message was hovered. We now ensure the
level icon is always visible and that we only show the Jump icon when the
message is hovered.
Finally, the button was styled targeting the title attribute in CSS, which
seemed a little brittle. We now use a dedicated class which should
be safer and more future proof.

Differential Revision: https://phabricator.services.mozilla.com/D8533

UltraBlame original commit: aa628ce24d80d9966a0bcdfce6bc0469da509c34