Bug 1501632 - Return an unsubscribe function from SourceMapURLService.prototype.subscribe; r=loganfsmyth.
The SourceMapURLService has a subscribe and an unsubscribe
functions to respectively listen and stop listening for
source map changes on a given location (url + line + column).
The unsubscribe function need to be called with the same
parameters as the subscribe function, which means the consumer
need to keep a reference to the callback.
By making the subscribe function return the unsubscribe function,
this makes things a bit easier.

Differential Revision: https://phabricator.services.mozilla.com/D9784

UltraBlame original commit: 8967e494e4e35d42acbefe4919625d18f74b4ab8