Bug 1502421 - fix undefined behavior in PendingFullscreenChangeList; r=mccr8
The order of evaluation of x and y in x->Func(y) is undefined, but the
code in PendingFullscreenChangeList depends on x being evaluated before
y.  We should make the ordering explicit.  (The ordering is fully
specified in C++17, but we're not ready to make that switch yet.)

UltraBlame original commit: 6baeb0cb9aeb5fd629fc697ea68f814328aaba43