Bug 1500963 - Use the Stacktrace component in Error reps; r=bgrins.
This patch makes use of the new renderStacktrace
prop in Reps to render Error objects stacktraces
using the shared Stacktrace component.
This way we are showing stacktraces in a consistent
way, and as a bonus, we support sourcemap in Error
object.

Depends on D9922

Differential Revision: https://phabricator.services.mozilla.com/D9923

UltraBlame original commit: 006509a8370fc1a1c83ba9da96737319b44b0cf2