Bug 1487022 - Use MOZ_RELEASE_ASSERT for builtins that take int32 values. r=tcampbell
The arguments to the intrinsics here are not guaranteed to be int32s,
however, if a double enters the intrinsic, that's probably a bug in our
builtin javascript. Values should be converted to int32 before being
passed in, with `|0`.

Differential Revision: https://phabricator.services.mozilla.com/D8446

UltraBlame original commit: 1cf53a43031167a372e7d3e5bfb096124d4e74af