Bug 1492128: Add `VOLUME` definitions to base debian images; r=gps
For tasks that only need python, the base images are sufficient. The in-tree
code verifies that the caches are setup as volumes in the image being used,
so set the caches in the base images as well.

Differential Revision: https://phabricator.services.mozilla.com/D10148

UltraBlame original commit: 778d969a1dabf52739c0d05f0a8ef4e6239d7d3d