Bug 1505024 - Disable codeMirror search addon in JsTerm; r=Honza.
This patches fixes a bug where hitting Cmd+F while
the filter input of the console was focused was
triggering codeMirror search addon, displaying the
search UI *in* JsTerm.
To prevent this, we add the ability for any consumer
of the Editor component to disable the search addon.
In the editor, in such case, we simply don't initialize
the event listeners that activate the search addon.
A test case is added to ensure we don't regress this.

Differential Revision: https://phabricator.services.mozilla.com/D11037

UltraBlame original commit: e9a8115d4c8f9eae72304b266f0307214ca4a857