Bug 1505328 - Remove the theme transition from the title bar. r=dao
This patch solves half of the problem but it will still show the transition when the window regains focus. I am OK with removing the transition on the title bar, we will still have it on the navbar and bookmarks toolbar.

Differential Revision: https://phabricator.services.mozilla.com/D11282

UltraBlame original commit: c58d0aab32b0d95a59306d954766bf5ef7abfcfe