Bug 1376873 - Disable Mid support in RtpDemuxer; r=mjf
The only use of Mid in the current webrtc.org code is in the unit tests.
RtpStreamReceiverController only allows adding sinks using SSRCs. Because
of this, we'll end up dropping packets in the RtpDemuxer with the current
code as none of our Mids will be recognized.

Tip of webrtc.org fully supports using Mids, so we'll be able to enable this
code again after the next update.

Differential Revision: https://phabricator.services.mozilla.com/D7442

UltraBlame original commit: b3ba8452e77105c72f6ddbc49cbe5a53dbea1507