Bug 1504020 [wpt PR 13840] - Add support for resizeMode in MediaStreamTrack.getSettings(), a=testonly
Automatic update from web-platform-testsAdd support for resizeMode in MediaStreamTrack.getSettings()

This CL exposes the current resizeMode setting to the Web Platform.
As of this CL, the resizeMode is 'none' if the track resolution
corresponds to the native resolution of the video source, without any
resolution adjustment. It is 'crop-and-rescale' if rescaling is
necessary.

The resizeMode currently cannot be set explicitly with getUserMedia() or
applyConstraints(), but this functionality will be added in a follow-up
CL.

Drive-by: Add separate WPT tests for all getSettings() properties.

Intent to Ship:
https://groups.google.com/a/chromium.org/forum/#!topic/blink-dev/V2srjdzRCXE

Bug: 854980
Change-Id: Id4c683e6b2f228b17cf2354cb7fc589ca9e593a6
Reviewed-on: https://chromium-review.googlesource.com/c/1307448
Commit-Queue: Guido Urdaneta <guidouchromium.org>
Reviewed-by: Henrik Boström <hboschromium.org>
Reviewed-by: Philip Jägenstedt <foolipchromium.org>
Cr-Commit-Position: refs/heads/master{#604562}

--

wpt-commits: dcf7730d7300f236459aa20990e14d46f3517916
wpt-pr: 13840

UltraBlame original commit: 163da2b81d74629d03d43e2a431b1b0ffa1d38ab