Bug 1485324 - (Part 2) Remove obsolete subset of fonts not used on current element; r=gl
Depends on D11505
The `otherFonts` field in the Redux store slice for Fonts was used by the old Font Inspector to differentiate
between fonts on the element and fonts elsewhere on the page. This is no longer used with the Font Editor.

Differential Revision: https://phabricator.services.mozilla.com/D11506

UltraBlame original commit: 13a15cc93107cbe2a17a332ba364f79c4e04c5ce