Bug 1503718 - Part 1: Use the `unwrapped` prefix. Covers spec sections 3.2 to 3.4. r=tcampbell
Each patch in this stack deletes comments that are redundant with the new
naming convention.

In ReadableStreamTee_Cancel, we have a variable named `unwrappedReason` whose
purpose is to create a properly wrapped verison of `reason`. It's a little
vertiginous. But I think this is what the new convention demands and it's not
so bad.

Also in ReadableStreamTee_Cancel, step 4.c., we wrap `cancelResult`, which does
not have an `unwrapped` tag.  This is because we switched realms between the
declaration of `cancelResult` and the line of code where we're going to use it.
I think this just means the convention is never going to make all correct code
obviously-correct and all wrong code obviously-wrong. Still an improvement.

Differential Revision: https://phabricator.services.mozilla.com/D11683

UltraBlame original commit: 9fb1b0aeef48a633c30593c5cef6ee10bcfd7f6d