Bug 1506118 - Use chromeUtilsWindow to start the parser worker; r=Honza.
As we were using the browserWindow to start the
parser worker, this was causing an exception
when evaluation from the browser toolbox.
Using chromUtilsWindow fixes the issue.

There is no tests yet as I'm not sure there's
an easy way to test things in the browser toolbox.

Differential Revision: https://phabricator.services.mozilla.com/D12101

UltraBlame original commit: 1bd7fb2e62580ec67db80611659ca162293df0ba