Bug 1501787 - Reset entered counts on stub attachment r=djvj
Resetting the entered count on an IC stub chain makes interpretation of the
counter values easier. For example, if we see an IC chain like this

    A (1000) -> B(800) -> C (400) -> D (200) -> FB (100)

We can say that there have been 100 cases not handled by chain, (and we did
not attach new stubs for those cases), and B handled the most (400) queries
to the IC chain.

Differential Revision: https://phabricator.services.mozilla.com/D11883

UltraBlame original commit: c376952fc91969b904b31cc0da568bb716cc17df