Bug 1507479: Make WorkerPrivate::mIsSecureContext const r=perry,asuth
This member was never changed after it was set, but it couldn't be const
because WorkerLoadInfo used a custom StealFrom() function instead of an
idiomatic move constructor and move assignment operator. This commit
replaces StealFrom with its idiomatic counterparts, adds a function that
determines a new worker's secure context status, and uses that function
to initialize the now const mIsSecureContext.

Making constant members const is part of a greater effort to ensure that
data is either modified from a single thread only, or access to it is
synchronized properly.

Differential Revision: https://phabricator.services.mozilla.com/D12030

UltraBlame original commit: 7efd8b5c5b934ba3897fb33ac19720c07f4d4057