Bug 1508905 Remove from clang-format-ignore WebAudio files forked from Blink r=Ehsan
Upstream of this fork has reformatted with clang-format, and so there seems
little value in leaving this fork in its original format.  It can be
reformatted to the new Mozilla style with clang-format, which will probably be
closer to the format upstream is now using.

Differential Revision: https://phabricator.services.mozilla.com/D12508

UltraBlame original commit: 34e7b9fa3a761d8e761c9aeaa50983d0cb3cc50c