Bug 1508870 - Make a distinction between an empty returned string versus a null pointer, r=Jamie
When determining if an author provided an empty alt attribute versus no alt attribute at all, our image accessible name is either an empty string or a null string. Screen readers can then choose to guess the name from the source if the author did not explicitly provide an empty alt text and thus marked the image as decorative.

The accessible handler for the Windows remote processes did not account for this distinction, always returning what appeared to be an empty string. This was found while determining why NVDA wasn't guessing the SRC from a non-labelled graphical link. NVDA checks to see if the name returned by the link accessible is null, and then decides to look for the graphical children if that is the case.

Differential Revision: https://phabricator.services.mozilla.com/D12532

UltraBlame original commit: af8241e64ecb7096c314c1a174479d11ab18163d