Bug 1507785 - Use logical, not physical, frame size at block exit. r=bbouvier
Background:

When we branch out of a block we must adjust the physical stack
pointer along the edge so that it will be equal to the physical stack
pointer as it is when we fall out of the block at the bottom.  But
along the edge we do not do anything to adjust the logical stack
pointer, because the logical stack pointer is determined by the
non-branching path through the code.

But when we fall out of a block we must adjust both the physical and
the logical stack pointers: the logical stack pointer may need to pop
a few items to get to where it was on entry to the block (the stack
can be higher because of pushes followed by an UNREACHABLE), and if
the code is not dead we may also have to deallocate some stack memory
by changing the physical stack pointer.

Previously, we used values for the physical stack pointers as guards
on the block-exit code, but this is WRONG since it prevents us from
doing something when the physical stack pointer does not change, as
the case is on ARM64 when we do need to pop something logically but
the stack stays within the currently allocated chunk - we would end up
doing nothing when we should adjust the logical stack pointer.
Instead, we should use the logical stack pointers for the guard, and
then popChunkyBytes will take care of translating the difference in
logical pointers for us so that a chunk is popped if that is required
by the amount of deallocation.

To catch these errors earlier we add an assertion on the logical stack
frame size that is tested for every instruction we compile.

The test case is reduced from the fuzzing test that found the bug.

UltraBlame original commit: 45b378c260d9308b6b067e42da3412961d431128