Bug 1505774 - Introduce nullref type. r=luke
Mostly this is straightforward: add NullRef in the various type enums
and make sure we handle it everywhere (sometimes it's valid; other
times it's a fatal error because it should not appear in that
context).  Some type calculus code had to move out of WasmOpIter and
into ModuleEnvironment in order to be available from the decoder
generally.  Also, since NullRef is not an expressed type but only the
type of null constants in the type checker until we know better, we
have to be careful to avoid using it as the value type of null values.

The text syntax for a null constant in the s-expression form is
"(ref.null)", with the parens required.  In the stacky syntax the
parens are not required.

The encoding of ref.null is now incompatible with our old encoding, so
there's a required update to the gc-feature opt in version (to version 2),
code tagged with version 1 will no longer run but there's a sensible
error message printed for that.

UltraBlame original commit: 5219f57277c4b6b76b61d435227bd666afee59ea