Bug 1506546 - Refactor about:debuggings. r=yulia,jdescottes
I did a dedicated patch for converting both about:debuggings as it is slightly more
complex as we have to tweak redux data, wrappers, mocks.
This patch also convert a manual "reload" request being done by about:debugging
and instead use protocol.js front to do it.
Also, I moved isLegacyTemporaryExtension to the front as it depends on accessing the form
and it should be better to keep form processing to the fronts, if possible.

MozReview-Commit-ID: 16qZkuCBp9b

Depends on D12576

Differential Revision: https://phabricator.services.mozilla.com/D12577

UltraBlame original commit: 9abf7054a03ee08e6beb910e2b61448cc74934ad