Bug 1509462 - Make DecoderDoctorLifeLogger constructor/destructor protected. r=gerald
DecoderDoctorLifeLogger is used as a base class in many classes but it does need to have a virtual destructor to avoid adding a vtable in the class. Thus the constructor and destructor become protected to avoid being used as base pointer to delete the real object.

Differential Revision: https://phabricator.services.mozilla.com/D13074

UltraBlame original commit: 209ed62867bf0159b4d9a49fdd98bcea43c64287