Bug 1509930 - Simplify CycleCollectWithLogsChild's lifetime, r=jld
Previously this actor would call `Send__delete__` in its destructor, meaning
that the lifetime of the actor is a bit hard to follow. This changes the actor
to instead use a separate object for XPCOM, and use refcounting universally.
This should also help with the transition to universally refcounted actors.

Differential Revision: https://phabricator.services.mozilla.com/D12955

UltraBlame original commit: a06e4cff09e6fe6ca41626310d6494a3d276772a