Bug 1508748 - Avoid duplicate tracking of the same removed CSS declaration. r=pbro
When a declaration is disabled, it is tracked as removed in the Changes panel.
When deleting a disabled declaration, we take care not track it as removed again.

Includes stricter checks when matching previously tracked added and removed declarations.

Differential Revision: https://phabricator.services.mozilla.com/D13616

UltraBlame original commit: 9d1eeceace1b9b954791a2554c4e497c6fe8ab46