Bug 1507212 - Guard against the compositor receiving a delayed FlushApzRepaints message. r=botond
On Android the FlushApzRepaints API is used as part of the code flow of
zooming/scrolling to a focused input field. In mochitests this code can
run but the message might arrive late, after the test suite is down and
the browser is shutting down. Depending on timing, the parent process
compositor bridge instance might already have been cleared, in which
case the code would crash/fail an assertion. This patch makes the code
handle this case more gracefully.

Differential Revision: https://phabricator.services.mozilla.com/D13941

UltraBlame original commit: 537b39da583098454476d886dcf8edd9c203c719