Bug 1511138 - Fix LightweightThemeConsumer's use of getComputedStyle. r=jaws,mconley
See D13472 for spec quotes and such. Other browsers don't allow
getting computed styles in disconnected subtrees and we agreed to follow suit
(it does make sense because when you're not on the flat tree it's not defined
what you're supposed to inherit from, specially in presence of Shadow DOM).

Also, it allows the style system to rely on the DOM being in a sane state.

Differential Revision: https://phabricator.services.mozilla.com/D13551

UltraBlame original commit: a99600391704f8752e6f05e312fe8e8cd588715a