Bug 1507487 - Handling of importScripts by ESLint should fallback to the file name if there is no entry in modules.json. r=mossop
We already have a fallback for items that aren't in a worker scope, we should have the same for ones that are. This means we don't need single maps (foo.js -> foo) in modules.json, and also we can identify more as explicit variables, so that no-unused-vars can detect them.

Differential Revision: https://phabricator.services.mozilla.com/D13299

UltraBlame original commit: 6f69dc19ed7cd1277fda93cd8e3129528f23bbc4