Bug 1513023 - Wait for wpt tests to finish across reloads, r=ato
When wpt is used by developers they often want to reload a test that's
running to allow iterative development (or running under devtools,
etc.). The recent switch to target marionette commands at the test
window rather than the opener window broke this as waiting is
implemented as an async script that never returns (i.e. a promise that
never resolves) and when this is in the test window navigating that
window causes the script to return.

A simple, but not particuarly elegant, solution here is to simply
rerun the script in the new window.

Differential Revision: https://phabricator.services.mozilla.com/D14087

UltraBlame original commit: 0a36f92c457acb05b5f9ee147c3ae42a8a6d9de3