Bug 1512799 - Don't force-disable WebRender on non-qualified hardware when the qualified pref is set. r=kats
The gfx pref machinery asserts when user-level values are set that
environment and runtime prefs have not been set. This effectively means
we need to make all user-level decisions before calls to ForceDisable(),
and the code in question here violates that contract.

This branch is not necessary for correctness, since the pref basically
means "enable WR on qualified hardware", and we can just ignore that
request if the hardware is not qualified. It does provide extra
about:support information, but we also get that information from the
recording in WebRenderHardwareQualificationStatus, which creates a
separate decision log for WEBRENDER_QUALIFIED. In this case, it will
report something like:

"WEBRENDER_QUALIFIED: blocked by env: No qualified hardware"

which should be enough here.

Differential Revision: https://phabricator.services.mozilla.com/D14082

UltraBlame original commit: 2e17d019a6a41757eee90f72f204c20417f4a5eb