Bug 1440779: BaseAction: move _preExecute call out of constructor r=mythmon,Gijs
Having _preExecute be called in the constructor makes it hard to test
it, but because it's a lifecycle hook that subclasses are intended to
override, testing it is very natural.

While we're here, move the initialization of AddonStudyAction from
_preExecute to the constructor, since it doesn't really make sense for
AddonStudyAction to be constructed without all members initialized.

Differential Revision: https://phabricator.services.mozilla.com/D13572

UltraBlame original commit: ecb4a2b0d498e35dc99f76184554769c07bc5451