Bug 1514844 - Remove the per-process notification cache. r=asuth
This assumes that all of the notifications for a given origin must be in the
same process. With this patch, we'll always go back to the parent process to
get the notifications. Up next is limiting our search in the parent process to
only the notifications we're looking for.

Differential Revision: https://phabricator.services.mozilla.com/D14774

UltraBlame original commit: c4eb9f94bc4c52e0489f8e5b7f7fc931ef013816