Bug 1508956 - Display unknown runtimes in about:debugging;r=daisuke,ladybenko,Ola
Depends on D14310. This patch enables displaying devices on which the debuggable
runtime is not started yet. To test it, simply use an android phone with USB debugging enabled
but don't start Firefox yet. You should see the Unknown Runtime item in the sidebar.

Not that once you start Firefox on the device, you need to click refresh runtimes here to get
the sidebar to refresh (more on that in the last patch of the series).

Two things I am not completely happy with:
- calling this a runtime, when really it's just a device that has no runtime.
  but I don't want to start renaming all our state for that. Hope we can treat it as
  placeholder runtime, and forget about it.
- localization for the runtime item name can be done in many ways. Not using getString
  at all makes things really complicated, but I'm open to suggestions if you have any

Differential Revision: https://phabricator.services.mozilla.com/D14313

UltraBlame original commit: 3b9db2fd0679efef5d182d1193bcf6180c97ae70