Bug 1509388 - Part 1 - Don't use a toolbarbutton element for the download-subview-toolbarbutton binding. r=Gijs
Download items in the Downloads Subview of the Library Panel are implemented using a XBL binding that is associated to a "toolbarbutton" element, but inherits from "button-base" in order to avoid the display="xul:button" of the "toolbarbutton" binding. This is because the item contains an inner button, and using a button CSS frame for the outer element would break hover behavior and mouse events.

This changes the binding so it doesn't use a "toolbarbutton" element, which is a different way to avoid the outer button CSS frame, and eventually allows removing support for the "display" attribute in XBL.

Differential Revision: https://phabricator.services.mozilla.com/D14572

UltraBlame original commit: fe181f5f1a9673580dfa5e89d094f91f1621ae12