Bug 1508956 - Check for ADB runtimes updates periodically;r=daisuke,ladybenko
Depends on D14318. This patch introduces a setInterval to check if
new runtimes have appeared on monitored devices, directly from ADB. This is
one of the scenarios where we don't get an update from ADB. ADB is able to
track devices appearing or disappearing but here we are looking for a new
process, containing a specific string. Out of the box I don't think we can
do anything for this.

Maybe we could ask the devtools server to "broadcast" its presence when it
starts, but at the moment I am not sure how we should implement that.

Differential Revision: https://phabricator.services.mozilla.com/D14324

UltraBlame original commit: ebd7cddb27181ee6a102d013321d542e18693f6d