Bug 571074 - Refactor nsContentUtils::TransferableToIPCTransferable to make error handling more obvious. r=smaug
While trying to handle a failed GetTransferData call I made some mistake, mostly because this function has incredibily
confusing control flow. I would actually assume some of its current behavior is not intentional.

For this we assume that data == nullptr <-> NS_FAILED(rv), this is probably not 100% correct, but doesn't seem to
matter in practice. (This is a change in behavior that is intentional, because we want to check return value)

If we trace the data == nullptr case through the old code, we realize that for aInSyncMessage we will just continue.
(This is because nullptr doesn't queryInterface to anything)

Differential Revision: https://phabricator.services.mozilla.com/D15572

UltraBlame original commit: b90c87fbab033dc146f45ed86720800564a4544e