bug 1392603 - Remove toolkit/crashreporter/tools/win32. r=gsvelto
We have prebuilt dump_syms binaries in toolkit/crashreporter/tools/win32 from
the days when we didn't build it as part of the build. There's one per MSVC
version, since the APIs it uses are version-specific, but we don't actually
support building with the MSVC versions corresponding to the binaries in that
directory anymore, so we should just remove them.

The only code change necessary here is changing the functional test to skip
running the test if a dump_syms binary isn't available.

MozReview-Commit-ID: IBuJZMVfv7C

Differential Revision: https://phabricator.services.mozilla.com/D15717

UltraBlame original commit: 5b395d1d3585f20579c098cd000adfe26eb84831