Bug 1515197 - Remove the "max" from "Max buffer size" label since it's not resizable anymore r=julienw
We made the profiler buffer resizable before that patch but then we decided that
it's not a good idea to do since we are spending so much time to allocate memory.
Since buffer size is constant, we don't need to add "max" anymore.

Differential Revision: https://phabricator.services.mozilla.com/D15656

UltraBlame original commit: f46ff2feaf5f4deee1db87bf701f0ef1553044aa