Bug 1514898 - Disable user impersonation on release r=rstrong
This introduces a DISABLE_USER_IMPERSONATION define when EARLY_BETA_OR_EARLIER is set, if that is present the maintenance service will not attempt to get an impersonation token for the user's updater process, and the updater will not attempt to use any token it is given. The bulk of the changes are restoring the old failure status codes and the tests that expect them, sorry for the noise that causes.

Differential Revision: https://phabricator.services.mozilla.com/D15781

UltraBlame original commit: f184da8d467cc98ff07dfbf880557eac00566c57