Bug 1518799: Show a custom page on startup on Nightly on a specific date. r=Gijs
We want to show an informative message about dedicated profiles per channel to
users of Nightly on a specific date. We currently only have the ability to do
this when the version changes. This adds the ability to show a page once on
startup on a specific date.

This will probably be backed out past that date.

Differential Revision: https://phabricator.services.mozilla.com/D16249

UltraBlame original commit: eace4709948cdf060a36a60c00eeb286cb7acb17