Bug 1484263 - Clean up manifest permission parser and add tests r=aswan
The "permissions" array of the raw manifest is not (and should not) be
used for permission checking, so it is not necessary to strip the
"mozillaAddons" permission from it.

This commit moves the validation of the "mozillaAddons" permission to
classifyPermission, so that the "manifestPermissions" getter (that uses
this method too) accurately reflects the supported permissions of an
extension.

New tests has been added to verify the permission warnings for some
combinations of permissions. This also includes tests that verify
that only privileged extensions can use "mozillaAddons" to unlock
unrestricted schemes.

Differential Revision: https://phabricator.services.mozilla.com/D5527

UltraBlame original commit: 14fee8b5f596471935ef50d75f56d882977185c3