Bug 1489491 - Populate reverse search with input text selection; r=bgrins.
If the user selected some text in the console input
and opens the reverse search UI, we populate the
reverse search input with the selected text and do
a reverse search with this text.
A test is added to ensure this works as expected.

Differential Revision: https://phabricator.services.mozilla.com/D16843

UltraBlame original commit: 529a105310435244ef67c078a810ad92b22d9327