Bug 1521776 - Add new scrollable markup badges; r=gl
Detect when nodes are scrollable on the actor side, and use this
information to display new [scrollable] markup-badges in the inspector.
Also track this dynimically so badges are added/removed when necessary.
Hide this behind a pref for now, until this gets tested by users and
validated.

Differential Revision: https://phabricator.services.mozilla.com/D17246

UltraBlame original commit: 21cb7caef246d374a8f73a14806d9fd328adbf8a