Bug 1355588 - Implement slide-up/down tab opening/closing animation. r=dao
This patch implements a new, preffed-off tab opening/closing animation where the tab slides up and down in place. In addition, it adjusts some tests and code that were relying on the max-width transitionend event happening. The intent of the patch is to land it preffed-off and iterate on the design in the tree

Differential Revision: https://phabricator.services.mozilla.com/D15955

UltraBlame original commit: 5c6cdeb8648d342ef53db0e954189049cc640532