Bug 1517406 - Fix this test to listen in the proper process. r=asuth
Summary:

This test listens for an observer service notification from the
`ServiceWorkerManager`, but with parent-intercept on, it listens in the wrong
process. This uses the magic of SpecialPowers to listen in the correct
process.

Unfortunately, this test still fails because it leaks.

Reviewers: asuth

Reviewed By: asuth

Bug #: 1517406

Differential Revision: https://phabricator.services.mozilla.com/D15635

UltraBlame original commit: 5e1b2e58c22555a247e7fc2f115ada9fec5e6a5c