Bug 1522901 - Include packet.stacktrace when computing repeatId; r=julienw.
This was causing an issue with console.trace calls
triggered from different paths (i.e. the message
was repeated in the console, only showing the first
stacktrace).
This fixes the issue, and a test is added to
ensure we don't regress this.
This also revealed an erroneous test where we were
asserting the buggy behavior.
Doing this modifies the message shape, so we also
need to update the stubs.

Differential Revision: https://phabricator.services.mozilla.com/D18094

UltraBlame original commit: f041fc6b4544cee1d2c502b6659c8bf99bad37e4