Bug 1516665 - Part 3: Additional logging. r=JanH
Added additional logging. In case the agent is not attached but we are not on a
release or beta build, allow createNotification to be called with a null context
so that we may be able to gather more information from the reports.

Depends on D18115

Differential Revision: https://phabricator.services.mozilla.com/D18116

UltraBlame original commit: 45c0ce350507fbe5a5a7885b55c2e8bce39fc00e