Bug 1400426 - [try] Add --exact argument to |mach try fuzzy|, r=jgraham
Passing in --exact reverses the behaviour of the ' operator. For example,
take the query "foo 'bar".

By default: foo is a fuzzy match and bar is an exact match.
With --exact: foo is an exact match and bar is a fuzzy match

Differential Revision: https://phabricator.services.mozilla.com/D16734

UltraBlame original commit: 3ba052addf07b21c78fafad0f2ae5e6f65f09bcd