Bug 1524548 - (Part 1) Add support to filter by ruleId and sourceId when building the changes tree. r=pbro
The diff for this change looks needlessly scary. The only material change is the introduction of the `filter` parameter to the `getChangesTree()` selector to restrict the output to a subtree that includes only the martching rules and sources provided in as arrays of ids in the `filters` argument.

The sources and rules are filtered, if any corresponding ids are provided, then the same logic to build the tree is used.

The meaningful blocks of code introduced are:

```
.filter(([sourceId, source]) => {
  // Use only matching sources if an array to filter by was provided.
  if (sourceIdsFilter.length) {
    return sourceIdsFilter.includes(sourceId);
  }

  return true;
})
```

introduced right after `Object.entries(state)`

and

```
.filter(([ruleId, rule]) => {
  // Use only matching rules if an array to filter by was provided.
  if (rulesIdsFilter.length) {
    return rulesIdsFilter.includes(ruleId);
  }

  return true;
})
```

introduced right after `Object.entries(rules)`.

Adding a newline confused the diff algorithm to consider that much bigger changes have occurred.

Introducing these filtering options lays the ground work for building subtrees of changes per-rule and per-stylesheet to be copied to the clipboard or saved as a file.

Differential Revision: https://phabricator.services.mozilla.com/D18703

UltraBlame original commit: 08fd81ac849bd369901e1d00989499aa0d27dc1c