Bug 1521939 - Correct the offset for reading stack args on ARM64.  r=bbouvier
In the optimized stub, we forgot to account for the frameAlignExtra
word when computing the offset from the SP at which stack args are
read.  The test case finds the problem on the ARM64 simulator, and
does not need any special parameters, just to run long enough for a JS
JIT to kick in.

Also a drive-by fix for an incorrect NaN canonicalization along the
non-toValue path, cf the Double case right above it.  This code
changed recently when I introduced the ScratchFloat32Scope, but the
bug predates that.

Differential Revision: https://phabricator.services.mozilla.com/D18942

UltraBlame original commit: 5d720cbe1187b8ef658ab5098fdaf9f36e532504