Bug 1525129 - Fix intermittent in browser_webconsole_reverse_search_mouse_navigation.js; r=Honza.
Based on the failure logs, it looks like the test fails because
a node highlighter is being triggered while the toolbox is
closed.
Changing one of the input to not display any nodes in the output
should fix the issue.

Differential Revision: https://phabricator.services.mozilla.com/D19341

UltraBlame original commit: 1e8fa611b7ae159304697d3de11aeae23c715bfa