Bug 1526219 - Guard atomicity tests on a plausible core count. r=luke
The atomicity test cases require multiple cores so that multiple
workers can run concurrently; if enough cores are not available then
the tests will just take forever to execute because they'll be waiting
on preemption.  (In practice, the tests will time out.)

So guard execution of the test on the availability of enough cores.

The guard may mean that the tests are not run on all test platforms,
but they should be run on developer systems and on platforms where we
run tests non-virtualized, eg, android hardware -- provided there are
enough cores.  This is a real improvement over the current situation,
in which the tests are marked slow and thus never run at all.

Differential Revision: https://phabricator.services.mozilla.com/D19259

UltraBlame original commit: 491fc771924f074b997d18be782b70e22a07021b