Bug 1527317 - lazy load TabboxPanel in NetworkEventMessage.js; r=Honza.
The TabboxPanel component is the one we use to show the detail
of a given network request. In the console, it's not used until
the user expand a network request, but we do load it at
the same time as the NetworkEventMessage.
We can save some extra work by loading it lazily.

Differential Revision: https://phabricator.services.mozilla.com/D19497

UltraBlame original commit: 3405ab9c5c7b0fc01fe76c93e96eca2a8df8bd77