Bug 1460874 - Part 5: Move out SharedPreferences watching. r=geckoview-reviewers,snorp
For easier testing in Fennec, we want to continue watching the respective
SharedPreferences key and toggle the listener in response to *that* instead of
directly wiring it into our Settings menu (GeckoPreferences), however because
that functionality is Fennec-specific, we move that logic out of the font scale
listener itself.

In conjunction with this, we also decouple the enabled state of the listener
from its attached state.
The enabled state can now be toggled at all times, but unless the listener is
also attached to a context, it simply won't have any practical effect.

Differential Revision: https://phabricator.services.mozilla.com/D17746

UltraBlame original commit: eea63b0bc4ebea70e1305dfa5b5b286e699023c4