Bug 1443560: Convert checks to examine snapshotted events r=Gijs
Because the mock no longer checks whether the format of events is
correct, check the events that Telemetry actually records, in order to
make sure we conform to the event schema.

Some uses check that *no* calls were made to the stub. Although this
is technically still valid (because we don't care about the format of
events in this case), convert even these to check events for
consistency.

TelemetryTestUtils.assertEvents doesn't allow us to pass a message,
but most of these messages don't add anything useful and at least one
is just wrong so it's not a big loss.

Depends on D19540

Differential Revision: https://phabricator.services.mozilla.com/D19831

UltraBlame original commit: 48d98388c29cfa5d0e4dd92acfb3327f160fc774