Bug 1303429 - expose placeholder object attribute for HTML placeholder, r=Jamie
The appropriate case is when placeholder is not used as the absolutely last fallback for the accessible name, and the input is not labeled by the same text as in the placeholder attribute itself. So a label and the placeholder text must be different for the object attribute to be exposed.

In addition, for ATK, placeholder is being renamed to placeholder-text to comply with the platform specification.

Differential Revision: https://phabricator.services.mozilla.com/D19808

UltraBlame original commit: fd01e6363762255fa6ff7ebde9b30e65ff7219e5