Bug 1507491 - Let WebAssembly.Table.prototype.grow take a fill argument. r=luke
There is no real spec for this feature yet but we assume for now that
grow with a fill value is like growing without a fill value and then
storing the fill value repeatedly.  The default value is null.
However, since the second argument to this method was previously
ignored, non-function values are ignored (rather than causing an
error) when the table is a table-of-anyfunc.

Since we're extending the semantics for grow() on tables of anyfunc I
have opted to gate the functionality on ENABLE_WASM_GC.

(Additional WPT tests land separately.)

UltraBlame original commit: 6f42a0b6c8dea3a07850db4bfe5966f56c5a8a58