Bug 1524982 - Use browser loader in performance panel to avoid blocking toolbox destroy;r=gregtatum,ochameau
The main goal is to use the browser loader instead of script tags to avoid using promises tied to the document during the panel's destroy.
I tried to cleanup a bit the modules, to rely less on globals and more on explicit require calls, but there is room for improvement left.
I will write an additional test to check that toolboxes in WINDOW hosts can be closed and reopened for any of our panels in a second changeset.

Differential Revision: https://phabricator.services.mozilla.com/D19865

UltraBlame original commit: cb15b108719acd80d9c59b773a1838b9d4d6c7f1