Bug 1528884 - Change how --enable-eme is handled. r=ted
It used to be that --enable-eme on its own had some effect, but that's
not the case anymore. The only thing it does as of bug 1300654 is to
switch the defaults in the Firefox configuration.

Reflect that in how we handle --enable-eme.

Differential Revision: https://phabricator.services.mozilla.com/D20270

UltraBlame original commit: 43965f6107dcdd7051afae588ebe0c3b12612ec8