Bug 1304001 - Stop passing the top window to LoginManagerContent. r=sfoster
It wasn't clear in callee code that the window was the top-window and it wasn't necessary in many cases. Relying on the top-window would also cause problems with Fission if the windows are in separate processes.

Differential Revision: https://phabricator.services.mozilla.com/D20395

UltraBlame original commit: d6646400f45d1a63ca75ac04e6b9b109c41970f4