Bug 1501375 - Clear primed listeners on extension shutdown r=aswan
In D9959 and D10954, the implementation of background page machinery
was updated so that the `build` method of `BackgroundPage` is
guaranteed to return eventually, without error. Even if the load of
the background page was interrupted by an extension shut down.

When an extension shuts down during start-up, primed listeners have
likely not been re-registered. To prevent interruption of startup
from causing the loss of persistent listeners, this commit changes
the implementation such that persistent listeners are only updated if
the background page was fully loaded.
Otherwise listeners are cleared from memory, but not unregistered.

Depends on D10954

Differential Revision: https://phabricator.services.mozilla.com/D17700

UltraBlame original commit: 3a5c6e81192dbdb0a8adfb20e00181d3de5540f7