Bug 1529352 - Collect WebRender probes in release r=jrmuizel
A handful of probes that are useful for measuring the performance of WebRender are not currently collected in the release channel.

Since we're planning to begin deploying WebRender in release on an experimental basis in 66, we should make sure that these probes are available there, so that we can assess its performance.

Differential Revision: https://phabricator.services.mozilla.com/D20599

UltraBlame original commit: c6e08ed3909cfd0227f444df63d0de4064b263e5