Bug 1530575 - Reduce SubChunkStep r=sfink
I initially set this to 64K thinking that it'd prevent too much changing of
the nursery size and therefore (once implemented) too many calls to
madvise.  But that's not true, because the nursery only changes size when a
tenure rate threshold is hit, and once it does it always changes size.  So
we can reduce this to 4K for simplicity and alignment with other constants.

Differential Revision: https://phabricator.services.mozilla.com/D21157

UltraBlame original commit: 705b9ae4d7280aac57cd326467387b77424aaab6