Bug 1530234 - P2. Don't recalculate audio sample start time. r=bryce
The WMF audio decoder recalculated the timestamp of each audio sample according to the number of frames decoded so far.
This is incompatible with the trimming mechanism that rely on the timestamps of the audio to be matching what is found in the container.

All the other audio decoders do it that way already.

Depends on D20969

Differential Revision: https://phabricator.services.mozilla.com/D21305

UltraBlame original commit: 4e5ac450f2b6dd57e9264fddc97971f6b5d8dc18