Bug 1527811 - On Windows on ARM64, add ClearKey GMP to the GMPService in i686 subdir. r=Gijs
We need to tell the GMPService where to find the ClearKey GMP, and in Windows
on ARM64 builds we should run the x86 ClearKey so that we test the same x86
emulation path that Widevine uses.

This patch assumes that the ClearKey GMP and its appropriate directory
structure are placed in the "i686" subdir of the aarch64 firefox dir by
the build.

Differential Revision: https://phabricator.services.mozilla.com/D19900

UltraBlame original commit: aae2bb67624e50f6628306cbd3a2099928161bca