Bug 1527811 - Hide OpenH264 by default on Windows on ARM. r=dminor,r=Gijs
We don't have an aarch64 build of OpenH264 yet, and WebRTC on aarch64 is
blocked by DirectShow support from MS, so just hide OpenH264 from the addons
manager UI for now.

We achieve this by moving the isEME() check in GMPUtils.isPluginHidden() down
to after the isPluginSupported()||isPluginVisible() check, so we can use
the media.gmp-gmpopenh264.visible pref to hide OpenH264 in the addons
manager.

When we are ready to enable WebRTC, we can flip the pref.

Differential Revision: https://phabricator.services.mozilla.com/D19903

UltraBlame original commit: 56d43815607853dacf7ae7d272912ef3e8927229