Bug 1530861 - Acknowledged newSource events sent by remote devices on FF66 and less. r=jdescottes
newSource events used to be sent on the target actors on FF66 and before,
this is no longer the case. But we still have to accept them if we connect
to old remotes, otherwise the events are considered as a method reply
and confuses packet ordering.

Differential Revision: https://phabricator.services.mozilla.com/D21394

UltraBlame original commit: 9a25f61c3ca39ba331fcf9810a6464fd14c78414