Bug 1530273 - Make 'funcref' the canonical name. r=bbouvier
Wabt is now supporting funcref exclusively, and with the reftypes
proposal I think there's broad agreement that we will stop using
anyfunc.  So let's accept funcref both in the text format and in the
table creation dictionary, and let's use this name as the canonical
name in error messages and similar.

But let's also continue to accept anyfunc, since there may be
in-flight tests and other content that uses it.  This includes a
couple of emscripten-compiled benchmarks currently in the repo; I
chose not to change those.

Differential Revision: https://phabricator.services.mozilla.com/D21388

UltraBlame original commit: 5c4daba349dc2484188abd2c301b35ccd2d4aa08