Bug 1532402 - Part 2: Add a variant of nsContentUtils::ShouldResistFingerprinting() which accepts an nsIPrincipal* argument; r=baku
We can use this variant in cases where we don't know anything about the
provenant source of the check except its principal.

The principal may not be enough information in the future but at least
porting call sites to this variant will allow us to have a small number
of functions to extend in the future when we want to enable enabling
fingerprinting resistance in certain frames (e.g. third-party tracking
frames.)

Depends on D21990

Differential Revision: https://phabricator.services.mozilla.com/D21991

UltraBlame original commit: ed1e303084c4026cdbba075464ef1050013df6b0