Bug 1531886 - [tryselect] Handle templates in mach_commands.py, r=gbrown
This was previously only in the cli parser because that was the only shared
place that ran for all selectors. Now that we have the 'self.run' function in
mach_commands.py, we can move it there. This move is also needed to allow us to
remove 'templates' from kwargs (which happens in the next commit).

Depends on D22021

Differential Revision: https://phabricator.services.mozilla.com/D22022

UltraBlame original commit: ba832d587ae117729d39b1040f1f87e2655b31ac