Bug 1527922: Ensure the Reload button is disabled when testing against blank tabs in the browser toolbar key nav tests. r=Gijs
For a blank tab, the Reload button should be disabled.
These tests depend on this.
This seems to be true when setting the new tab page to blank in Firefox Options.
However, when we open about:blank with BrowserTestUtils.withNewTab, this is unreliable.
That is, sometimes the Reload button is enabled, sometimes it isn't.
I don't understand why this happens.
For the purposes of these tests, just force the Reload button to be disabled for new, blank tabs so we get consistent results.

Differential Revision: https://phabricator.services.mozilla.com/D22449

UltraBlame original commit: c26d4a8d43c4d0d6f6e13135b8f35924123fe99e