Bug 1509446 - part3 : run 'TimeMarchesOn' directly, instead of queuing a task. r=jya
According to the spec [1], we should run `TimeMarchesOn` algorithm directly when the specific situations happen, the spec doesn't say that we need to queue a task for it.

In addition, all the call sites for `TimeMarchesOn` are on the main thread, so we don't need to worry about race condition.

[1]
https://html.spec.whatwg.org/multipage/media.html#playing-the-media-resource:time-marches-on-2
https://html.spec.whatwg.org/multipage/media.html#playing-the-media-resource:time-marches-on-3
https://html.spec.whatwg.org/multipage/media.html#playing-the-media-resource:time-marches-on-4

Differential Revision: https://phabricator.services.mozilla.com/D22149

UltraBlame original commit: d1046aea2e0e1bca30de7d44b08bf56915a5aefc