Bug 1530255 - Load keyMap files on demand; r=bgrins.
We use to include all the keyMap files in the
codeMirror bundle, but these files can be
pretty large, and are useless if the user did
not select a custom key mapping in the settings
panel.
This patch changes this to only load those files
when they are needed, either when a new editor
is created, or when the user changes the key mapping
preference. The files are only loaded once.

Differential Revision: https://phabricator.services.mozilla.com/D22977

UltraBlame original commit: 2f7c1341cc1e39358cf79de3dc476c0104ae8279