bug 1534829: remote: map source code to bug components; r=yulia
DevTools :: Remote Agent was removed in favour of a new product,
Remote Protocol.  This corrects the file metadata mapping for the
agent.

The patch also adds mapping for new domains.  Not all of the
components exist yet, but this is tracked separately in
https://bugzilla.mozilla.org/show_bug.cgi?id=1534970.

Differential Revision: https://phabricator.services.mozilla.com/D23306

UltraBlame original commit: a710a25b6c5d3274fb80cf22af7fcc50ce306256