Bug 1521729 - P2: Fix failure tests r=mayhemer
Due to the p1 patch, the first time we get the result back from NS_ShouldSecureUpgrade is always asynchronously.
This means that the first http request could be completed slower than the http requests opened after the first one. There are some tests affected by this change, since these tests assume that http requests should be completed in the same order as these requests were created.
This patch fixes the tests by introducing a new method in nsIHttpProtocolHandler. This method returns a promise and makes sure the HSTS data is ready to read synchronously when the promise is resolved.
Once the HSTS is ready to read, the order of open/close channels will be the same.

Differential Revision: https://phabricator.services.mozilla.com/D21683

UltraBlame original commit: 70b9b6dbe5b8e2b55a240aa3b39ff6087de92276