Bug 1488673 - part4 : adjust cue's position. r=heycam
According to the spec 7.2.10, step1 to step10 (snap-to-line is true) and step1 to step2 (snap-to-line is flase) [1], we would adjust cue box's position depending on the algorithm.

The algorithm would calculate the cue box's top (or left) position depending on the writing direction, and move the box to the specific position.

However, if the cue box is overlapping with other cue boxes in the rendering area, we would still need to adjust its postion, which will be implemented in next patch.

[1]
https://w3c.github.io/webvtt/#ref-for-webvtt-cue-snap-to-lines-flag-12
https://w3c.github.io/webvtt/#ref-for-webvtt-cue-snap-to-lines-flag-13

Differential Revision: https://phabricator.services.mozilla.com/D19609

UltraBlame original commit: 9e2f80f5ec79e1d33826065ef40fb799da3ad574