Bug 1533890: Add simulator support for IgnoresReturnValue version of native functions r=mgaudet
Native calls in the simulator have to be redirected to a special swi instruction. In the old implementation, this redirection did not support calling the IgnoresReturnValue version of a native function. This patch fixes that problem in the new implementation.

Differential Revision: https://phabricator.services.mozilla.com/D22775

UltraBlame original commit: 3fad807ff7d4da01b6c161ef4eac3c3a155d5fc4