Bug 1535609 - Root pointer args in callExport. r=bbouvier
In the slow callExport path, we convert arguments left-to-right and
these conversions can both OOM and error out for other reasons, and
they may perform various computations, notably allocation.  Therefore
any reference arguments must be rooted until all the conversions are
finished.

However, the arguments are computed into a plain Vector<> whose layout
we need to control tightly and whose elements are not easily rootable.

So we keep a separate rooted vector of reference values and when we
encounter a reference argument we push that argument onto this, and
before the actual call we move the reference values (which may have
been altered by GC) into the argument array; this copying is atomic
with respect to gc + the invocation, and once we're in the callee the
normal rooting logic (via stackmaps) takes over.

Differential Revision: https://phabricator.services.mozilla.com/D24061

UltraBlame original commit: 009a7361fc1b4dd6852300e52cfc5b541d9f58f0