Bug 1530481 - Support encoded snapshots in Firefox Origin Telemetry. r=janerik
The threading and promise handling is kept at the TelemetryImpl level
deliberately to facilitate testing. It also turns out that hangs/modules do
more-or-less the same split for their off-thread collection, so in
retrospect it is the obvious choice.

For the batch ID we're using the app build id as discussed.

Differential Revision: https://phabricator.services.mozilla.com/D22606

UltraBlame original commit: 91a4fa9c15a838309e4abf7250f4e7267a603eb6