Bug 1532395 - Clear the DNS cache before each TRR test r=JuniorHsu
It is likely that the test is flaky because we resolve the same host with a different TRR endpoint and sometimes we use a cached record.
We should clear the DNS cache before each test to ensure that we're actually checking the right thing.

Differential Revision: https://phabricator.services.mozilla.com/D24355

UltraBlame original commit: 127610bec3fc4d268f0dfda04ffa5028523ae9ef