Bug 1538470 - Null-check the global when creating AudioNodes. r=karlt
This is now necessary, because it's possible to effectively create new
AudioNodes on during shutdown, when the global is not available anymore.

It is unnecessary to null-check in the AudioNodeEngine methods, because they
will not run in this scenario.

This also renames the function because it can now return nullptr.

Differential Revision: https://phabricator.services.mozilla.com/D24688

UltraBlame original commit: 58fe84489519e891e049ad1eccee5f7ad046c1ba