Bug 1538508 - P2. Don't limit search for SPS/PPS on keyframe only. r=bryce
We limited the search for a SPS/PPS change to keyframe only in bug 1469257.
However this causes issues if the first frame containing a SPS/PPS isn't a keyframe.

We also need to error on content with no SPS/PPS as to inform the caller that something is amiss. Such content was invalid to start with.

Differential Revision: https://phabricator.services.mozilla.com/D24853

UltraBlame original commit: ea116addae50b8e65e75f4c875b9e583c1c74f34