Bug 1352113 - Shippable Builds - Make source-test depend on shippable instead of pgo. r=aki
This is useful in order to not have to run linux64/opt on push, especially on autoland/inbound
when we need a source test. It is also required if we remove the linux64-pgo build type entirely.

Differential Revision: https://phabricator.services.mozilla.com/D23124

UltraBlame original commit: 9cf5c1259a858ee5e49e671e3f7d7f9aedf7992c