Bug 1533808: Recognize Chromium-sandboxed methods in WindowsDllInterceptor r=aklotz
Due to coming changes involving the IOInterposer, the WindowsDllInterceptor may be set up later than the sandbox.  The sandbox hooks some of the same functions, so the Interceptor is running into its hooks instead of the original implementations it anticipated.  This patch allows it to recognize and efficiently patch those hooks when that happens.

Differential Revision: https://phabricator.services.mozilla.com/D24654

UltraBlame original commit: 88df194b6f92e1cb6623398323f2e593d58bfddc