Bug 1525762: Part 1e - Don't hide add-ons in the built-in location. r=aswan
We want add-ons in the built-in location, particularly the built-in themes, to
appear in the add-on manager so that users can enable and disable them. We
don't want them to be able to uninstall them, though, so this also prevents
users from uninstalling them.

Differential Revision: https://phabricator.services.mozilla.com/D24619

UltraBlame original commit: cc45935ed2c2f3ae4627584f5c7318cb52271875