Bug 1539262 - Support multiple origins in the same metric in Origin Telemetry r=janerik
Since reporting intervals are ~1 day/1 session, the Origin Telemetry prototype
must support the possibility that multiple origins will be recorded for the
same metric.

For example, if the user is sampled to record two pageloads where the same
ultra-common tracker is present and blocked we must record that tracker as
having been blocked twice.

This requires a bit of a shift in storage and plaintext snapshot. Instead of
being an array of origins with duplicates, now we're storing origins as a bag
(aka multiset, aka hashtable of origin->count).

Differential Revision: https://phabricator.services.mozilla.com/D25283

UltraBlame original commit: 7c940e9caee902fd9da00798ce2fee013aa82591