Bug 1520929 - part 1 - add an isWowARM64 property to nsSystemInfo; r=aklotz
In addition to knowing whether we're running x86-on-x86-64, we'd also
like to know about the x86-on-arm64 case.  The current code doesn't
provide enough information to determine that, so we need to query a
little bit harder.

UltraBlame original commit: d49485b4f4221077442dc994c6e69b11f32ddc64