Bug 1541121: [try-staging] Allow staging release to not have partials; r=aki
We check that partials as a safety check. But we don't need that for staging builds,
and it is often useful to be able to test things that don't depend on partials.

The shipit UI currently still requires partials, but that can be worked around using
the react dev tools.

Differential Revision: https://phabricator.services.mozilla.com/D24834

UltraBlame original commit: 948b7b1003afaea2c3f03bde1d43adcf0cd108ba