Bug 1539210 - Do not destroy the content process session on page unload. r=ato
A session is bound to a browsing context, not just one document.
`unload` is fired when navigating to another document. The document goes
away, but not the browsing context. So, we should keep the session up and running
when the current document is destroyed and we navigate to another URL.

Depends on D25577

Differential Revision: https://phabricator.services.mozilla.com/D25578

UltraBlame original commit: d0558202539c2f1f0977b332b8ea65472039a77c