Bug 1541413 - Fix Error rep parseStackString function. r=davidwalsh.
If the stacktrace wasn't similar to SpiderMonkey structure,
we could end up calling the match function on a null variable,
which would throw (and ultimately display the safenet Invalid Object
element).
The fix is simple, we guard on the variable before calling match
on it.
Test cases are added to ensure we can deal with different kind
of stacktraces.

Differential Revision: https://phabricator.services.mozilla.com/D26125

UltraBlame original commit: 1365a123d8e23af9fccbd656ead27f37c5796be2