Bug 1541925 - Avoid accounting for font-face rules the style set when there's no pres shell. r=heycam
This preserves the behavior of not having FontFace objects for font-face rules
in display: none iframes.

Note, however that we may want to always track stylesheets to both not have this
limitation and to fix bug 1471231 & co. But that's a separate bug in any case.

The assertion that is failing ultimately in this bug is not really relevant
anymore, so remove it while at it.

Differential Revision: https://phabricator.services.mozilla.com/D26196

UltraBlame original commit: 1b9a86e36a4c161794d82f6cadc3a6ccb11dadf4