Bug 1508961 - Show a notification icon after automatically dismissing permission prompts. r=MattN
This adds support for a "post-prompt" that appears dismissed (only the icon showing) when
the browser decides to automatically deny a permission request from a site (either based on
internal heuristics or because the user prefers not to receive permission prompts).

In order to move quickly and not overload the patch I have only implemented the post-prompting
for desktop-notification so far, however, implementing it for other permissions should be
a trivial follow-up.

Differential Revision: https://phabricator.services.mozilla.com/D25415

UltraBlame original commit: b73c90f76abd7c0fc7d287135a8d0ab9cd13d7df