Bug 1542854 - Make [System Principal] into a static atom. r=froydnj
Whenever chrome JS code leaks, we leak the dynamic atom "[System
Principal]". This causes an assertion, which hangs the browser in
stack walking code for multiple minutes on OSX. As a bandaid, make
this atom static so it won't leak.

My first instinct is that it would make sense to use this new atom
directly in SystemPrincipal, but atoms are represented internally as a
kind of nsAString, but principal code uses nsACStrings, so you end up
with an extra conversion from 16-bit to 8-bit.

This atom is also going to be present in any process running chrome
JS, so this will save a little bit of memory.

Differential Revision: https://phabricator.services.mozilla.com/D26590

UltraBlame original commit: f09e9d011ba193397c2f0035ab74b3120355b99d