Bug 1541819 - Ship allocation-tracker when running tests and test it. r=jdescottes
Using official build flag was breaking tests on try as the builds
are done with the official build flag turned on.
Convert this module into a "supports" file, like other test-only files.
And also introduce a test for this module. It helps as we have to have
a browser.ini file in order to reference allocation-tracker.js,
and you can't have a browser.ini without any test in it.

Depends on D26105

Differential Revision: https://phabricator.services.mozilla.com/D26106

UltraBlame original commit: 251858ceaf0caec574c4ba65fe70aacdd014f22f