Bug 1533968 - Don't add animations to layer for TYPE_TABLE_BACKRGOUND_COLOR. r=hiro
nsDisplayTableBackgroundColor inherits from nsDisplayBackgroundColor,
and both display item types are mapped to background_color property.
However, nsDisplayTableBackgroundColor is not animated on the
compositor, so we shouldn't add animations for TYPE_TABLE_BACKRGOUND_COLOR.

It seems the test case has a weird timeout in Android test-verify-opt. I
checked the live.log and the tests are passed, but got a final time-out.
It is not related to our change because this still happens if we skip
the test, and I saw that other intermittent bugs have the same issue.

Differential Revision: https://phabricator.services.mozilla.com/D24780

UltraBlame original commit: 6f0b0ab56a478e99853a465ceaa7a5d8c7e2b1b8