Bug 1529772 - Part 2: Factor out PropertyName parsing from Parser::propertyName(). r=khyperia
The existing propertyName() method parses a little more than just
PropertyNames; this patch renames it, to clarify that, but the behavior is
unchanged.

There was a lot of redundancy in the implementation of getters and setters, so
this deletes a bunch of code.

Differential Revision: https://phabricator.services.mozilla.com/D26036

UltraBlame original commit: 067e3a99b02750fa89a16a451ca1cfbf42354b71