Bug 1529772 - Part 2: Factor out PropertyName parsing from Parser::propertyName(). r=khyperia
The existing propertyName() method parses a little more than just
PropertyNames; this patch renames it, to clarify that, but the behavior is
unchanged.

There was a lot of redundancy in the implementation of getters and setters, so
this deletes a bunch of code.

Differential Revision: https://phabricator.services.mozilla.com/D26036

UltraBlame original commit: 0f4a343adc59b26a574c1a74b87c4990ffe77057