Bug 1543560 - Move NotifyWebRenderError() to CompositorManagerParent r=nical
NotifyWebRenderError is sent via CompositorBridgeParent/CompositorBridgeChild. It is not good, since they are re-created for each nsBaseWidget::CreateCompositorSession() call. Then there could be a case that the NotifyWebRenderError is not notified to GPUProcessManager and a crash was caused by NotifyWebRenderError message during ipc close.  CompositorManagerParent/CompositorManagerChild are stable for sending NotifyWebRenderError.

Differential Revision: https://phabricator.services.mozilla.com/D27030

UltraBlame original commit: 719725292e4cafffb19ff16688879875d4053b0e