Bug 1544602 - Make media.gmp.insecure.allow an atomic. r=cpearce
Since this pref can be accessed from the GMP thread (per the bug), it should
be an atomic.

See bug 1465852 for the code the added the assertion hit in this bug, as well as
further context as to why this should be an atomic.

Differential Revision: https://phabricator.services.mozilla.com/D27831

UltraBlame original commit: a14a4e0b60ebc2ca635bbc777474cf2fa8cbc62d