Bug 1534738 - Title and url in RTL mode are not properly overflowing in Quantum Bar. r=dao
Moves the mask to the url and title elements, uses a flex layout to ensure
elements properly overflow, and fixes alignments and paddings because the new
layout would be more compact. rows are slightly taller than before, but they were
smaller than the legacy bar, so in the end we should be good.

Differential Revision: https://phabricator.services.mozilla.com/D27868

UltraBlame original commit: 841e248130aeed83a650c3a2049310946279efa8