Bug 1541767 - Make Animation::Cancel line up with the spec a little better; r=hiro
Animation::Cancel calls UpdateTiming() which in turns runs the procedure to
update the finished state. However, the spec[1] doesn't require that.
Furthermore, calling UpdateTiming here hides the fact that we end up triggering
a restyle.

It would be better to move the parts of UpdateTiming we require into Cancel
itself so that we align better with the spec and to make it a bit more clear
what side-effects of UpdateTiming we actually rely on.

[1] https://drafts.csswg.org/web-animations-1/#cancel-an-animation

Differential Revision: https://phabricator.services.mozilla.com/D28020

UltraBlame original commit: a751128e51f71ddd7df3eaa1ba8c5c34b6876e31