Bug 1545207 - Mock the Search Service region test in Telemetry Client Tests r=raphael
We don't want spurious net connections, so we should mock the result.
In addition, we need to mock it to a specific non-North-American region to
ensure we don't cause an extra change to browser.search.region which may split
the subsession in the middle of our tests.

Differential Revision: https://phabricator.services.mozilla.com/D28059

UltraBlame original commit: 2f36aca8e78bd627f3a006b2582819489a675ad0