Bug 1544198 - Use the proper writing-mode for scroll-snap for viewport. r=jfkthame
In the CSS writing mode spec [1], the writing mode for the document should be
taken from body, GetFrameForDir() is the function to get the corresponding frame
for writing-mode.

A web platform test for this case will be added at the last of this commit
series.  Unfortunately as of this commit, we can't introduce proper test cases
since there is another issue on scroll-snap-type which will be fixed in
subsequent commits.

[1] https://drafts.csswg.org/css-writing-modes-4/#principal-flow

Differential Revision: https://phabricator.services.mozilla.com/D27984

UltraBlame original commit: f413cb36905d980cccb30577f827938f80cbe913