Bug 1544749 - Fix HTML Tooltip tests when running in content frame r=bgrins
Depends on D27680

Not strictly related to win.top/parent, but those tests would fail if the toolbox is in a content frame.
tooltip-01.js is creating a blank tab for no reason which prevents interacting with the actual test frame.
tooltip-02.js is not properly targeting an iframe to simulate a click

Differential Revision: https://phabricator.services.mozilla.com/D27674

UltraBlame original commit: 970ef36e6c09e748140ebf96c9ccfec6eb52ab5e