Bug 1538054 - Toggle off 'Pause on Exception' immediately. r=loganfsmyth
After toggling off 'Pause on exceptions' in the debugger, exceptions were still being paused on once.

This was because the server's exception-pausing behavior was not toggled off until an exception was paused on. This patch makes it so that the behavior is changed immediately after unchecking the 'Pause on exceptions' checkbox in the debugger.

Differential Revision: https://phabricator.services.mozilla.com/D27453

UltraBlame original commit: c31ccea52937e0c889f392b8d52094ad367c5f60