Bug 1542826 - Rebuild the debugger front-end. r=gw
This patch is the output of rebuilding the debugger frontend with a
recent npm installation. The changes to dist/ and the addition of
package-lock.json are a result of the build. The changes to package.json
and main.js were done manually to work around an incompatible change
in beufy 0.6.7 (the lib/ folder changed to dist/).

Differential Revision: https://phabricator.services.mozilla.com/D28357

UltraBlame original commit: 1dba743761b70dcf224def15834f5d32559ea72e