Bug 1545108 - Use a diagnostic assert to check for nullptr when pushing into MediaQueue. r=alwu
We're rarely getting nullptrs out of MediaQueues. It's not clear where these are
coming from, as we have many guards against them. Upgrade this assert to a
diagnostic to help us track the source and determine if the value is null before
entering the queue.

Differential Revision: https://phabricator.services.mozilla.com/D28866

UltraBlame original commit: b9bd90dd38bfcc9ea7231d0dd326fced00750749