Bug 1440537 - Fix a cross-origin navigation test to navigate reliably. r=bzbarsky
This subtest (of test_iframe_sandbox_navigation.html) starts intermittently
failing with my first patch of this bug. It relied on the pres-context being
created when sendMouseEvent is called in order to navigate away (we only
navigate away by clicking a link if there's a link handler).

sendMouseEvent calls getBoundingClientRect() which used to do this. It no longer
does though.

I could make sendMouseEvent do that automatically using SpecialPowers or such,
or from the DOMWindowUtils code, but I think I'd prefer not to do that. This is
the only test that wasn't trivially fixable, and this awkwardness can be removed
when bug 1218456 is fixed.

Differential Revision: https://phabricator.services.mozilla.com/D28910

UltraBlame original commit: 5b63c5b38dcc382a4fa0afdbbf67b147057c4fac