Bug 1544686 - [raptor] Fix Gecko profiler settings. r=rwood
When the Gecko profiler is enabled the interval and entries as specified
in manifest files aren't correctly set. Instead the default values from
the extension are always used.

Also with WebRender enabled the threads are "WR", and "Renderer", but
not "WR,Renderer".

Differential Revision: https://phabricator.services.mozilla.com/D28045

UltraBlame original commit: 9ff6837b1526aaae3eca6546245ace6710c20a40