Bug 1533211 - Remove assertion for if MP4 sample description index is invalid. r=jya
It's possible for a malformed mp4 to contain invalid sample description index in
fragments, that do not reference any sample description entries found in the
header. E.g. the header may contain 2 sample description entries (which should
be indexed with indices 1 and 2), but for a fragment to contain an index to 4.
Instead of asserting in this case we should gracefully fail.

Bug 1547328 plans to add logging for this case, so we have a means to still
detect failures here from bad files.

Depends on D29733

Differential Revision: https://phabricator.services.mozilla.com/D29734

UltraBlame original commit: 9de7d7669f41a5df9c61b753767a5713aa9b965d