Bug 1543607 - Store the tabbrowser scrollbox button width when tabs are closed with the mouse so it can be used on underflow to keep the close button under the mouse r=dao
The button is already hidden when underflow fires, so the clientWidth is 0. Instead,
store it during _lockTabSizing so we know how much space to fill when tabs are being closed
by the mouse, to allow the close button to remain underneath the mouse cursor.

Differential Revision: https://phabricator.services.mozilla.com/D29227

UltraBlame original commit: 2f0db0b04a913275f91ede9fc984a40065c08266