Bug 1540200 - Part 3. Reuse snapping offsets calculated on the CPU in the WebRender shaders. r=kvark
Historically we calculated the snapping offsets in the GPU shaders.
Because this information is always needed on the CPU side, we now just
pass the values into the shader instead of recalculating again. This
ensures we will use the same set of values consistently and makes it
easier to adjust how we snap in the future.

This patch should have no functional change on the output of WebRender
itself.

Differential Revision: https://phabricator.services.mozilla.com/D28883

UltraBlame original commit: 34510ca46cc61984ca775f30a7c4f288371cf329