Bug 1546248 - Loosen Client-ID check in unit test r=rpl
The client_id part of browser_html_discover_view_clientid.js was failing
on TV because of a pre-existing, test-specific issue in Telemetry.
Fixing this is not trivial, so just check that the ID was set instead of
checking its exact value.

See the comment for more details; the fix will be part of bug 1537933

Differential Revision: https://phabricator.services.mozilla.com/D30001

UltraBlame original commit: 912b44cef0bcf2ddf29778ee2f5d8586933095af