Bug 1549388 - Render non-string messageText with GripMessageBody. r=bgrins.
Sometimes the messageText can be a longString, which was crashing
the whole console output since we were always assuming messageText
was a plain string.
We now render non-string messageText with reps so it handles longStrings.
A mocha test is added to ensure this works as expected.

Differential Revision: https://phabricator.services.mozilla.com/D30875

UltraBlame original commit: b90c24bb6358c8ad4b0b56cd312457e0548c26f3