Bug 1550737 - When shutdown is finished use only a reference of the child. r=jya
When `RemoteMediaDataDecoder::Shutdown` is finished it is not necessary to hold a reference of the `self` any more. Keep the `mChild` alive, which is the only one needed to destroy the IPDL. In addition to that, deleting the IPDL and destroying the child will be happening at the task queue similar to what was happening before Bug 1545416.

Differential Revision: https://phabricator.services.mozilla.com/D31261

UltraBlame original commit: 5ba6d20ec80ad5dcabead733c7bd77df18aa05d7