Bug 1552089 - Don't tweak snapport position even in the case of RTL scroll containers. r=botond
In RTL scroll containers, the right most x-axis scroll position is 0 and
leftward scroll positions are negative values.  And also
nsLayoutUtils::TransformFrameRectToAncestor, which is used to tell whether the
snap target element is inside the destination snapport or not [1], returns
negative x-axis positions for elements in RTL scroll containers if the element
is positioned at places where the elements are outside of the initial scroll
position (0, 0).  So we don't need to tweak snapport postion even in the case
of RTL scroll containers.

Instead, what we needed in the first place is that we choose a proper x-axis
scroll position that the targe element appears inside the snapport.

[1] https://searchfox.org/mozilla-central/rev/11cfa0462a6b5d8c5e2111b8cfddcf78098f0141/layout/generic/nsGfxScrollFrame.cpp#6604-6605,6616-6617

Depends on D31409

Differential Revision: https://phabricator.services.mozilla.com/D31410

UltraBlame original commit: 14743da36853e6b1d745418e128030b89bbb6fe2