Bug 1551086 - Do not finish edit action when unbinding the frame from the editing state. r=masayuki
The code here was introduced to fix bug 1053048, but the test there and the
test-case there no longer need this, and it's generally unsound to do stuff that
changes the state of the input from UnbindFromFrame().

I'll file a new bug for the bogus disabled styling in the test-case in a second.

I don't know how to add a test for this (no less because the testcase crashes on
debug builds regardless of this patch, see bug 1551192).

Differential Revision: https://phabricator.services.mozilla.com/D30914

UltraBlame original commit: f6b2612701f631484e59c793308e955b01eada99