Bug 1533277 - Use SearchBox component in console filter bar. r=ntim.
The code is adapted to work with the SearchBox, which means
we can get rid of an event listener in webconsole-ui.js.
Also, we take this as an opportunity to remove the element
refs as they were only used in tests, where we can more easily
use query selectors to get those elements.
Some changes were needed in SearchBox's `onChange` function since
`setState` is asynchronous.

Differential Revision: https://phabricator.services.mozilla.com/D31762

UltraBlame original commit: 4039c6761d62239b4e55691f381cf2953969dcca