Bug 1550606 - Fix a race condition in WPT related to postMessage r=smaug
We used to use 2 rAFs to stop the load event from racing with
onmessage event, but this was not the proper solution and it was
still racy, especially after Bug 1534012 is landed. So we fixed it
by instead letting racy postMessage to be sent, we wait for the
proper postMessage to be sent, and the test would automatically
timeout if the proper postMessage failed to sent.

Differential Revision: https://phabricator.services.mozilla.com/D32029

UltraBlame original commit: 28a981abcc6c2b80648cc8b4162b87b28fddd4ef