Bug 1548612 Part 2 - Fix GetTextWidthHeight for strings with no line breaks. r=agashlin
The current implementation of GetTextWidthHeight attempts to guess how much
height a string needs to fit into a given width based on how long the string
is when rendered onto one line of unlimited width. This doesn't work because
breaking up the string into lines introduces additional space at the end of
the lines that the single-line method doesn't account for.
This patch replaces all of that logic with asking DrawText to render the
string into the width of interest and then just seeing how much height it
ended up needing in order to do that.

We also take the opportunity to clarify what GetDlgItemBottomDU was doing,
because it isn't exactly what it claimed to be doing.

Depends on D31139

Differential Revision: https://phabricator.services.mozilla.com/D31140

UltraBlame original commit: 27fc874734e1f91a1a9e2835fd1c77a609cc9aa4