Bug 1536997 - Fix broken assert for recycling animated images with WebRender. r=kats
By default, recycling frames for animated images is enabled. However the
first frame is never recycled because we may need the first frame
immediately if the animation is reset. If only the first frame of an
animation is displayed before shutting down a tab, then it will not
correctly register with the layer state manager prior to its
destruction. This trips an assert incorrectly.

Now we just always register with the layer state manager if recycling is
enabled, and never if recycling is disabled. This allows us to remove
from state information at the cost of requiring a restart to toggle to
recycling pref (which is almost never done now that the feature is
stable.)

Differential Revision: https://phabricator.services.mozilla.com/D32526

UltraBlame original commit: 29c76bc4b5901f6fd331435e6db5ff47bcde04bb