Bug 1554496 - Don't remove the IS_IN_SHADOW_TREE flag from CharacterData unlinking. r=edgar
We used to do that for FragmentOrElement too, but it was removed in D3533. The
explanation is a bit hidden in the revision, but quoting:

> Needed to remove the IN_SHADOW_TREE flag removal from unlinking since it may
> make a ShadowRoot claim not to be a ShadowRoot (nsINode::IsShadowRoot() would
> return false).
>
> In practice it was there because there was code that assumed that
> IsInShadowTree() -> GetContainingShadow() != nullptr
>
> But that stopped holding a while ago, and callers have been adjusted in the
> meantime to not assume that, so it should be fine to not remove the flag now.

We also unbind the Shadow tree on unlink now, which didn't use to happen.

This makes CharacterData consistent with FragmentOrElement.

Differential Revision: https://phabricator.services.mozilla.com/D32627

UltraBlame original commit: 1173ab4276f6c62a3015a923d41d0cc4f79ede5e