Bug 1554317 - String for showing add-on recommendations r=rpl,flod
This string will be used as the alt and title text for the recommended icon
in about:addons. The patch to add the image will follow, once we have a way
of identifying add-ons that were recommended.

Differential Revision: https://phabricator.services.mozilla.com/D32542

UltraBlame original commit: a58efeaf9c94e23e12cc883f313fe70dd35dcdea