Bug 1553692 - enable wpt 'track-text-track-cue-list.html' r=jib
For getter, it will return undefined [1] when we can't get the element. We should use assert_equals because undefined is not an object.

For getCueById, it will return null [2] when we can't get the element. We should use assert_equals because null is not an object.

[1] https://heycam.github.io/webidl/#dfn-attribute-getter
[2] https://html.spec.whatwg.org/multipage/media.html#dom-texttrackcuelist-getcuebyid

Differential Revision: https://phabricator.services.mozilla.com/D32246

UltraBlame original commit: 15f3eb25315331a08c0cd739d69d1ce94ca36e90