Bug 1555102 - Cranelift: Generate better code for indirect-global accesses. r=lth
Wasm globals that are mutable and either imported or exported are, in our
implementation, accessed through a level of indirection. The instance's global
area doesn't hold the value itself. Instead it holds an immutable pointer to
the location where the value is really stored.

CL doesn't know that the pointer is immutable, and so fails to CSE/GVN
together multiple reads of the pointer, even when it's obvious it would be
safe to do so.  This patch marks it as `readonly` so that such commoning
up can happen.

Differential Revision: https://phabricator.services.mozilla.com/D33007

UltraBlame original commit: 7d78bc0694c6c30b0250ab6bc05f92a7da924dec