Bug 1551563 - Move shortcut test to own file r=mstriemer,mixedpuppy
In bug 1522757, a new test was added to verify that extensions without
shortcuts were hidden in the shortcut view, and shown in a list of
extensions without shortcuts.

The test was inserted in an existing test that did not have anything to
do with the change. In the next patch, I am going to add similar tests,
so the test logic was moved to a separate file.

Differential Revision: https://phabricator.services.mozilla.com/D31794

UltraBlame original commit: 6e43b0f4322b717606c2b14bd55d9c2958a1f7c7