Bug 1547752 - Support jump tables in cranelift. r=bbouvier
This patch uses the machinery submitted in
https://github.com/CraneStation/cranelift/pull/774 to split the jump
table (and any other read-only data) apart from the compiled code, so
as to make room for the epilogue and proper alignment.  The only
difficult part about this is to update the instructions referencing
the tables, and the table entries, when the tables are moved.

Differential Revision: https://phabricator.services.mozilla.com/D32458

UltraBlame original commit: 28264a198767c9081a59c119d240297890785de7