Bug 1464032 Part 2: Remove LoadEvent and replace with DoWithEvent. r=jrmuizel
This patch modifies DoWithEvent so that we can more easily have a
DoWithEventFromStream callable from outside of Moz2D similar to
LoadEventFromStream. We will add that in a later patch for the new
EventRingBuffer. It also changes the only user of LoadEventFromStream
over to it, so we can can get rid of it and LoadEvent entirely.

UltraBlame original commit: 274a9d9596a2b7a2d046e198ebcddaa4c45f82e0