Bug 1556511 part 1: Mark the outer SVG frame as needing a reflow, when its viewBox attribute changes. r=longsonr
The viewBox attribute establishes an aspect ratio, which may influence the size
of the outer SVG element. So when the viewBox attribute changes, we have to
reflow to potentially update the frame size.  This patch achieves this by
sharing an existing codepath for handling changes to width & height.

Differential Revision: https://phabricator.services.mozilla.com/D33898

UltraBlame original commit: aba9b5d525af1b48929de552224626799931ff3e