Bug 1555632: Add separate AddonManager.beforeShutdown barrier and use where appropriate. r=aswan
The AddonManager.shutdown barrier blocks the final phase of AddonManager
shutdown, which means that things which use it to mean "I need to do this
before the add-on manager shuts down" are generally broken.

This patch adds a separate beforeShutdown barrier which blocks the start of
provider shutdown, and changes existing broken consumers to use it. It also
renames the existing barrier to `AddonManagerPrivate.finalShutdown` to make it
clearer a) what it does, and b) that only internal AddonManager code should
use it.

Differential Revision: https://phabricator.services.mozilla.com/D33195

UltraBlame original commit: 2983751114de7a26f6d509b7369c129baee5ac69