Bug 1551490 - Fix browser_ext_browserAction_contextMenu.js when HTML about:addons is enabled r=rpl
When HTML `about:addons` is enabled, the "Manage extension" button still
opens the add-on's details page, but without automatically loading the
inline options.

The test assumed that it did and failed. This patch switches to a
different way of detecting that the menu item works as expected.

Differential Revision: https://phabricator.services.mozilla.com/D34209

UltraBlame original commit: 0ed314d2412307695a3835968c35c41450c3da49