Bug 1554075 - Allow direct access of source buffer's data from demuxer, eliminating most copies. r=jya
A lot of the overhead in MoofParser::RebuildFragmentedIndex(BoxContext&) is
copying data into the BoxReader's storage. In the MSE case the underlying data
being read is actually buffered in memory, so we may be able to just read the
in-memory data directly, avoiding the copy.

Note that the data may not be stored in a single contiguous ResourceItem in the
SourceBufferResource's ResourQueue. If so, it's not as straightforward to hand
out a uint8_t* to the underlying data. So we just copy it in that case for
simplicity. In most cases, the data in a single MP4 box would be appended in a
single append, and so is likely to be in a single ResourceItem.

Differential Revision: https://phabricator.services.mozilla.com/D33877

UltraBlame original commit: d5543a60f8335228f68cbaf760da511cf849b969