Bug 1494796 - fix blob stylesheet test to not fail retrieving sourcemap; r=nchevobbe
this test was failing because the threadClient was posting messages due to being unable to
find the source map. After investigating, I found that the sourcemap was formatted incorrectly. I do
not know if this was intentional. If it wasn't this fix works. If it was, then we need to find a way
to wait on source map failures, and the test name might need to be updated

Differential Revision: https://phabricator.services.mozilla.com/D32715

UltraBlame original commit: 986402a59722ad800f07eae5384d95dbf37c2c17