Bug 1557882 - part1 : delay setting 'unicode-bidi' until finishing the position calculation. r=heycam
Because of bug1558431, now the text alignment for `bidi` won't take effect on the initial reflow of the document, so we have to delay setting 'unicode-bidi' in order to trigger reflow again as a workaround.

It would make the `bidi` text showing correctly after reflow.

Differential Revision: https://phabricator.services.mozilla.com/D34640

UltraBlame original commit: 50efff6cae914849fff8ee9fbf6dc2d153210839