Bug 1553127 - Display a warning when `place-self` is used on non flex/grid items. r=pbro.
We take this as an opportunity to add tests for `align-self` as well.
This requires the test to change a bit so we can create more than one
element in order to test the inactive property helper on grid/flex item
(i.e. with a parent flex/grid container).
This is done by providing a `createTestElement` function in the test case,
that  creates whatever nodes it need and append it in the rootNode parameter.
The function then returns the element that needs to be tested with isPropertyUsed.

Differential Revision: https://phabricator.services.mozilla.com/D34714

UltraBlame original commit: 9ce8678568c46e262c285074ca7dd13a3b006081