Bug 1548687 - Send OOP frame transforms to the main thread more often. r=botond
In cases such as APZ scrolling with paint-skipping, there is an
additional transform component that needs to be included in the
transforms that APZ sends to the main thread for the purposes of
tracking OOP iframe positions. This patch includes the missing transform
component in the calculated matrices. This component updates more
frequently, generally on each call to RequestContentRepaint, as
mExpectedGeckoMetrics is updated, so we also need to send the frame
transform matrices over with this frequency.

Differential Revision: https://phabricator.services.mozilla.com/D35201

UltraBlame original commit: 6d41cf61f1f89ae0752669384b8c2b3b794431af