Bug 1548687 - Optimization to reduce the number of IPC messages. r=botond
The case being fixed by this bug is currently relatively rare, in that
it requires the scrolling of a frame that contains OOP content. This
patch adds a bit of optimization to ensure that we only send the
affected transforms (i.e. for the scrolled OOP layers subtrees) in this
scenario, so that we don't cause unnecessary IPC overhead.

Differential Revision: https://phabricator.services.mozilla.com/D35202

UltraBlame original commit: 1171aa4d8d4330a9568271b03e478bb000562b5b