Bug 1479960 - Fix the main thread I/O tests to handle the IPC shared memory changes. r=florian
The tests for unexpected main thread I/O had exemptions for the specific
paths that were being used for shared memory, which would cause it to
fail with the changes in this bug.  This patch does two things:

1. On Linux, /dev/shm is always tmpfs (a memory filesystem), so it's not
going to cause disk I/O, and it's used by glibc to implement the POSIX
standard shm_open API.  This allows all /dev/shm paths instead of
limiting it to a hard-coded prefix.

2. On MacOS, with the patches in this bug, we'll no longer use temporary
files for shared memory on current OS versions, but we still need them on
older versions to avoid an OS bug (https://crbug.com/project-zero/1671),
and they are backed by disk in this case, so we want to allow only the
IPC files.  However, the path prefix has changed.

Differential Revision: https://phabricator.services.mozilla.com/D34628

UltraBlame original commit: c418095bb7b78aa89668291e59e81926f877c4b5