Bug 1551614 - Simpler way of restarting our app in a new process; r=VladBaicu
Our Restarter Service would have issues on Android Q:
It would run on a separate process that allowed it to kill the app's process
but when trying to restart the app it would fail because it would've been a
background Service and as such it would've been restricted from starting other
Android Application components.

To make it work on Android Q we would need to migrate it to a foreground
service or piggy-back the SYSTEM_ALERT_WINDOW permission given by the user for
the Tab Queue functionality.

The chosen solution is simpler and not affected by any of those limitations.

Depends on D35402

Differential Revision: https://phabricator.services.mozilla.com/D35403

UltraBlame original commit: 0e1180cd2a89aa974ae7190839631249727cfca4