Bug 1557771 - Trigger fewer GCs via nsJSContext::MaybeRunNextCollectorSlice r=smaug
This patch changes MaybeRunNextCollectorSlice so that it doesn't start a new GC/CC if one is not running already.

This reduces the number of GC slices when running the benchmark from ~400 to ~230.

Olli, what do you think?  It seems to me that this is closer to the original intention of the code.  I haven't run this through perfherder because I haven't worked out what try command I need to do this, but I figure that it won't make things worse.

Differential Revision: https://phabricator.services.mozilla.com/D35556

UltraBlame original commit: ec4aac5a6c9f6b0172894a44bed7ad3393bfbae0