Bug 1560969 - Fix MediaStream-idl.https.html per spec. r=jib
The "inactive" event for MediaStream does not exist in the spec. Instead we can
check that the `active` attribute is set to false synchronously after stopping
tracks. For completion, we do this on all MediaStreams in the test.

Differential Revision: https://phabricator.services.mozilla.com/D35664

UltraBlame original commit: feb980ef7f050e76b921be9016d8390d56f35767