Bug 1557343: Move AutoRealm outside addGeneratorFrame. r=jorendorff
Without this patch, addGeneratorFrame may be called from any realm, and enters
the debugger's realm to call DebuggerFrame::setGenerator. However, we would like
to merge addGeneratorFrame and setGenerator, and call the combined from various
points which are already in the debugger's realm, so it would be a little nicer
to simply make the function assume it is called from the debugger's realm.

Differential Revision: https://phabricator.services.mozilla.com/D35605

UltraBlame original commit: 9b9bb81548d1d69bb4296f1b190acbc4452efbe0