Bug 1558598 - Force WebRender on or off in the cppunit harness. r=ahal
Also adds it to the mach command, which is a little weird, because the
mach command doesn't expose the option but does parse it via the cpp unit
argument parser. So I just exposed it on the mach command and after that
it Just Works for mach.

Differential Revision: https://phabricator.services.mozilla.com/D35859

UltraBlame original commit: db9b8bf09e16cf4c09a01b7cfa3ebf43b38b2e33