Bug 1558598 - Add ignored option for enable-webrender in jit_test harness. r=ahal
WebRender being on or off doesn't affect the JIT code, so the option is
not really useful in this harness. However, it is a downstream harness that
we'll be passing this flag to so for consistency we have it accept the
flag and silently eat it.

Differential Revision: https://phabricator.services.mozilla.com/D35865

UltraBlame original commit: 459c8d5d00480e04949148abbf9dcfded25a8c0f