Bug 1558598 - Simplify the webrender flags in android_emulator_unittest.py. r=ahal
Instead of using --setenv to control WebRender being on or off, just propagate
the --enable-webrender flag to the downstream "remote" test harness. This
is more reliable than passing --setenv because not all harnesses support
the setenv flag, and it's easier to explicitly add support for --enable-webrender
to the harnesses that need it (i.e. the ones we want to run with WR enabled in
automation).

This also drops the --disable-webrender flag and asserts that lack of
enable-webrender implies WR will be disabled.

Differential Revision: https://phabricator.services.mozilla.com/D35867

UltraBlame original commit: 64a3dc9c192ac036ceb10c7337b30fbb2de3e6b8