Bug 1560193 - Always specify the Python executable to use when running `./mach browsertime` r=nalexander
If `./mach browsertime` runs browsertime with a globally-installed node, due to
an existing bug in [execa][1], the wrong Python will be executed. We now
specify the full path of the Python binary we wish to use (via the `PYTHON`
environment variable that our fork of browsertime supports) and avoid this
issue altogether.

[1]: https://github.com/sindresorhus/execa/issues/153

Differential Revision: https://phabricator.services.mozilla.com/D35374

UltraBlame original commit: d5f3bcfcd6bdf0245264e94e5050aa3dde799938