Bug 1562585 - Don't store urls in input history for dropmarker empty queries. r=adw,dao
When the search string is supposed to be "empty", we should not store a url in
inputhistory's input column, instead we should store an empty input. This happens
for example when clicking the dropmarker and picking something, or when pressing
DOWN on a valid pageproxystate (no user input).
In the future we may evaluate not storing an empty string at all (Bug 1562823).

Differential Revision: https://phabricator.services.mozilla.com/D36568

UltraBlame original commit: a48314ea7f0228413a945bdb88dad355d73aca92