Bug 1525314 - Auto-fuzz WR on Android with maxDifference<=2. r=jmaher
Due to the sheer number of tests that exhibit a random fuzz with maxDifference=1
and maxDifference=2 with WR on Android, it's easier to just tweak the harness
to autofuzz these away. This adds machinery to do so, and also adds a new
annotation that can be used to disable the autofuzzing on specific tests.

Depends on D36794

Differential Revision: https://phabricator.services.mozilla.com/D36796

UltraBlame original commit: 95790a07a93c5a9faf1ededcb0c4c05191e784b4