Bug 1541452 - update cue display when window is resized. r=heycam
When window is resized, the cue would usually be zoomed in or out automatically with the video and keep its relative position to video.

However, if video is being applied the explicit percentage value on its 'width' or 'height', we have to recompute cue's position in this situation, because the width or height of the video would be scaled again after applied the first size scaled which is caused by resizing.

Differential Revision: https://phabricator.services.mozilla.com/D36138

UltraBlame original commit: 16485dcacef6087da868ac4fef2ad829850db195