Bug 1559044 - Improve criteria for deciding whether to kick off download of a font-face resource, so that we don't defer resources that will actually be needed. r=heycam
The original patch in bug 1422530 checked whether we're currently hiding a given piece of text
because we're waiting for a resource to download, and used this as a signal that we should not
initiate another download.

However, that's not really a good criterion to use when we're doing font selection for a given
character, both because it's an indirect signal -- whether painting the text is suppressed will
depend on timing and the font-display property -- and because it doesn't consider whether any
resource that's already being downloaded will actually be relevant for the specific character
we're trying to display.

So this patch instead checks for a relevant in-progress load earlier in the font list during
the specific FindFontForChar call, so that when unicode-range is used to split a font across
multiple subsets, an in-progress load for one subset won't make us defer the downloads of other
subsets if they are also needed for the characters present in the text.

With this, the repeated blinking no longer happens on the site here, as loading of all the
required font subsets is initiated early.

Differential Revision: https://phabricator.services.mozilla.com/D35634

UltraBlame original commit: b6fa827753928bc580ec13edad54d3970e5671f5