Bug 1562645 - Extend mozlint to have license check r=ahal
The goal is to have something simple and not relying on other tools.
I tried with licensecheck and licensee but, with if we use a different
wording to declare "public domain" (as example), they might not
catch it. Requiring to contribute upstream, etc

Instead, I just create a list of line of license to catch it.
From my trials, it works well enough and it is trivial python.

Differential Revision: https://phabricator.services.mozilla.com/D37082

UltraBlame original commit: a817e3f2e0c9c0c5e7ee6cbcf161471a1bd25cc2