Bug 1218456 - Simplify CheckHandleEventForAnchorsPreconditions. r=smaug
I've checked that this doesn't change behavior in the test-case for bug 30178,
which is just an imagemap inside a link.

Overall it'd be weird if we needed this for area-inside-a but not for stuff like
nested links.

Differential Revision: https://phabricator.services.mozilla.com/D37407

UltraBlame original commit: 90ce604ee5c5ee818743b2c7ec559aaabed1bf2b