Bug 1564932 - Fix XDRScript handling of lazy scripts. r=jandem
XDR should be a structural clone, so check if we have a lazy script
directly when deciding if it should be coded. Whether relazification
should be supported is a VM decision, not an XDR decision. This gives a
cleaner round-trip of data when using XDR.

Differential Revision: https://phabricator.services.mozilla.com/D37578

UltraBlame original commit: 9223ff1024a67d3d160b75481905a7079031223f