Bug 1563755 - Add tenured heap survival rate telemetry r=pbone data-review=chutten
This records the size of the heap at the start of collection and maintains a count of bytes freed during collection. From this we calcuate a 'survival rate' at the end of collection. The intention is that this is like the nursery's promotion rate so it's the fraction retained rather than the fraction freed.

Differential Revision: https://phabricator.services.mozilla.com/D37452

UltraBlame original commit: 55605e3b82ef47f64e2f793d85196b1e81e057ad