Bug 1545970 - make sure we only divide or modular TimeUnit on unsigned int32. r=jya
The way we use `operator/` on TimeUnit is usaually to divide 2 for the fuzz or to divide the playback rate, and the rate is always a unsigned int.

The `operator%` should also have same constraint, because there is meaningless to do such a operation on a negative value.

Differential Revision: https://phabricator.services.mozilla.com/D37310

UltraBlame original commit: bf18e1748c4bd33b5ec54a2bd13eacf2d6c9f69f