Bug 1563617 - Quantumbar: Set rolluponmousewheel="true" so that you can scroll in pages while the popup is open. r=dao
With awesomebar we got this for free because nsXULPopupManager checks the "type" attribute of the popup element.  If it starts with "autocomplete", rolluponmousewheel is effectively true [1].  We got rid of the type attribute in quantumbar since it wasn't otherwise necessary.

[1] https://searchfox.org/mozilla-central/rev/07f7390618692fa4f2a674a96b9b677df3a13450/layout/xul/nsXULPopupManager.cpp#361

Differential Revision: https://phabricator.services.mozilla.com/D38136

UltraBlame original commit: caa7933e91e4854fddedbc9ade55e53ab512b21d