Bug 1562881 - Part 3: Merge the connection icon and the extension icon into the identity icon. r=johannh
This patch merges two icons, the connection icon and the extension icon,
into identity icon. And display differently according to the site where
it is. For the site needed to show a lock icon, it will show lock icon.
Otherwise, it will show a proper icon according to the page, such as a
brand icon for chrome page, or a extension icon for a extension page.

Differential Revision: https://phabricator.services.mozilla.com/D37789

UltraBlame original commit: c727ba3536a2b7afdf4eca34ecedc92da702a543