Bug 1566687 - Use mac-cross clang for all macosx builds. r=chmanchester
We've been lucky that non-sanitizer cross-builds for macosx have not
required the clang runtime so far, but they soon will. And it's only
available in the mac-cross clang toolchain, so we need to use that on
all macosx builds.

Differential Revision: https://phabricator.services.mozilla.com/D38260

UltraBlame original commit: bc97fe6dbaa2b29a6da418934abe9d2d8f9424bd