Bug 1545278 - Add js::ObjectMayBeSwapped and restrict transplanting. r=iain
JS_TransplantObject is a very powerful API that results in the JITs
having to worry about the type of objects changing in surprising ways.
In practice though, there are very limited uses of this API so we can
add an API to determine which objects have to worry about transplanting.
This can then by asserted in JITs to document places that optimize
performance by expecting not to deal with transplants.

Differential Revision: https://phabricator.services.mozilla.com/D27975

UltraBlame original commit: 0385c65eab373ec4ec7d7d1e27eefc03950432b0