Bug 1566330 follow-up - De-duplicate DebugEpilogue calls in the interpreter. r=iain
The call needs to be part of the bytecode op for the JIT, but the interpreter can emit
a single call at the end. Results in more compact code and we can now assert unique
callVMs in recordCallRetAddr.

Depends on D38477

Differential Revision: https://phabricator.services.mozilla.com/D38633

UltraBlame original commit: 499c8fa689ad7be86250f89d31a1cca5039abb11