Bug 1557738 - Huffman table lookup;r=arai
This implements lookup in Huffman tables. We try to avoid pulling bits one at a time from the stream (because that's annoying and inefficient and requires abstractions that we don't really have), so we pull about 32 bits at a time and keep them in a bitbuffer. The Huffman table is in charge of looking at
~32 bits and keeping as many as it needs to produce a value.

For the moment, Huffman lookup is O(n), where n is the number of entries, which is very inefficient, but should be simple to test. In the future, we may wish to change representation of Huffman tables to something like
  vector (indexed by a 32 bit key) -> vector (indexed by a 8 bit number of bits) -> *T
this would give us bounded time lookup, which should be better.

Differential Revision: https://phabricator.services.mozilla.com/D34864

UltraBlame original commit: d028bd4f77db5940d5d97667c02f7615b3dee553