Bug 1568269 - Update console architecture diagram in doc. r=ochameau.
The diagram we had was outdated. We delete
the image, and add 2 ascii diagrams that
reflect the current architecture:
- one for the react part
- one for the non-react part

The style of the page is modified so diagrams can take
the whole horizontal space if needed.

Differential Revision: https://phabricator.services.mozilla.com/D39157

UltraBlame original commit: b0a5c300e5257e1d3b0a91261de04f485ef90910