Bug 1242221 - Close IPC channel before MessageLoop gets destroyed r=mayhemer
From the log, I found that the intermittent crash happened because SocketProcessBridgeParent's message loop is destroyed before the IPC channel is closed. Before this patch, the only way to close the IPC channel between SocketProcessBridgeParent and SocketProcessBridgeChild is relying on "content-child-shutdown" notification in child process. If socket process gets terminated before "content-child-shutdown", the IPC channel is never closed.
So, this patch tries to close the IPC channel at SocketProcessChild::CleanUp if the channel is not closed.

Differential Revision: https://phabricator.services.mozilla.com/D38287

UltraBlame original commit: 648bcac47c6c96531732742440280d25fbd80279