Bug 1565301: Disallow too large line numbers for asm.js; r=luke
This is due to an implementation detail of the IonMonkey backend, which stores
call sites bytecode numbers on 29 bits, for packing purposes. The wasm baseline
compiler doesn't have this issue because it stores them on 32 bits. Check it to
avoid an easy browser DOS.

Differential Revision: https://phabricator.services.mozilla.com/D37845

UltraBlame original commit: 11a9758d17c39dd917fbd9aca7830a205a50ed35