Bug 1562301: Stop URL bar hiding tab switcher in GVE. r=geckoview-reviewers,agi
The URL bar was expanding to cover the tab switcher button, meaning that a new tab could not be navigated to after opening.

This is needed so that members of the Gecko media team can replicate an issue in GVE that causes images not to be displayed in Fenix.

Differential Revision: https://phabricator.services.mozilla.com/D39678

UltraBlame original commit: aa7e431e2e2a59add0cd1f1152f44cb9deabd3ef