Bug 1568150 - Stop spinning the event loop when destroying the inspector front. r=gl
InspectorFront.destroy isn't returning a promise. So the `await` in the toolbox
code is not relevant and is only spinning the event loop.
Removing it introduce races which should be fixed.
The toolbox buttons shouldn't try to unhighlight if the inspector isn't used.

Differential Revision: https://phabricator.services.mozilla.com/D39407

UltraBlame original commit: 03824089a0203ffe97f3c567ec9923c9acaa0be2