Bug 1506293 Part 1 - Use next-in-flows instead of next-continuations when converting overflow containers. r=dbaron
In the beginning of nsOverflowContinuationTracker::Insert() there is an
assertion MOZ_ASSERT(aOverflowCont->GetPrevInFlow()). Assume this is
correct, we should use next-in-flows instead of next-continuations when
converting overflow containers.

The crashtest is not added because the assertion can also be triggered
in the reftests added in Part 3 without this patch.

Differential Revision: https://phabricator.services.mozilla.com/D15392

UltraBlame original commit: f2893e234358d30af76f16f9952f3f5f4a351c69