Bug 1488475 - Reject string values starting with 'if ' in wpt manifests, r=ato
Writing disabled: if foo was a common mistake. The easiest fix is just to
require that 'if ' at the start of a string is quoted (one could also
fix this by interpreting these as real conditions, but that would be
more involved, and this fixes the immediate problem of invalid metadata).

Depends on D4952

Differential Revision: https://phabricator.services.mozilla.com/D4953

UltraBlame original commit: 0f926caee5ca3f7a68ebc8b2f27c45522603190d