Bug 1569756: Check for S_FALSE after calling IShellWindows::FindWindowSW; r=jmathies
`IShellWindows::FindWindowSW` may return the dreaded `S_FALSE`, so we need to
ensure that we're handling that.

I checked the remaining functions called by this code and none of the others
do this; this is the only call site that requires the check.

I'm not sure why we're seeing this error code. I added an explicit cast to
ensure that `CSIDL_DESKTOP` is being interpreted as an `int`, though I doubt
that this actually makes any difference.

Differential Revision: https://phabricator.services.mozilla.com/D39799

UltraBlame original commit: 82449212da88f1ffdb4d2b50df3f77d817cb314c