Bug 1570575 - Remove `MirrorKind`. r=KrisWright
This commit removes `MirrorKind` by splitting the `VARCACHE_PREF` macros in
two, giving `ALWAYS_PREF` and `ONCE_PREF` macros. This is good because most of
the time the generated code is quite different for the two cases.

The commit also removes the examples of code produced by the macros in
comments. These are kind of useful, but also quite verbose and a pain to
maintain.

Differential Revision: https://phabricator.services.mozilla.com/D40166

UltraBlame original commit: 89b4edfa815032849012bbb7d24f980348201831