Bug 1387894 - Fix browser_animation_setCurrentTime.js for unconditional clamping. r=birtles
This has to do with double imprecision. The test originally had toPrecision(6) to
account for this imprecision. It'd round up 499.9999 into 500. When we send
double(500) (which is an epsilon below 500) into ReduceTimePrecision we wind up
coming out with 499.98. By reducing our precision requirement in this test
we can handle that and round 499.98 back up to 500

Differential Revision: https://phabricator.services.mozilla.com/D38809

Depends on D38808

UltraBlame original commit: f07ba3e1ae0d98af129619c5f211b795375ac0ba