Bug 1568138 - Reverse order of Huffman keys;r=arai
In Context 0.1, Huffman keys are represented with a bit order that is the reverse of the bitstream. This
patch ensures that Huffman tables reverse the bit order of their keys upon initialization, so as to allow
fast comparison later in the code.

Differential Revision: https://phabricator.services.mozilla.com/D38979

UltraBlame original commit: dd401c5ee827abb735909f5ab09488a10e7a20bb