Bug 1548866 - Fix browser_webconsole_keyboard_accessibility.js intermittent. r=davidwalsh.
The test was supposed to pass custom values to waitFor (for
the number of tries and the interval), but what we end up
using in the webconsole is the waitFor function from the
debugger helper, which is loaded via loadSubScript.
Since the debugger waitFor doesn't allow to pass custom
maxTries nor interval, we are not waiting long enough
for all the messages to be displayed on slow machines.
The quickfix here is to allow debugger waitFor to have
custom maxTries and interval.
A better fix would be to actually not load the waitFor
from the debugger, for example by avoiding loadSubScript
and using modules instead.

Differential Revision: https://phabricator.services.mozilla.com/D40163

UltraBlame original commit: 60ba89028ca86c6fddfb59009f99ccd21167f41e