Bug 1569713 - Add wpt-metadata YAML export to metasummary, r=maja_zf
https://github.com/web-platform-tests/wpt-metadata/ contains metadata
for web-platform-tests in YAML form. This will be searchable on
wpt.fyi and so it can be the basis for a triage feature in the wpt
dashboard.

This adds a basic export of the most obvious annotations from the wpt
metadata-summary mach command. It can be used by setting the
--meta-dir variable to the root of a wpt-metadata checkout.

Differential Revision: https://phabricator.services.mozilla.com/D39736

UltraBlame original commit: 4c812f05096ca71823501c263d41f50e24688012