Bug 1570663 - Align favicons in the awesome bar with the shield icon if the pageproxystate is valid. r=adw
This patch changes the alignd icon of the awesome bar to the shield icon
if the pageproxystate is valid. This is needed since the tracking
protection icon would be at the first position if the 'pageproxystate'
is valid.

Differential Revision: https://phabricator.services.mozilla.com/D40981

UltraBlame original commit: 790e7615ee182a308244d9d5856cbab84f308f4b