Bug 1566643 - Part 11: Use null principal for initial document when changing process, r=kmag
Previously, the principal of the embedder would be used, which would mean a
document with the wrong principal would be created in the target process. By
using a null principal we can ensure that this initial document will be replaced
almost immediately.

In a follow-up, we should switch this to using the principal of the document to
be loaded to avoid unnecessary intermediate window globals.

Differential Revision: https://phabricator.services.mozilla.com/D41236

UltraBlame original commit: 9dd2bc7bc4aff1b0dab44abda8f386fbaa7195b4