Bug 1572389 - Add pref to show normal lock icon for sites with EV (Extended Validation) certificates. r=johannh
This is probably the simplest way to implementing this. We might also decide to show the normal lock icon, but still
show a bit of a different text in the popup. I think this that might just be confusing though.

Differential Revision: https://phabricator.services.mozilla.com/D41285

UltraBlame original commit: b1f185fcab6203063b7b7a531a35da10e638eee7