Bug 1572779 - Update test expectations for perspective-scrolling-1.html. r=kvark
This bug became an unexpected PASS on tier-2 android platform after
landing the fixes for bug 1572197. The fixes in that bug were
specifically to fix this test on windows platforms, so it's not
surprising that it also fixes it on android.

Differential Revision: https://phabricator.services.mozilla.com/D41493

UltraBlame original commit: 5c879cc0473ca4ff19898d3016292c56299aded3