Bug 1569570 - Select elements fail to display dropdown in RDM r=bradwerth,layely
The `xul:menulist` appears a little offset but this is the same as it was prior to the fission patch that completely broke it.

If you have an alternate approach or a way to get the correct position for the `xul:menulist` I am all ears.

Differential Revision: https://phabricator.services.mozilla.com/D40805

UltraBlame original commit: 81d748db6e07fe475edceb6bd95a689097db2ad4