Bug 1508177 - Expand the minimum scale height even if the expanded area doesn't contain any contents. r=botond
As a result of the expansion, position:fixed elements are attached to the
expanded layout viewport.

The expanded value is used behind a pref which is enabled by default on nightly
initially, and the pref will be fliped in bug 1571599 on other channels.

scrollbars-in-landscape-content.html still fails since the vertical overlay
scrollbar doesn't appear since we are not yet using the expanded value during
reflow to tell whether we need overlay scrollbars or not.  This will be fixed
by the next commit.

Differential Revision: https://phabricator.services.mozilla.com/D40771

UltraBlame original commit: eb738d246d4059431c420843e86ee485084d8ddc