Bug 1574930 - Part 3: Add a test case for making sure that the track counter displays properly and fix an issue for BrowserTestUtils.waitForAttribute(). r=nhnt11,johannh
This patch adds a test case to check the visibility of the track counter
is correct in different situations, including zero tracker, one tracker
and multiple trackers.

Also, this patch fixes an issue that the
BrowserTestUtils.waitForAttributes() won't work if the waiting
attributes doesn't have a value or has an empty string as its value.
We check the value of the observing attribute to verify if it is
existing, however an empty string will be treated as a false, so it
won't pass the check if the attribute doesn't have a value. Hence, we
should use hasAttribute() instead of getAttribute() to check the
existence of the attribute.

Differential Revision: https://phabricator.services.mozilla.com/D42718

UltraBlame original commit: 5dbac06bc222985e95f30d6b372c64fcac8b817d