Bug 1575783 - Fix bug 1551639 after bug 1573566. r=froydnj
Bug 1573566 moved libxul in a subdirectory, which means the list of
files it links are now relative to that directory, rather than the
directory where buildid.cpp is. So the dependencies for buildid.cpp
need to account for that.

Differential Revision: https://phabricator.services.mozilla.com/D43032

UltraBlame original commit: 4dfe4abe05d8b9c7c692c7187df7435674a6d8b7