Bug 1547224 - Enable inactive CSS detection for all releases. r=gl
As discussed with mbalfanz, we want to flip the pref so Inactive CSS is enabled and rides the trains.
This will enable the feature on Firefox 70 Beta/DevEdition allowing us to collect early user feedback and issues.

If all goes well during Beta/DevEdition, the feature will ride the trains to Release.
If there is a major blocker, we will uplift a patch to disable the pref so it doesn't proceed to Release.

Differential Revision: https://phabricator.services.mozilla.com/D43265

UltraBlame original commit: 8889774ae7bc97042846df84e7f9addc952249a8