Bug 1559592 - [marionette] Take screenshots for content in parent process. r=webdriver-reviewers,ato,automatedtester
This moves all the screenshot related code from the framescript to
the parent process, so that canvas.js is no longer called from
within a content process.

The remaining code in the framescript is only needed to compute
the dimensions of the screenshot, and from all the to highlight
elements.

This move is necessary to allow switching to the new drawSnapshot
API which only works from within the parent process.

Differential Revision: https://phabricator.services.mozilla.com/D40209

UltraBlame original commit: 5fee1e8e06c6cab1755f15e0fcf6ffb0cb186ba9