Bug 1559592 - [marionette] Make use of the new Fission compatible screen capture API. r=webdriver-reviewers,ato,automatedtester
To allow Marionette to capture the content of OOP iframes, the
new Fission compatible screenshot API has to be used.

For callers requesting a read-back of the pixel data as
currently in the compositor/window, the 'drawWindow()'
method still has to be called.

Differential Revision: https://phabricator.services.mozilla.com/D40654

UltraBlame original commit: c4ce2527e763199e2c43ef24b1f4043aeefd21e0