Bug 1576707 - Allow "absolute" fetch artifacts to ignore `artifact_prefix`. r=aki
The way that `artifact-prefix` works is that it applies to the
_consuming_ task's fetches... meaning that changing the prefix
requires working out the new paths for every fetch in the consuming
task.  This allows to "opt-out" of the artifact prefix mechanism for
individual fetches.

Depends on D43534

Differential Revision: https://phabricator.services.mozilla.com/D43535

UltraBlame original commit: 4a7e698d5033d9703aa7e91d68ffe793d55a5a08