Bug 1572013 - Fix grammar walk order;r=arai
The walk order in the grammar is a bit unusual. Some items are treated immediately, while others are enqueued for later use (maybe this is a form of manual TCO?) This patch fixes the order to make
it match with the Python implementation.

Also, we add `readUnpackedLong` as some numbers are not stored as varnums.

Depends on D42845

Differential Revision: https://phabricator.services.mozilla.com/D42846

UltraBlame original commit: 1837e5c1b1b2600854862ef5c0cd727ac731a506