Bug 1579025 - Split schedulegc testing function into two separate functions r=sfink
This splits schedulegc() into schedulegc() and schedulezone() and adds an option parameter to gcstate() to report the GC state of a zone. I moved gcstate() so that it is not only present in zeal builds and updated test uses of schedulegc() as appropriate.

Differential Revision: https://phabricator.services.mozilla.com/D44840

UltraBlame original commit: 11d7643d276d3e1c7707b48a9c01bfa0c2f9bd36