Bug 1579820 - Handle window.open() from remote subframes. r=nika
The CommonCreateWindow code requires having a BrowserHost for the tab that's
creating the window, which it tries to get from the requestor's BrowserParent.
For remote BrowserParents, though, there is no BrowserHost, so we need to get
it from the top-level embedder instead.

Differential Revision: https://phabricator.services.mozilla.com/D45172

UltraBlame original commit: dd9684586ba31d92037cb0cf16227d5969fb670b