Bug 1570476 - Don't fetch the request stacktrace in StackTracePanel's componentWillReceiveProps if the request won't change. r=Honza.
This was causing an intermittent in console test because
multiple requests were sent to the server, and sometimes,
some of them after the connection being closed, which
was making the test fail.
With this patch, we only try to fetch a new stacktrace
on componenetWillReceiveProps if the component will have
to deal with a new request.

Differential Revision: https://phabricator.services.mozilla.com/D45808

UltraBlame original commit: cfc5a1ad7e1075bf7ca7e37e0eb834483932d764