Bug 1579972 [wpt PR 18942] - [LayoutNG] Fix line breaking behavior after out-of-flow objects, a=testonly
Automatic update from web-platform-tests
[LayoutNG] Fix line breaking behavior after out-of-flow objects

This patch changes line breaking behavior after out-of-flow
objects to match to ICU, assuming the out-of-flow object is
an object replacement character.

The behavior is not well-defined, and not all cases are
interoperable across existing implementations. Legacy/WebKit
resets the prior context so that text after out-of-flow
objects behave as if it is at the top of the line. Gecko and
Edge seem to allow break in most cases, even when NBSP
follows, but not when the line overflows.

The behavior implemented in this patch is interoperable for
all tests that pass on existing 4 implementations, but not
exactly same as legacy when implementations do not agree.

Bug: 1001438
Change-Id: I4c7077d2c287a6e897175b29a73ce1c308f29f4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1792467
Reviewed-by: Ian Kilpatrick <ikilpatrickchromium.org>
Reviewed-by: Aleks Totic <atoticchromium.org>
Reviewed-by: Emil A Eklund <eaechromium.org>
Commit-Queue: Koji Ishii <kojiichromium.org>
Cr-Commit-Position: refs/heads/master{#694888}

--
Load ahem as a webfont to fix lint error.

--

wpt-commits: 0dba391be24cebb9bbbd22d63df729193e502902, 992be2f9e48c9a8cab93ef3b393d548f3beea4ef
wpt-pr: 18942

UltraBlame original commit: c052b1b762b54edebf5ae3e01fb260eae348d364