Bug 1580772 [wpt PR 19007] - Update wpt scroll-to-visible-areas tests, a=testonly
Automatic update from web-platform-tests
Update wpt scroll-to-visible-areas tests

The original tests were based on an inaccurate understanding of
"visibility" for a snap area, where a snap area was considered to be
visible if it is in the viewport at the scroll location before the
snap. The definition of visibility per the specification is that a snap
area is visible if it would be in the viewport at its snap point.

See:
https://github.com/w3c/csswg-drafts/issues/2526#issuecomment-380371186

Also added tests for the consideration of scroll-margin when
calculating visibility as well.

Bug: 954851
Change-Id: Ic2ffef8d02fc5ea2439054b6ea5185bb66622d1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1795308
Commit-Queue: Kaan Alsan <alsangoogle.com>
Reviewed-by: Yi Gu <yiguchromium.org>
Reviewed-by: Majid Valipour <majidvpchromium.org>
Cr-Commit-Position: refs/heads/master{#696435}

--

wpt-commits: 224ecb27547a9b49b28b7ae06abc8bc09f4b91fc
wpt-pr: 19007

UltraBlame original commit: 71069e4c04c1df9f346829c6860c171b554cbd46