Bug 1572281 - In DeviceChangeCallback class separate the observer from the subject functionality. r=pehrsons
DeviceChangeCallback class implements the observer pattern. However, the role of the observer and the subject is integrated into the same class which makes use of virtual methods to allow a separation of the roles. This makes code reading difficult. Also, it does not allow from a class to inherit only the observer role or the subject role. This patch breaks the DeviceChangeCallback class into two classes according to the observer or subject functionality.

Differential Revision: https://phabricator.services.mozilla.com/D46270

UltraBlame original commit: 90286c5d1ef171475be1d02d337825ef5fcf3b14