Bug 1583138 - Prepare clang-tidy for clang-9 r=andi
In clang 9, the signature of `registerPPCallbacks` changed and it no longer provides us a `CompilerInstance` [0]. I couldn't find a way to reach the `CompilerInstance` from the new structures, so I've inserted a level into the class hierarchy that mimics the old functionality.

With this patch, clang-tidy compiles with either clang-8 or clang-9, so we can land this independently in preparation for bug 1573211.

[0] For reference: the new signature was introduced in https://github.com/llvm/llvm-project/commit/bbc89dcb29035ad450ac0981b8f51a6cfa2aa8ba, the old overload was marked deprecated in https://github.com/llvm/llvm-project/commit/e9087fe75c8652a442bada0a7e02cba0df781074 and removed in https://github.com/llvm/llvm-project/commit/ee737a84d75b701c2ae3a8c9a0abbae63fc330af.

Differential Revision: https://phabricator.services.mozilla.com/D46702

UltraBlame original commit: a96435a6dc5417a0d2a8c356ac5ddd7e9852804c