Bug 1577574 - Only assert that string is propely lowercased in AlternateServices::SchemeIsHTTPS r=kershaw
According to the comment above, SchemeIsHTTPS should return an error if the
input is neither http or https.
But we don't assert that the string is not lowercased (as the comment says),
instead we assert if it's not http or https, meaning in debug mode the
function would always panic for such an input.
This patch changes the MOZ_ASSERT to match the comment and only assert if
for some reason the input scheme is not lowercase.

Differential Revision: https://phabricator.services.mozilla.com/D46279

UltraBlame original commit: 455e683e5479f14cfae6002d556e238a28fc4e9d