Bug 1579270 - Disable TLS 1.0 and TLS 1.1 in Nightly, r=jcj
This flips the default for security.tls.version.min to 3 (TLS 1.2) for the
Nightly channel.

Having had this pref at this level for the last year, I can confirm that this
does break the occasional site, but it is quite rare.  The intent of this change
is to start making it more obvious when sites don't support TLS 1.2.

I'm asking for wider review because this is a disruptive change.

Differential Revision: https://phabricator.services.mozilla.com/D45627

UltraBlame original commit: 0781e60dd54c1a2dc2f622445ef8c207ce5ff96c