Bug 1582865 - Don't use document/window in tree component. r=yzen.
When using the ObjectInspector in an inspector extension sidebar,
document is not an HTMLDocument but a Sandbox, which means we
don't have access to the same properties than in a document.
Since the Tree and the ObjectInspector components do use `document`
and `window`, this means some interaction would lead to the sidebar
crashing.
To prevent this, we retrieve the document and/or the window from
a node reference instead.

Differential Revision: https://phabricator.services.mozilla.com/D47271

UltraBlame original commit: b19d946a62e06c2df1027393a7679d0b6467113b