Bug 1581875 - Introducing MultiLookupHuffmanTable;r=arai
The MultiLookupHuffmanTable is designed to allow fast lookup in tables that
have a max bit length too large for the SingleLookupHuffmanTable.

We will probably need to tune the bit lengths at which we switch from SingleLookup to TwoLookups or ThreeLookups in followup patches.

Differential Revision: https://phabricator.services.mozilla.com/D46806

UltraBlame original commit: 9ed9365ba660bed1e72369472d322be2244b4135