Bug 1546185 - Also check mEventDispatchingSuspended to see if we are in sync event loop r=baku
Only useing `mFlagSyncLooping` to check if the current code is in sync event loop is not enough. The crashtest in this patch shows that `mFlagSyncLooping` is set to false in OnStopReequets while we are still in sync event loop.
`mEventDispatchingSuspended` is set to true right before calling `SpinEventLoopUntil` and set to false after, so we should also use it to check if `SpinEventLoopUntil` is finished.

Differential Revision: https://phabricator.services.mozilla.com/D47240

UltraBlame original commit: b24092b7bd4989e85494022922d766d120536cf2