Bug 1584244 - Fix invoke getter action on quoted property. r=Honza.
When clicking the invoke button on a property that
needs to be quoted, e.g. hello-world, the result
was undefined.
That's because we were sending the property name to
the server with the surrounding quotes.
This patch fixes the issue and adds a couple of test
case to make sure we don't regress.

Differential Revision: https://phabricator.services.mozilla.com/D47399

UltraBlame original commit: 046a8c9987d95355780330e6c9a66523b8a5e235