Bug 1582210. Support splitting blob images into different sizes. r=nical
This makes things better especially when the bounds of the combined blob
is changing but the bound of the separate ones are not.

The current implementation is a bit ugly, but it's simple and
can be cleaned up in the cleanups I have in mind for the future.

Differential Revision: https://phabricator.services.mozilla.com/D47983

UltraBlame original commit: 4c696f6996de4681b114a6cf17f4b750d5363f68