Bug 1586911: Silence SHEntry assertion after local->remote->local iframe navigation. r=peterv
When we have a parser-created iframe which starts out in-process, transitions
to remote, and then transitions back to in-process, we create separate
DocShells for the first and last in-process loads. Since both are
network-created, and have the same child index, they both try to add
themselves as children to their parent's SHistory at the same index. And since
the entry for the first DocShell already exists at that index when we try to
add the second, that triggers an assertion.

This isn't really ideal, but it is expected given the current state of session
history under Fission. It should hopefully be solved more gracefully when the
Fission-aware session history rewrite is done, but in the mean time, I think
we should just ignore the conflict, since it's expected.

Differential Revision: https://phabricator.services.mozilla.com/D48437

UltraBlame original commit: c0029f42afd2873d7a7f3a5db4b40bfbb3865e18