Bug 1568883.  Remove the QI implementation from Web IDL elements.  r=peterv,mossop
The XBL test is being removed because it was the only remaining consumer of
xbl's implements="interfacename" in the tree, and was triggering QI on elements
for that codepath.

I've verified that a try run that MOZ_CRASHes when the C++ binding
QueryInterface implementation is invoked is green with these changes.

Differential Revision: https://phabricator.services.mozilla.com/D48249

UltraBlame original commit: 02578c5bb2032fa0d91e712a03667b9757314ed0