Bug 1584368 - Fetch Normandy recipes from capabilities-aware locations r=Gijs,leplatrem
This indicates to the Normandy server that this client understands the
capabilities model, which is used for backwards compatibility. Without this
change, the server will only send recipes that are compatible with the
functionality that existed in Firefox 70, and other recipes will not be
included. With this change, we'll get the full picture of recipe.

Differential Revision: https://phabricator.services.mozilla.com/D48448

UltraBlame original commit: e1a1b2085ee21cd5a02cd69c6f7b56185841f8da