Bug 1586903 - Wait for track listener removals rather than just informing them of shutdown. r=bryce
This does a bit of a cleanup, where changing the notification to waiting for
removal is the major task. It also removes the special handling of not informing
listeners of shutdown on Cancel(), and a bit of cleanup around MozPromise usage.

Differential Revision: https://phabricator.services.mozilla.com/D49416

UltraBlame original commit: 509719f98963a8f1c2663dac01279a734d30e71a