Bug 1586201 - Add a function to get a nodeFront from a ContentDomReference. r=pbro,jdescottes.
A function is added on the walker actor that creates a NodeFront
from a ContentDomReference, e.g. an object containing a browsingContextId
and a unique DOM element identifier.
A trait is added on the walker actor since the ContentDomReference API was
only added in Firefox 69.
We then add a function on the toolbox that can return a NodeFront from a
element grip.

Differential Revision: https://phabricator.services.mozilla.com/D48808

UltraBlame original commit: ba7304bbc693865f655b9fb4c4196fdda44192b9