Bug 1577896 - Fix GridElementResizer on overflowing container. r=fvsch.
We remove the pseudo elements we were using to draw the borders and
put them on the grid elements instead.
We make the GridElementResizer 6px and translate it so there's
handle on "both sides" of the splitter.

Differential Revision: https://phabricator.services.mozilla.com/D45302

UltraBlame original commit: 8ddcca74cb17d67e1ac21a5198b3665de82d0db9