Bug 1590083: Disable wasm multi value with Cranelift; r=lth
Since wasm multi-value is enabled by default (without a shell switch), there
was no way to signal it was disabled in certain configurations, namely
Cranelift.

Instead, assume multi-value is enabled by default and disable it with
Cranelift, and store this information in the CompilerEnvironment so we can use
it when iterating on the wasm binary.

Differential Revision: https://phabricator.services.mozilla.com/D49925

UltraBlame original commit: 8204dd5d6f50b6090446793947ccfde46bc9bb2d