Bug 1589766 - Experiment with implementing zoom as a transform + transform-origin shorthand. r=xidorn
This is a gross hack, of course, but has the advantage of not breaking sites
that use both zoom and -moz-transform / -moz-transform-origin.

There should be no behavior change when the pref is off, of course, and the
webcompat team wanted to experiment with this.

Differential Revision: https://phabricator.services.mozilla.com/D49792

UltraBlame original commit: 97514cd3237ce93aca88458d53643b2ac551128e