Bug 1591968 - Put window.mozPaintCount behind a default-off pref. r=bzbarsky
It's not the kind of thing we want people to allow observing, generally, and
even less so the kind of thing that we may want people to rely on.

Move internal callers (all of them tests) to a new DOMWindowUtils.paintCount
method.

Differential Revision: https://phabricator.services.mozilla.com/D50817

UltraBlame original commit: 8c8611f1f35f0579d8272019927fe55589393607