Bug 1584602 - Enforce eval restrictions in Workers and do not enforce restrictions in Release r=ckerschb
This commit does two things. Firstly it enforces eval restrictions in Workers per Bug 1584602.
We're collecting telemetry on these in Beta (and not seeing any) so we can let enforcement ride up to Beta.

Secondly, it disables enforcement checks on Release (and late Beta, as explained in the comment) until
we can gather data about what's happening in Release. This is a counterpart to Bug 1592349 for -central.

We have two separate commits because the first part of this is a change in the same code and we'd have
rebase problems if we tried to do them both separately.

This does tie enforcement to a build-time constant instead of leaving it as a pref. This doesn't make
me very happy inside, but I don't think the extra complexity is worth it...

Differential Revision: https://phabricator.services.mozilla.com/D50970

UltraBlame original commit: 8cce153f4722511b81123e289987a29c267b80fc