Bug 1591101 - bookkeep orphaned registrations r=asuth
Keeping track of these registrations is needed for when
ServiceWorkerManager::MaybeStartShutdown is called for the first time. At this
point, all living service workers must register themselves with the
ServiceWorkerShutdownBlocker. Non-idle registrations that are unregistered and
not controlling clients would otherwise not be found in the SWM's data.

Differential Revision: https://phabricator.services.mozilla.com/D50500

UltraBlame original commit: 6c15753220e01a46d57a5b8694ef1c51dcbd2262