Bug 1591047 part 2 - Split memCopy/memFill implementations for shared/non-shared modules. r=lth
Whether a module uses shared memory or not is fixed throughout its lifetime. We
can use this to specialize the implementation of memCopy/memFill and remove a
branch on the memory type. This will also be useful when acquiring the memory
length in a future commit, which will require different code per shared-ness.

Differential Revision: https://phabricator.services.mozilla.com/D50375

UltraBlame original commit: 2a4f822999200ebbe1600ace0c0670dd813f7a20