Bug 1545123 - store flash information in prefs instead of pluginreg, r=handyman
By storing the plugin information in prefs when only flash is allowed, we can
avoid reading pluginreg and doing a plugin scan on the mainthread on startup.

As part of this, we're now keeping track of the 'is flash allowed' pref on the
plugin host, and no longer write 'valid' plugin info into pluginreg if so.

Also note that in this commit, we're changing `mPluginRegFile` to actually
refer to the file, rather than the containing directory.

Differential Revision: https://phabricator.services.mozilla.com/D48328

UltraBlame original commit: dc85aaeb9ee91f2df0a5c03106538138a9adbadb