Bug 1593622 - Test console autocompletion after ?? (nullish coalescing operator). r=Honza.
The nullish coalescing operator was added last week in Spidermonkey.
It looks like there isn't anything to do to support it on the console
side, so we only add a test for the autocompletion after it, to make
sure we don't regress this in the future.

Differential Revision: https://phabricator.services.mozilla.com/D51632

UltraBlame original commit: 0f0c1fc3c0ca693f43dbc5968dcdbd041c7d5b5d