bug 1549708: docshell: add nsIDocShell.isForceReloading; r=bzbarsky
Exposes a new nsIDocShell API, isForceReloading, to determine if
the loaded document was force-reloaded or not.

It relies on the underlying behaviour of nsDocShell::IsForceReloading(),
which again relies on nsDocShell::IsForceReloadType(mLoadType).

The getter is used in the remote agent to test that
Page.reload({ignoreCache: true}) works as intended.

Differential Revision: https://phabricator.services.mozilla.com/D51435

UltraBlame original commit: ece31d77808bb891c83cf4dacfd9284782a6c936