Bug 1594530 Make tooltool via the taskcluster proxy work on both legacy and firefox-ci clusters at the same time. r=tomprince a=Aryx
MANUAL PUSH: (a) This patch will cause a ton of toolchain rebuilds, and might as well do that on central right now rather than autoland, and (b) We want to test the new Taskcluster instance today, and will be testing THAT on m-c, so we'll need this patch on m-c before we can test the new cluster as well.

tooltool at present needs to support production (legacy cluster) but its auth system is tied to that cluster.

Which means that using tooltool in the new cluster ahead of TCW is harder. We have swapped the credentials for the tooltool staging deployment to use the new tc cluster, so when we're using the taskcluster proxy we need to have it swap between legacy and new tooltool url's depending on which cluster (ROOT_URL) we're using.

This patch is intended to be ok to land on production code today, and could be backed out after the TCW when production tooltool will be configured to work with the firefox-ci cluster itself.

Differential Revision: https://phabricator.services.mozilla.com/D52089

UltraBlame original commit: 69049d8a2c4b9ee877d2153ce6f01fd0549f90c3