Bug 1594275 - Parse the exportparts attribute. r=bzbarsky
But don't hook it into style yet, that'll be a follow-up patch.

I had this patch in my local queue for a bit and there was no point in not
landing it I guess.

The value of this attribute could be stored only in the shadow root (as this
only applies to shadow hosts), but that would make invalidation harder, I think,
so do the obvious thing for now.

Differential Revision: https://phabricator.services.mozilla.com/D51963

UltraBlame original commit: 4d77157a8980ce8842f53ef8a8911aa2982c0321