Bug 1593203 - Proper initialization of nested dictionaries for MediaDebugInfo.webidl  r=pehrsons,bzbarsky
The debug dictionaries in  MediaDebugInfo.webidl all have default values, and the intent when the debug structure is created by the C++ promise is to initialize all values, including nested dictionaries, to the provided defaults. **required** was not the right way to do this.

Differential Revision: https://phabricator.services.mozilla.com/D51822

UltraBlame original commit: a04cf95c417d672597ff34e0f50218333ac7608e