Bug 1594417 - Ensure panel-list is hidden when added to document r=Gijs
When the stylesheet needs to load for panel-list it is missing the rule to
hide the panel while it's closed. Move this style into the shadow DOM so
it's always present even if the other styles haven't loaded yet.

This was especially noticeable when the "Disable Cache" option was enabled
with the devtools open.

Differential Revision: https://phabricator.services.mozilla.com/D52063

UltraBlame original commit: dff542b772e590137a2e3d9785f5c32878f256f1