bug 1590965 clear GtkClipboard on nsIClipboard::EmptyClipboard r=stransky
This notifies GTK that the data is no longer available for clipboard_get_cb(),
so that GTK will no longer advertise nor attempt to store the data.

The clipboard is not cleared immediately before it is set in case that might
trigger a system clipboard manager to take ownership of the selection, which
could lead to gtk_clipboard_set_with_data() failing to take ownership back due
to an older timestamp.

This also fixes a GtkTargetList leak.

Differential Revision: https://phabricator.services.mozilla.com/D50764

UltraBlame original commit: 2905fbf1625f2613c3532bdf64b93676fb3d6ae1