Bug 1565902 - Test the impact of removing startup dll readahead r=glandium
We haven't tested this in recent times, and it would be good to understand
what the impact is looking at telemetry measures of startup in Nightly.
This doesn't rip out everything, but we will need to do that if we
determine that the readahead has a neutral / negative effect.

Differential Revision: https://phabricator.services.mozilla.com/D50512

UltraBlame original commit: 5f0b392beadb7300abdaa3e5e1cc1c0d5a9f0791