Bug 1592444 - Allow E_ACCESSDENIED from CoAllowSetForegroundWindow in debug build. r=aklotz
Bug 1588500 reported some of Thunderbird's tests failed because
`CoAllowSetForegroundWindow` failed.  In that case, those tests should not launch
a default browser of the system, but we should allow `E_ACCESSDENIED` because
it legitimately happens if the window is not in the foreground.

Differential Revision: https://phabricator.services.mozilla.com/D52618

UltraBlame original commit: 138b1dd63e101b22e3b3aa008516efe543e1dcc9