Bug 1585410 - Implement and document 'deletion-request' ping r=janerik,Dexter
The 'deletion-request' ping, which supercedes the 'optout' ping, notifies the
pipeline when a profile opts out of FHR upload. (IOW, when a user on a specific
profile unchecks the box in about:preferences#privacy about sharing
technical and interaction data with Mozilla).

This ping tries its best to reach the pipeline to let them know that we need
to delete data associated with the provided clientId. This means it will remain
pending on the client even after opt out and it will try to resend if upload is
ever re-enabled.

Differential Revision: https://phabricator.services.mozilla.com/D51710

UltraBlame original commit: 5205263485b69a2a9661a8a3a8572c829053df31