Bug 1590907 - Remove obsolete macro JS_OOM_DO_BACKTRACES. r=sfink
This macro isn't defined anywhere and doesn't seem to do anything. It
affects the oom-backtraces property of the build configuration object in
the testing functions, but since the macro is never defined, it seems to
be always set to false anyway, so just hardcode it.

Differential Revision: https://phabricator.services.mozilla.com/D51769

UltraBlame original commit: bedd96446c984c7bee5dae402274e0427a42cb98