Bug 1596342 - Update sending patch documentation to only mention Phabricator + moz-phab. r=rcaliman.
The documentation was mentioning arc as the primary
tool to push to review (and was discouraging using
moz-phab). moz-phab is now much more stable and
is the prefered way of pushing patches, so we shouldn't
mention arc anymore.
There was also a (not recommended) mention of attaching
patch to Bugzilla for reviews, which I think isn't
something we recommend anymore.

Differential Revision: https://phabricator.services.mozilla.com/D52969

UltraBlame original commit: ee728f3ea085900757752d009ea271d21087096a