Bug 1588241 - P3. Move classification methods to PHttpChannel. r=mayhemer
We can guarantee that those callbacks will occur prior the call to OnStartRequest.
But running them all on the PHttpChannel we remove the potential race between the PHttpChannel which runs on the main thread and PHttpBackroundChannel which runs on the necko thread.

Differential Revision: https://phabricator.services.mozilla.com/D52262

UltraBlame original commit: 2dda521c0591c002217453ea3e8358651ad6a45d