Bug 1593170. Make the reftest harness deal with SynchronizeForSnapshot and FlushRendering returning promises. r=mattwoodrow
The code comment mostly explains the design. Basically, we force nothing to happen while we wait for the promises to finish and instead record what we need to do once the promise is finished, and do those pending tasks when it's finished.

Differential Revision: https://phabricator.services.mozilla.com/D51344

UltraBlame original commit: a004de6493422688953b5b8ea2fc5e1522240f47