Bug 1592150 - Make NativeLayer::NextSurfaceAs* copy existing drawing from previous buffers so that only the update region needs to be drawn. r=jrmuizel
There are three reasons for doing this.
1. It makes the NativeLayer API more compatible with DirectComposition.
2. Copying existing content may be faster than redrawing those pixels. Redrawing
   might have some amount of overdraw which takes up more memory bandwidth.
3. Most importantly: Partial updates now have "unidirectional flow of information":
   The renderer decides which area to redraw, and it redraws exactly that area.

In the past, partial updates required the following dance:
 - Figure out what area changed in this frame. Call that area A.
 - Invalidate that area in the NativeLayer.
 - Get the next surface for drawing from the layer.
 - Request the actual invalid area in the current surface. Call that area B.
 - Redraw B.

Now with this change, the renderer no longer needs to care about B, and can just
redraw what changed in the current frame (A).
This is useful for WebRender because WebRender prepares drawing commands on a separate thread
before it executes them on the render thread. And at the time of preparation, WebRender does
not have access to the native layer. It needs to know what to draw ahead of time.

Differential Revision: https://phabricator.services.mozilla.com/D51760

UltraBlame original commit: ffcddfbaf321cb2f7ca87fe3ab68a1347b4328aa