Bug 1582944 - Get the Search cache working properly with the modern configuration. r=mikedeboer
This makes visibleDefaultEngines in the cache redundant, and replaces it with a new builtInEngineList that is indexed via the id and locale of the WebExtension. This gives a better indexing method that just display name.

Also bump the cache version to 2, so that we can do safe downgrades/upgrades with pref flips that might not happen at the time of an upgrade.

Differential Revision: https://phabricator.services.mozilla.com/D53189

UltraBlame original commit: 5c271dd2d41a5e1a288bb0ef2580575a8008610d