Bug 1594795 - Fix local browsertime issues on Windows and MacOS. r=perftest-reviewers,barret,stephendonner,davehunt
This patch fixes local `./mach browsertime` issues on Windows.

First issue that is fixed is that the archive is extracted into the browsertime folder instead of a separate folder for imagemagick within the browsertime folder. Now, it is extracted into it's own folder whose name depends on the `fetches` path entry. Second issue fixed is that the environment was being contaminated with bad strings because of the `GECKODRIVER_BASE_URL` addition - using `str` on the entries solves this issue.

Finally, convert and compare were not passing in the visualmetrics.py check call. This is because ImageMagick expects to find the path to them within `ProgramFiles` on windows. Since we don't have them there, we have to add an entry in the `PATH` environment variable to point to the imagemagick directory to get around this.

There is also an issue with the macosx imagemagick directory which was being extracted as 7.0.9, but we expected 7.0.8. The imagemagick version used is modified to the correct version with the expected directory.

Differential Revision: https://phabricator.services.mozilla.com/D52751

UltraBlame original commit: 72fd29e8c8eed83c494b32165ec052d27f541694