Bug 1567949 - Move resampler setup one layer down to be able to set the channel count per chunk. r=pehrsons
It is technically possible to have different channel count per chunk, so we
should reinitialize the resampler when that happens. This is a bit awkward
because the rates are on the track, but the data is in the chunks, and the track
cannot access the chunks, so we have to plumb the resampler state down.

Differential Revision: https://phabricator.services.mozilla.com/D53806

UltraBlame original commit: 1bed49fa4f66d08b02e4980d93ce3f07fd831ef6