Bug 1596479: teach `run-task` to get worker location from `TASKCLUSTER_WORKER_LOCATION` environment var r=tomprince
Previously we inspected the `TASKCLUSTER_WORKER_GROUP` environment variable,
which now only returns the cloud provider of the worker. This commit teaches
`run-task` to instead use the `TASKCLUSTER_WORKER_LOCATION` to gather
information about the location of the worker. We also use the extra data
about the cloud provider for the worker to construct a key for use in the
config, in the form `cloudprovider/region`, so GCP hgweb mirrors can be
amended to the `hgmointernal` config when they are ready.

While we're here we make the error handling for a missing environment
variable slightly nicer.

Differential Revision: https://phabricator.services.mozilla.com/D53209

UltraBlame original commit: efb782988830a6d815cd8c7e39947f9f94dc2efd