Bug 1592155: Trace Debugger.Frames for suspended generator/async calls directly. r=jonco
For collection purposes, there must be an owning edge from an
`AbstractGeneratorObject` to each `Debugger.Frame` with hooks set that refers to
it. If a `Debugger.Frame` refers to a suspended generator or async call, and the
`Debugger.Frame` has `onStep` or `onPop` hooks set, and the call's
`AbstractGeneratorObject` is reachable, the garbage collector must retain the
`Debugger.Frame`: collecting it would cause the hooks not to fire, which would
be visible to JavaScript.

At the moment, we obtain the effect of that owning edge in a needlessly
circuitous way. The iterative marking loop in `GCRuntime::markWeakReferences`
calls `DebugAPI::markIteratively` to scan all `Debugger` objects with live
debuggee globals to see if they have any `Debugger.Frame`s for marked
`AbstractGeneratorObject`s. If they do, then we mark the `Debugger`, which marks
the `Debugger.Frames` and causes `markIteratively` to return true, so we go
around the `markWeakReferences` loop again.

With this patch, tracing an `AbstractGeneratorObject` in a debuggee realm checks
the `Debugger`s weak maps for `Debugger.Frames`, and marks them. The edge is
still only inferred (via the Realm's isDebuggee flag; its list of Debuggers; and
the Debuggers' tables of generator frames) rather than being represented
directly as a pointer, but overall, this is much closer to ordinary tracing
behavior: tracing one object causes more objects to become markable.

Differential Revision: https://phabricator.services.mozilla.com/D53078

UltraBlame original commit: 9f5acf34610fefe7f9d1065ae263dcff3b7595b6