Bug 1598674 - Don't assume sync dispatch of iframe load event in tests, r=bzbarsky
Blink and WebKit dispatch the load event of the iframe synchronously,
whereas Gecko and (currently) the spec assume it's async. This causes
a hang in some tests using this helper library because in Gecko a
subsequent load ends up running in the event dispatch of the initial
load event, and so the load event is suppressed and the tests are
unable to complete. In other browsers the event is not suppressed and
so the tests run.

Avoid this by ensuring that the event loop always spins after the
iframe load.

Differential Revision: https://phabricator.services.mozilla.com/D54343

UltraBlame original commit: 2b978939b653b8531e27dc20fa12587db6f3338a