Bug 1597086 - Limit size of the search bar panel. r=Standard8
Wrapping one-off buttons into a container confuses layout that lets the richlist
box to grow unconditionally. This problem is already partially visible by
opening the searchbar continuously, sometimes strings are not properly overflowing
with ellipsis, they are just overflowing.
Setting a width on the panel limits its size properly, though this won't fix the
layout sometimes not properly cutting the richlistbox. It's likely there an
underlying bug in the widget or calculations when widgets are not visible.

This also correctly forces showing at least 3 one-off buttons per row, the
existing code is not sufficient for that because search-one-offs.js sets a
max-width that is smaller than 3 buttons.

Similarly, it corrects a calculation in the urlbar one-offs, where we should
show at least 4 buttons, but due to a not considered margin we only show 3.

Finally, this removes a few rules that are not necessary with the current design.

Differential Revision: https://phabricator.services.mozilla.com/D53968

UltraBlame original commit: f912ca0ab78b79db21c9c6c1d4335fd406b86e34