Bug 1595564 - Dispatch TestRendered event for wpt reftests with reftest-wait, r=maja_zf,mattwoodrow,marionette-reviewers
The TestRendered event is dispatched for reftests containing a
reftest-wait class on the root at the time the test would complete if
there wasn't such a class. This occurs after the initial paint is
complete, so any dynamic changes can be performed in such a way that
they are not batched with the initial layout/paint.

Differential Revision: https://phabricator.services.mozilla.com/D52684

UltraBlame original commit: 033df21b3afec1410d9e8190b5eaa1637e60537e