Bug 1598516 - Handle malformed uri load error with an error page. r=kmag
Previously this error occurred synchronously during AsyncOpen, and we handled it there.
With DocumentChannel we don't find out until it's handled in the parent, so the error is returned to the docshell via a failed status on the request.

Differential Revision: https://phabricator.services.mozilla.com/D54248

UltraBlame original commit: 72f2256ae171e64398efccf910a6d1b7d202ec34