Bug 1599117 part 2 - Restore empty-block + improveTypesForTest behavior for test instructions. r=tcampbell
When an MTest instruction results in a triangle structure, we now create an
empty block (like we did before bug 1595476) to get a diamond structure.
We did this for short-circuit operators but not for other cases.

We now also call improveTypesForTest for this block.

Differential Revision: https://phabricator.services.mozilla.com/D54535

UltraBlame original commit: 2c25a18171ba0a181781d6395b2e6443e26ac30e