Bug 1536523 - Support performance.getEntries result in console.table. r=Honza.
PerformanceEntry values are put in the `getterValue` property in
the descriptor, so whenever we want to display a table we need
to check if the value could be in there.
This highlighted an issue in the console layout when there is a
large number of cells, which we fix in this patch.

Differential Revision: https://phabricator.services.mozilla.com/D54261

UltraBlame original commit: 1680a0b08a7602cd06228d8e484de2233c7a5612