Bug 1599324 - Disable layout.css.zoom-transform-hack.enabled for now. r=boris
People do like to use zoom along with transforms for legit reasons, which means
that we get this wrong.

This is unfortunate, as the whole point of the hack was fixing sites that only
used zoom without regressing sites that would do:

  zoom: 0.5;
  -moz-transform: scale(0.5);
  transform-origin: 0 0;

So we're a bit stuck here. The only way to deal with the known regressions is
parsing `zoom: 1` as invalid, which is insane, and it would probably cause other
compat issues... For now disable the pref.

Differential Revision: https://phabricator.services.mozilla.com/D54685

UltraBlame original commit: 9c10d8e6c8a4f414c8ae8d5a792b4a7709a80bc7