Bug 1591874 - Move VariablesView files to devtools/client/storage. r=miker.
Since the storage inspector is the last consumer of the VariablesView,
it makes sense to move the component directly in the storage inspector
folder.
Since it can't have a controller, the bits where we were checking this
are removed.

Differential Revision: https://phabricator.services.mozilla.com/D54734

UltraBlame original commit: 0b5282256e0094eeae41d6c6377c6b37893cbf06