Bug 1599248 - Use local raster space when animating a double tap zoom. r=botond
In bug 1531142 we made it so that when a spatial node is being pinch-zoomed we
use a local raster-space to avoid rerasterizing glyphs for every slight change
in zoom level. This makes it so that we also apply the same trick when
being asynchronously zoomed by a double-tap gesture.

Differential Revision: https://phabricator.services.mozilla.com/D55699

UltraBlame original commit: 37582d3167c5f6a838039162d9e376baf165b495