Bug 1600488 - Allow FinalizationGroup targets to be unregistered after they have been queued for cleanup, in accordance with the spec r=sfink
The fix is to queue pointers to finalization records rather than just holdings when the target dies.  These are still in the resgistration map and so can be cleared by unregister().  We detect this in the iterator's next() method and skip any such records.

Differential Revision: https://phabricator.services.mozilla.com/D55808

UltraBlame original commit: 37b5be84c26cea97feb86784f8f0ca9ca6608055