Bug 1601775. Stop needlessly calling PMPrintSettingsSetJobName in nsPrintDialogServiceX::Show. r=mstange
We want the print job title set when actually submitting the job to a printer
so that the job appears with a useful title in any print job management
software the user may be using. However, we do that in
nsDeviceContextSpecX::BeginDocument, which is the place that it needs to
happen.

This removed code is called much earlier, while setting up to open the print
dialog and using a temporary nsPrintSettingsX in order to configure the
dialog. The dialog does not make use of the print job name (and we discard
the nsPrintSettingsX before going on to do the print using BeginDocument).

This removes the last piece of code that requires us to pass a
nsIWebBrowserPrint to the print dialog opening code. Removing that clears the
way to simplify the code and get printing working with Fission enabled.

Differential Revision: https://phabricator.services.mozilla.com/D56087

UltraBlame original commit: 7f8a7ac4bfe49199d34e8581d1fbbce4e25b9650