Bug 1427141 - Request a break before the row-group if its first row doesn't fit.  r=TYLin
FTR, the origin of the Reset() call in the else-branch I'm removing
is bug 397428.  I think this is before we had a break-before status,
or at least before our table code knew how to deal with it, so it
might have been correct to change an INCOMPLETE status into COMPLETE
in those days.  It seems the intent was to push the row-group:
https://bugzilla.mozilla.org/show_bug.cgi?id=397428#c14
"... and they all set aStatus to NS_FRAME_COMPLETE, which is logical
because we want to push this entire rowgroup to a new page and try
to reflow the row again..."

Nowadays though, returning a break-before status is the correct
way to do that.

Differential Revision: https://phabricator.services.mozilla.com/D56157

UltraBlame original commit: 111c247a2bca158383d6c142cc6c2209456aa637