Bug 1601992 - Set write timeout on udp socket; r=ng
We don't currently set a write timeout on the udp socket which could cause
write calls to block indefinitely. It is possible that this is blocking
long enough to cause the shutdown hangs seen in Bug 1601992.

This also bumps the number of times we retry failed queries from 2 to 3 to
account for the increased likelihood of not sending a query or answer.

Differential Revision: https://phabricator.services.mozilla.com/D56536

UltraBlame original commit: a336ad33f55374536c61865e1a04ae022e0ef8b2