Bug 1602972 - Keep FxR PC window on top r=Gijs
This change adds the 'alwaysontop' features to the window created for FxR on PC. This is done to address issues where scrolling and entering fullscreen require the window to either be visible or have focus. While this does not fully solve the problem, it significantly reduces its likelihood for a small cost.

Differential Revision: https://phabricator.services.mozilla.com/D56648

UltraBlame original commit: 8732ab855b2e234e35b85d609f3232b8291f07db