Bug 1596166 - Remove "height: 30px" on RDM toolbar r=bgrins
The UA input field gets caught off when the browser viewport is too small. This is because the RDM toolbar has an explicit height of 30px set on it, so it can't accomodate the second row created for the UA input.

Differential Revision: https://phabricator.services.mozilla.com/D56783

UltraBlame original commit: 2b7a473e92d31da7290305aa452421cfbb6a4f56