Bug 1603396 - Do not test XR permissions in browser_temporary_permissions_expiry.js when dom.vr.enabled is disabled r=pbz
dom.vr.enabled is enabled by default on all platforms in Nightly, but only enabled for Windows and macOS in Beta and Release. This results in the test failure after the merge to Beta.

To fix this test, we can skip the VR/XR related portion when the dom.vr.enabled pref is not set.

Differential Revision: https://phabricator.services.mozilla.com/D57044

UltraBlame original commit: 65ab90c240ecdf6585f2eda8b5edeff14a970756