Bug 1603726 part 1 - Switch back to main repo for wasm spec-tests. r=lth
This commit updates the spec tests to be based off the main repo with no
proposals.

Two commits are grafted on the branch. The first is a pre-existing commit to
drop a test that isn't valid with a bottom type. The second is a new commit to
allow segment initialization failures to be reported as a runtime or link error.

[1] https://github.com/eqrion/wasm-spec/tree/spidermonkey-tree-tests
[2] https://github.com/eqrion/wasm-spec/commit/f8044be9f7bc2b3c7add51251efc6276c6ce0ce4
[3] https://github.com/eqrion/wasm-spec/commit/b8836785dd82bdcb9fd0f75406fa7418ce7417d2

Differential Revision: https://phabricator.services.mozilla.com/D57335

UltraBlame original commit: 6ed5d68689f1c6a1f26f09902416553fa099ec6e