Bug 1600208 - mark test_bug756984 as permafail on ubuntu1804 r=jmaher
Changes:

This test is a permafail on ubuntu1804, and should be marked as such as the migration to ubunt1804 is actively underway with target date of end of December 2019.

With a fail-if annotation, the hope is that if this test suddenly begins passing we will be notified of it (as opposed to skipping it outright).

Differential Revision: https://phabricator.services.mozilla.com/D57381

UltraBlame original commit: 2d188d7909343bae73064921b8d4c44af2bb1b40