Bug 1603349 - Use try_recv rather than recv_timeout in mdns_service; r=ng
There are known bugs with recv_timeout which may explain the crashes we're
seeing in Bug 1603349. This patch switches to using try_recv which returns
immediately if no data is available. This thread already has timeouts set in
the UDP socket reads and writes, so the timeout with the receive channel
should not be necessary.

Differential Revision: https://phabricator.services.mozilla.com/D57448

UltraBlame original commit: 5b6ef512edd26985a9350961d6974ea6096fdda4