Bug 1405416 - Assertion failure: mDocument->GetReadyStateEnum() == nsIDocument::READYSTATE_INTERACTIVE (Bad readyState) [ txMozillaTextOutput::endDocument]. r=bzbarsky
We only set a transform observer when transforming through an XSLT PI (nsXMLContentSink::LoadXSLStyleSheet), not when the transform is initiated through the JS API. The only time we can assert that the readystate is READYSTATE_INTERACTIVE is when transforming to a fragment. Because transformToDocument doesn't set an observer, we can't check for a non-null observer but need a state that both XSLT PI transforms and transformToDocument set (which we can do when they create a document).

Differential Revision: https://phabricator.services.mozilla.com/D56860

UltraBlame original commit: e49dc46498238e04c0571a06f3e95574e1a0ddf5