Bug 1603714: Do not treat third-party module data integrity error as IPC error; r=mhowell
I still do not yet have a good grasp of the root cause of this data integrity
issue, however I also think that having the browser process crash because of it
is far worse.

This also adds a check in our telemetry processing that the deserialized event
is valid.

We'll probably need to add more diagnostic information in the future, but I'd
like to get rid of the IPC errors in Nightly in time for holidays.

Depends on D58046

Differential Revision: https://phabricator.services.mozilla.com/D58047

UltraBlame original commit: c3a21a4b5a45d865f7d717191cc3b6ce3398e303