bug 1603930: toolkit: fix remote agent build targets; r=firefox-build-system-reviewers,mshal
The remote agent component was meant to only be built on the
Nightly release channel, but 868b1bf043dc accidentally reversed
--enable-cdp and --reverse-cdp, not understanding that there's a
semantic difference in the build config parser for this.

This has had the undesired effect that the remote agent has been
built on all release channels since around 25th April of this year.
This in itself is not a huge concern because the feature is disabled
through the remote.enabled preference in modules/libpref/init/all.js,
except on Nightly.

This patch reverses s/disable/enable/ which causes ENABLE_REMOTE_AGENT
to be set only on the Nightly release channel, or when --enable-cdp
is explicitly used, thus achieving the original, intended effect.

We caught this when running the bc test
remote/test/browser/browser_agent.js as part of beta simulations
for Firefox 73.  I wrote that test explicitly to test for this
eventuality, so I will consider it good enough for this change to
ship without additional tests.

Differential Revision: https://phabricator.services.mozilla.com/D58452

UltraBlame original commit: 405d6889de7cf4df42cff2163c992a534c39d3f3