Bug 1605747 - Fix ErrorRep for Error with undefined name property. r=Honza.
Since the name can be an undefined grip, there
could be case where the name would be an object
and not a React element, making the rendering fail.
This patch fixes that and adds a test to ensure
this does not regress.

Differential Revision: https://phabricator.services.mozilla.com/D58747

UltraBlame original commit: 1b223c4783bb64d595886b074d1d955011cf24bd