Bug 1606124 - mozbuild/util.py supports Python 3 r=mars
Many of the utility classes very heavily rely on multiple inheritance which is broadly unncessary and made debugging difficult, so I ripped it out wherever it was necessary. The indented_repr() function is also more or less a re-implementation of pprint, so to support Python 3 I've instead made it a small wrapper around pprint. (This results in the output of the function being slightly different than we're used to, as pprint does sorting/indentation/line breaking differently than indented_repr() currently does, but it'll be nice to not have to maintain this code that is slightly different than pprint for no real reason.)

Differential Revision: https://phabricator.services.mozilla.com/D58306

UltraBlame original commit: e610cb1d72aa940c0bcac6b0c993712fb2349025