Bug 1599043 - Part 4: Notify the OnContentBlockingEvent in the parent process in UrlClassifierCommon. r=dimi,Ehsan
We make the OnContentBlockingEvent to be notified in the parent procees
for UrlClassifierCommon. There are two place would trigger this,
UrlClassifierCommon::SetBlockedContent and
UrlClassifierCommon::AnnotateChannel. But we still send to the child
process to notify the content blocking since we need to update the log
in the content process in this stage.

Differential Revision: https://phabricator.services.mozilla.com/D55647

UltraBlame original commit: 3de15e4fce42094f7e07a18725106e48d47b5597