Bug 1607703 - Initialize the JS context if -reset-profile is set and -migration is not. r=froydnj
ProfileResetCleanup ends up opening a browser window and doing browser
things, so it needs the JS context to be initialized. Migration does
that, so if you do both then things are fine. But if you just pass in
-reset-profile and not -migration then you end up crashing
immediately. This patch just checks for that case and initializes the
context.

Differential Revision: https://phabricator.services.mozilla.com/D59810

UltraBlame original commit: 92c0e70bf4ec3575a623058de6c51b23a28791b2