Bug 1412401: accessible/tests/mochitest/events/docload/test_docload_root.html: Wait for focus event after opening the dialog. r=MarcoZ
Previously, when opening the dialog, we only waited for a reorder event.
This could mean we programmatically closed the dialog before it was ever focused.
That in turn would mean that focus would never be fired on the previous window when the dialog was closed, causing the close dialog test to time out.

Differential Revision: https://phabricator.services.mozilla.com/D59826

UltraBlame original commit: b9f97507c65183b7429147f3d7274d44e4821c75