Bug 1608509 - Part 1: Fix a few odd comments. r=jandem.
The comment about JSOP_GLOBALTHIS and JSOP_FUNCTIONTHIS may never have been
sensible; JSOP_FUNCTIONTHIS never walks the environment chain, as it is used
only in non-arrow functions, which take `this` as an argument in the CallArgs.

JSOP_LOOPENTRY and JSOP_TRACE are gone. So is JSOP_SPREAD.

"JSOP_POPs" was eccentric in a way that would be unclear with CamelCase.

Differential Revision: https://phabricator.services.mozilla.com/D59799

UltraBlame original commit: 725809d9e561a55d27b8f2b10360a3d00a0cd6a8