Bug 1610165 - Failing to insert an SVG into the surface cache continuously repaints the image. r=tnikkel
I was poking at the missing images bug by dropping cache sizes and such and
found that we always ended up repainting some SVGs over and over. This seems to
be the cause. When we fail to insert into the cache we still end up notifying
over and over (because we forget to say aWillCache=false), so we'd end up
repainting, and failing again, etc.

This is easy to see adding some logging to
nsDisplayBackgroundImage::PaintInternal.

Any ideas as to how to add a test for it?

Differential Revision: https://phabricator.services.mozilla.com/D60372

UltraBlame original commit: 206cec28723abd20274126812c861e16f9f683d5