Bug 1610192: Remove unneeded array-index in ObjLiteral array mode. r=mgaudet
The ObjLiteral format allows for names (as atom indices) or numeric
indices to be attached to each property in the literal. The numeric case
is limited to 23 bits by the packed format in memory.

The object-literal case ({}-literal with some numeric indices) already
checks for in-range values before deciding to use the ObjLiteral
infrastructure, but the array case did not check this.

Fortunately, the indices in the array case were only used as a sanity
check for debug-mode assertions, and never for any actual indexing
operation, so no out-of-bounds behavior could result, even before this
fix. However, the sanity-check assert will fire at
object-construction-from-ObjLiteral time in cases where the ObjLiteral
mode should not have been used, because no range check was done at
ObjLiteral-construction time.

This change removes the use of indices for array-mode ObjLiterals
entirely.

Differential Revision: https://phabricator.services.mozilla.com/D60474

UltraBlame original commit: 5ea36d82a283c479423faaabf0c7f2eb622167ed