Bug 1608286 - Crash in [ mozilla::dom::MediaKeySystemAccessManager::CheckDoesWindowSupportProtectedMedia] r=bryce,bzbarsky
This changes fixes a failfast in a call to Navigator.requestMediaKeySystemAccess for the case where there is no browser available from a window when e10s is on. This can happen when a document is disconnected from the DOM.
The fix is to reject the promise in this case.

Differential Revision: https://phabricator.services.mozilla.com/D60022

UltraBlame original commit: b71a801eabd42db42bec3332767b59ab6cb5e714