Bug 1609711 - Compute gradient positions in css pixels. r=jrmuizel
The gradient code is the only one that does a really weird thing with
LengthPercentage values, by getting the percentage and length separately and
turning the length into a percentage relative to the line length (which is in
device pixels).

This won't work once we have min() / max() / etc. in CSS (as we can't access
the length and percentage components separately, as which one you choose may
depend on the percentage basis). So instead of that, use the regular
ResolveToCssPixels there are lengths involved.

We change a bit the surrounding code to work in CSS pixels, so as to avoid
unneeded CSS -> device pixel conversions.

Differential Revision: https://phabricator.services.mozilla.com/D60159

UltraBlame original commit: 627ee9bf55737123a114b3494b769d9eeb0e04d2