Bug 1610269 [wpt PR 21259] - [LayoutNG] Remove IsIntermediateLayout flag., a=testonly
Automatic update from web-platform-tests
[LayoutNG] Remove IsIntermediateLayout flag.

Previously if we synthesized the min/max sizes by performing a layout
on the orthogonal subtree - we could end up in a situation where the
"DCHECK(!constraint_space.IsIntermediateLayout() || box_->NeedsLayout())"
was hit.

This was caused by a legacy layout fieldset needing a preferred widths
recalc.

This is an indirect fix for the below DCHECK bug. Removing the
IsIntermediateLayout flags means that we now perform a copy back when
we are in this (relatively rare) phase.

The IsIntermediateLayout was initially introduced as a performance
optimization - however it could also hurt performance if we perform
a correct layout of a subtree with a fixed size.
(we'd need to layout that sub-tree twice, getting exactly the same
layout result).

Bug: 1039241
Change-Id: I468a9c8e26b5e7855a72ddb2b1f2614538cedd21
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2007422
Reviewed-by: Morten Stenshorne <mstenshochromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrickchromium.org>
Cr-Commit-Position: refs/heads/master{#733344}

--

wpt-commits: 676c9d957b75188553d3dfd75b7c19856561e773
wpt-pr: 21259

UltraBlame original commit: 4740263d1dc2a32aa2b41671624621156236fa38