Bug 1612025 - Avoid conflict with pread64/pwrite64 defined by libc r=gsvelto
musl libc defines these to pread and pwrite respectively, since its
off_t is always 64-bit. This causes unintended macro replacement
when defining the pread64/pwrite64 wrappers.

To prevent this, save the macro definitions at the start of the
header and restore them at the end. This is what is already done
to work around the same issue on Android NDK.

This is a cherry-pick of the following upstream change:
https://chromium.googlesource.com/linux-syscall-support/+/f70e2f1641e280e777edfdad7f73a2cfa38139c7%5E%21/

This change was also applied to google-breakpad here:
https://chromium.googlesource.com/breakpad/breakpad/+/647158232ba153d33c984f201b0ee35384e684cc%5E%21/

Differential Revision: https://phabricator.services.mozilla.com/D61205

UltraBlame original commit: 660da1ec99c0b7407bd6683a1504681e3e959f45