Bug 1606901 - P1. Don't assume that nsIAsyncInputStream are always non-blocking. r=baku
From the documentation:
" While this interface is almost exclusively used with non-blocking streams, it is not necessary that nsIInputStream::isNonBlocking return true"

So we can't assume that a nsIAsyncInputStream is always non-blocking. Handle the case where it is.

Differential Revision: https://phabricator.services.mozilla.com/D61368

UltraBlame original commit: 57f2c7a63fcf6999e1c25b4c02f239ae08328b67