Bug 1613430 - Go back to LibLoadingStrategy::ReadAhead on Nightly r=glandium
The original issue in bug 1565902 which prompted this investigation
seems to have gone away with a Windows OS update. We saw mixed results
in telemetry over this anyway, and no clear and obvious path for
getting the benefits without the drawbacks, so removing this seems
to be the best solution right now.

Differential Revision: https://phabricator.services.mozilla.com/D61720

UltraBlame original commit: f0cb6e48329b48f141f5c75edbcb38ba20228414