Bug 1613679 - Fix a misleading-indentation error in mach_override.c r=froydnj
Every day clang's warnings continue to amaze me. As does our ability to trigger them.

I'm assuming that this line was meant to go within the scope of the `if`, because if it were outside the scope, then we'd be duplicating the previous `fprintf`.

Differential Revision: https://phabricator.services.mozilla.com/D61835

UltraBlame original commit: b372743705c9c955e3f5cc08a1aa4cf09b6a0c2d