Bug 1541095: Remove an unnecessary static_cast in WritingModes.h. r=TYLin
The variables involved in the bitwise arithmetic are all of type uint8_t, so
we end up with that type automatically (no need to bother with a static_cast).

Note that we *do* need to declare the type (we can't use `auto`, or else
the compiler wants to upgrade the local variable to have a wider type --
"int", I think -- and that then produces an build warning when we pass
that variable as a uint8_t param further down).

Differential Revision: https://phabricator.services.mozilla.com/D61893

UltraBlame original commit: f1f26a8866c1cab9d332d7106aec326a892753c7