Bug 1612483 - Remove mozilla::dom::TabContext::IsMozBrowser(). r=bzbarsky
Now, this method returns simply the result of of `IsMozBrowserElement()`.

In the old days, We used to have IsMozBrowserElement and IsMozBrowserOrApp,
where the latter was true if we had a mozapp or mozbrowser.
But with b2g removed, the app thing went away.

`IsMozBrowser()` is less used than `IsMozBrowserElement()`.

I think the former should be replaced by the later one.

Differential Revision: https://phabricator.services.mozilla.com/D61699

UltraBlame original commit: 9246a8ba73d89f9c6e1fa842c555fc576755716c