Bug 1613527 - Stop using mUpdateNestLevel for MaybeInitializeFinalizeFrameLoaders, r=smaug
This has been tightly coupled with script blockers, which are already being
used, for a while. This simplifies the situation, and avoids issues which can
occur when the script blocker for a document update has yet to be removed, but a
scriptrunner has.

Differential Revision: https://phabricator.services.mozilla.com/D62462

UltraBlame original commit: fecefca18fc617da2e1b4e8d834d145c6620f240