Bug 1604140 - Part 1: Fix focus trap within shadow DOM when host is scrollable; r=smaug
The checks for `*TopLevelScopeOwner` are to skip the scope that we have already checked.
But when the shadow host is scrollable, we will traverse anonymous children for the scroll frame first in frame traversal and `oldTopLevelScopeOwner` will be reset.
Then we don't realize that we have already checked the host's scope.

Differential Revision: https://phabricator.services.mozilla.com/D60923

UltraBlame original commit: 8897a2187a9537310df8b283f65844fe46ad075d