Bug 1613047 - Fix CSS Warnings in a Fission world when navigating. r=ochameau.
If the CSS filter was enabled, when the user would navigate to a page
on a different origin, the CSS Warnings from the new page wouldn't
be displayed in the console.
This is related to how we manage the CSS Warnings. Since emitting those
messages is costly, we only do so when the console is opened, if
the user already set the filter, or when they turned it on.
The issue is that it was only done on the main target, and only
when the console would start, or when the user clicked on the css
filter button.
So with Fission enabled, we could switch to a new target, but we
wouldn't trigger the code that parses the stylesheets of the new
page.

The browser_webconsole_message_categories was asserting this issue,
and is now fixed (after setting the proper target switching target).

Differential Revision: https://phabricator.services.mozilla.com/D61558

UltraBlame original commit: de7ad7ae633913be9b152cc27486c30a65204312