Bug 1610899 - avoid text track list being cut off, r=jaws
The text track list should not exceed the size of the video controls. Because it is positioned
absolutely, and because the height of the video is not known in the CSS, the only way to
ensure the list's height does not exceed the video's height is to set a 'top' value as well as
a 'bottom' one.

Unfortunately, that then means that if the list is quite small and the video quite tall, the
popup becomes way too large.

To avoid this, I've added another level of nesting, and used flexbox to align the actual list
to the bottom of the container for the list.

Differential Revision: https://phabricator.services.mozilla.com/D62452

UltraBlame original commit: ea48938d4cf677c4e00e093c2c92d7da5b52b0eb