Bug 1378189 - Part 4: Support JSOp::BuiltinProto in Ion. r=jandem
This operation is similar to JSOp::GetIntrinsic, where we can bake in a constant
value for known prototypes. In the future we may want to assign a type-set to
this operation, similar to JSOp::GetIntrinsic, but for now we don't need more
exact type information here.

Differential Revision: https://phabricator.services.mozilla.com/D58779

UltraBlame original commit: be0da60bc0a377ec1ddc3811bea166df977dae50