Bug 1612620 - Fixing dashed/dotted border issue Android devices with Adreno. r=jnicol
On GLES, the default shader behaviour is to use highp(32bit) integers
in the vertex shader and mediump(16bit) integers in the fragment shader. This
causes issues in the border shader due to bitshifting with 16 bits. The
fix here is to only shift by 8 bits as the data can be represented in a
16bit integer and forcing mediump in the vertex shader as well.

Differential Revision: https://phabricator.services.mozilla.com/D62784

UltraBlame original commit: a2caf09a5e7cadd78e5a12c4d715d3f0c996e48c