Bug 1616000 - Fix broken push notifications r=handyman
Bug 1604412 enabled "remote backbuffer" compositing when using non-accelerated
layers. However, due to my misunderstanding of how nWindow and compositing
handles layered windows, I broke the usecase where a child window is
transparent.

In this case, it is the top-level anscestor window that has the
WS_EX_LAYERED attribute and not the child window. The ancestor is the one
that needs to be repainted when a present is requested.

Differential Revision: https://phabricator.services.mozilla.com/D63377

UltraBlame original commit: d69f53c5968cf816bf07c5664b274863c155d19f