Bug 1616332 - Add all fenix pageload tests to raptor-browsertime. r=perftest-reviewers,ahal,AlexandruIonescu
This patch adds all Fenix pageload tests to raptor-browsertime. To handle how the task is specified (specifying run-on-projects by raptor-test) the raptor.py taskgraph transforms had to be modified.

The transform modifications change a couple things. First, the `raptor-subtests` is now the first thing to have it's keyed-by entries handled because it's values are used as keyed-by entries on other fields such as run-on-projects. Once it's values are resolved, we immediately split/clone the task into one per subtest entry (adding a `chunk-number` field that is used for more setup in the next transforms). After this, the rest of the fields have their keyed-by entries resolved.

Differential Revision: https://phabricator.services.mozilla.com/D63222

UltraBlame original commit: 1ac5a1a3d1c7e1f8a3900c11c6c3147bcafed646