Bug 1615921 - Skip writing to pdf file if shutting down r=jwatt
We are intending to advance the toolkit.shutdown.lateWriteChecksStage
pref, to collect information (and crash on DEBUG) about writes which
happen during and after the xpcom-shutdown-threads notification. This
is producing failures in the PrintTargetPDF.cpp destructor because
we end up calling write_func and writing to (I presume) the target
PDF. This _feels_ like something we can just skip, so that's the
review request I am sending, but please let me know if it's critical
that we write to this file during shutdown.

Differential Revision: https://phabricator.services.mozilla.com/D63218

UltraBlame original commit: edf46a901a6bb4986c29d5a292f22d18728bdc32