Bug 1597378 - Remove old popup implementation; r=julienw
This commit removes the old popup implementation, but does not implement the new one
yet. This is in a follow-up commit. The client is in a bit of an awkard state where
it is mid-transition to a new architecture. Some of the code comments and documentation
were updated to reflect the changes coming up.

Differential Revision: https://phabricator.services.mozilla.com/D62913

UltraBlame original commit: 4d1ff8bd4e645f5ee35be6aaf81cc916e5cade11