Bug 1604726 - Use a ResizeObserver to responsively size the toolbar. r=bgrins
The RDM toolbar currently scrolls with the page when there is overflow in `.browserStack`. This is because the browser viewport can be set to a size larger than browser window. The solution here sets makes the toolbar `position: fixed` and utilizes the `ResizeObserver` to adjust the size of the device modal/selector to take up the size of the browser stack.

Differential Revision: https://phabricator.services.mozilla.com/D63245

UltraBlame original commit: 2a3597aa8d9e6fc158b02b93f26dd1355a955f47