Bug 1616675 - Add an IPv4 secure service to mochitests' ssltunnel r=keeler
WebAuthn needs to write a test to confirm it's prohibited when accessed via
an IP address. This adds the capability to get a SecureContext for an IP host.
It uses 127.0.0.2 so as to bypass restrictions on 127.0.0.1, and the use of .1
as a special-market in ssltunnel.

Differential Revision: https://phabricator.services.mozilla.com/D63570

UltraBlame original commit: f8632635a60ccfb5f702512f9d747621927594f7