Bug 1601940 - Don't use bitfield members in AsyncPanZoomController. r=kats
A write to a bitfield can legally be optimized to a write of the entire
machine word containing it, which makes ensuring thread safety more
complicated (we'd have to make sure the entire word forms a critical section
such that one bitfield in it can't be read while another is written).

Thread safety in APZ is hard enough as it, and AsyncPanZoomController has
a large sizeof() already, we can just take up a byte for each field.

Differential Revision: https://phabricator.services.mozilla.com/D64486

UltraBlame original commit: 9aaf2c037d8fabfd5b1395661d82ea901ba96eee