Bug 1596655 Keep DestinationNode track after offline completion for worklets r=padenot
This keeps the worklet/graph thread alive for control and MessagePort messages
until there are no references to main thread objects.

In test_getUserMedia_audioCapture, the track of the OscillatorNode being
recorded does not end, and, even if it did, MediaRecorder would not listen for
ended notification on the node's track but its pipe track.  Recording via
DestinationNode works around this because the DestinationNode track ends and
MediaRecorder listens to this track directly because it has
EXTERNAL_OUTPUT. Previously MediaRecorder would stop when the MediaTrackGraph
was shutdown, but this is no longer happening.

Depends on D65069

Differential Revision: https://phabricator.services.mozilla.com/D65070

UltraBlame original commit: 93abf903275b5f7fdab64f37fdce588bc780e63a