Bug 1616570 - Part 3: Change the CookieJarSettings::CreateBlockingAll to be a singleton and rename it to GetBlockingAll(). r=Ehsan
We don't need to create the blocking-all cookieJarSettings every time we
request it since it is basically the same. Instead, we should make it as
a singleton.

The patch makes this change the rename the
CookieJarSettings::CreateBlockingAll() to GetBlockingAll() which
reflects the fact that it is a singleton.

Differential Revision: https://phabricator.services.mozilla.com/D64564

UltraBlame original commit: 1bad5855c84fd691bff61f688470ef03c8222192