Bug 1620163 Fix perfdocs to not skip missing tests r=sparky,perftest-reviewers
There is a logic error in the code that is not catching missing tests. Also, the manifest path for each test needs to be added along with the test name so the logger can be able to print it.

Differential Revision: https://phabricator.services.mozilla.com/D65473

UltraBlame original commit: 17c8c9f21b69fb45ba3f14cf47473f93c56914fb