Bug 1592037 - part4 : update metadata to the event source. r=MeFisto94
Whenever the main controller updates its metadata, we should also notify the event source to update the metadata, because we don't want to show the wrong metadata on the virtual control interface.

In this patch, we add a new media event to notify this situation and create a new method `SetMediaMetadata()` on the `MediaControlKeysEventSource`. But we leave that function empty for now, and will have follow-up bugs to handle setting metadata on each different platforms.

Differential Revision: https://phabricator.services.mozilla.com/D64847

UltraBlame original commit: b27118f20ad52aed19adee1da21a3b89bc5db572