Bug 1615370 - Add fallback to ShellExecuteExW in nsMIMEInfoWin. r=aklotz
If a system uses a custom shell instead of Windows Explorer,
`ShellExecuteByExplorer` always fails because it relies on explorer.exe,
but `ShellExecute` stil works because it's in-proc implementation.

We added a fallback for a local file to address Bug 1602726.  This patch adds
a fallback for a uri.

Differential Revision: https://phabricator.services.mozilla.com/D63095

UltraBlame original commit: 08488953cff81039d248a949a1701dcd2d600f2c