Bug 1611817 - Fix race condition in DevTools shutdown and stopProfilerAndDiscardProfile; r=julienw
This race condition is really only exposed in automated testing, but it
was making the DevTools performance-new mochitests intermittent.

The race looks like this:

* stopProfilerAndDiscardProfiler() request
* perf actor process the request
* The gecko profiler sends an event notifying that the profiler stopped
* DevTools updates the UI upon receiving the event
* The test suite sees the UI update, and triggers a close of DevTools
* Error: Connection closed, pending request
* stopProfilerAndDiscardProfiler still hasn't sent the response yet

This patch fixes it by not throwing an error if the panel is already destroyed.

Differential Revision: https://phabricator.services.mozilla.com/D65150

UltraBlame original commit: 47f2ae8298a9195d7fd428aea55204a07ec6120d