Bug 1607474 - Add logging and expected results to puppeteer-test, r=remote-protocol-reviewers,maja_zf
Use mozlog to handle logging from the `mach puppeteer-test`
command. This makes the output compatible with treeherder and allows
generating error summaries that can be used by the push-health
feature.

Presently the integration is based on parsing the output from mocha
with regex since that's easy to implement without digging into the
many layers of js code too much. In time it might make sense to
replace this with a custom output formatter integrated into mocha
itself.

We also add the ability to store and use expected results, so that we
are able to detect regressions even when not all test are passing. The
format of expected results is Dict[String, List[String]], where the
keys are test names and the values are lists of statuses. If the list
has more than one value the remaining values are treated as possible
intermittent statuses.

The updated expected statues can be generated by passing the
--write-results flag to the mach command.

Differential Revision: https://phabricator.services.mozilla.com/D65700

UltraBlame original commit: 0cb0288d1fe08eb4d5216bc0a4531025dde313e5