Bug 1619371 [wpt PR 22044] - Make more robust the timeout check WebDriverProtocol.is_alive(), a=testonly
Automatic update from web-platform-tests
Make more robust the timeout check WebDriverProtocol.is_alive()

In 58988ab31a (#21889) the WebDriverProtocol.is_alive() check was modified
to use a 2 second timeout, but this only works when the connection with
the webdriver has been lost and a new one will be made for the is_alive()
check because calling socket.getdefaulttimeout() doesn't affect the ongoing
httplib connection with the WebDriver.

To fix this we introduce the timeout value as an optional parameter of the
WebDriver send_command() function. When timeout its specified, then the
socket timeout of the current httplib connection its modified just only
for this request (and restored after). This works both when the connection
with the WebDriver is lost (and a new one its made) as when the connection
it is working as expected.

--

wpt-commits: 5b7d10d8b299774f420ffb452d70ad9c7ead92c7
wpt-pr: 22044

UltraBlame original commit: 6ff1b4f0a7cf7e9bba96f61342e040d8a1343c1c