Bug 1615509: Move place where calculate the issues to the action. r=rcaliman
To avoid storing declarationBlocks in reducer.
The reason why I held the information is because had thought that want to update
as soon as possible when change the target browser.
However, as it seems that to update the target browser is not so frequently,
remove the information to reduce the memory.

Differential Revision: https://phabricator.services.mozilla.com/D64305

UltraBlame original commit: ee220b50d171ab9bf324e8e02dec3a9a29683507