Bug 1618198 part 14 - Implement environment chain initialization. r=iain
The previous patch introduced more jit-test failures because we didn't
initialize the environment chain yet (a good thing because it means we're
beginning to support non-trivial tests!).

WarpEnvironment is added to store information WarpBuilder needs to build the
environment chain. The goal is to make most of the decisions in WarpOracle so
that the WarpBuilder code can be as simple as possible.

Differential Revision: https://phabricator.services.mozilla.com/D66387

UltraBlame original commit: 8cd7ab297247251be3bb476b5db3c18bb0827a5e