Bug 1619803 - Unify JSScript and LazyScript instances r=jandem,mgaudet
Use a single GC allocation for the JSScript and the LazyScript of a
function. This reuses the script during delazification / relazification
transitions.

The debugger and JITs are simplified as a result of having a single identity
for a BaseScript rather than tracking both a lazy and non-lazy form.

Differential Revision: https://phabricator.services.mozilla.com/D65226

UltraBlame original commit: 4dd9206018a378f9412b057d00646a10379060ee