Bug 1621076 - Fix intermittents caused by recording state race; r=canaltinova
I originally wrote the tests to check for the "not-yet-known" recording state. This
check is a race that did not fail locally, but does when run on the try servers.
This patch makes the tests a bit more permissive. It still includes the check for
"not-yet-known", but also allows for the recording state to already be known
at the beginning of the test.

Differential Revision: https://phabricator.services.mozilla.com/D66609

UltraBlame original commit: 51cd675065e68ef8a02bd68aae77e90a532b0bed