Bug 1619692 [wpt PR 22066] - Enforce mime-type check when loading Origin Policy manifest., a=testonly
Automatic update from web-platform-tests
Enforce mime-type check when loading Origin Policy manifest.

This CL modifies OriginPolicyFetcher to enforce that Origin Policy
manifests load with the mimetype

application/originpolicy+json

If an incorrect mime type is received, loading is aborted, and
the OriginPolicy object is returned with an empty policy and
status kNoPolicyApplies. Since this same pathway is hit if the
manifest load returns 404, we can also set kNoPolicyApplies in that
case as well, and a test is included for that.

The test includes a valid CSP which, if applied, would lead the test
to fail as eval() would not succeed. If the manifest is rejected due
to its incorrect mimetype, then the test passes.

Bug: 1051169
Change-Id: I59e5bde20c8a21a4fae8dbc3c6f58f34cb292195
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2080696
Reviewed-by: Daniel Vogelheim <vogelheimchromium.org>
Reviewed-by: Domenic Denicola <domenicchromium.org>
Commit-Queue: James MacLean <wjmacleanchromium.org>
Cr-Commit-Position: refs/heads/master{#748364}

--

wpt-commits: c9cc069222fdbfec600129f485b456ddc965a1b7
wpt-pr: 22066

UltraBlame original commit: 22cdeeabc792e0181b7020073cc07557ae88eae0