Bug 1617427 - Prefer using the root frame's displayport to the root composition bounds. r=tnikkel
If we can get the root frame's displayport, then we should use that
rect instead of the root composition bounds when clipping the scrollframe's
displayport. That way if APZ is disabled on the scrollframe, but the root
frame scrolls to bring a part of it into view, it will be fully painted and
not perma-checkerboard-y.

Note that this patch is the main fix, but leaves a bunch of comments/variables
with bad names; the next patch cleans that up.

Differential Revision: https://phabricator.services.mozilla.com/D66421

UltraBlame original commit: 275f9a696ae18baf32eea13303cc7afe393f4edf