Bug 1617427 - Resurrect IsCurrentlyCheckerboarding and make it more correct. r=botond
This improves the implementation of IsCurrentlyCheckerboarding (which is not
invoked from anywhere prior to this patch) so that it takes into account the
recursive clipping applied by ancestor layers' composition bounds. In other
words, the visible rect for a layer may be additionally clipped because
ancestor scrollframes have scrolled, and this patch accounts for that.

It also records the currently-checkerboarding state into the APZTestData
at the time that the compositor APZTestData instance is fetched.

Differential Revision: https://phabricator.services.mozilla.com/D66427

UltraBlame original commit: 35d9d70da6c9058c3498b7004089ad70674eb0b5