Bug 1620020: Fix Zone implementation r=mgaudet
For obvious reasons, irregexp does not contain any calls to EnsureBallast. We therefore need to make our LifoAlloc allocation fallible so that it can allocate a new chunk if necessary.

Also, we want to use the current size of the LifoAlloc, not the peak size, to decide whether we've allocated too much memory. Nobody was using the old ComputedSizeOfExcludingThis, so I rewrote it to use the value we're already tracking.

Depends on D66100

Differential Revision: https://phabricator.services.mozilla.com/D66101

UltraBlame original commit: ded485063dd6bfd1be0b070d0abe42a9e52f1759