Bug 1617959: Switch to the babel-eslint parser and turn on support for optional chaining and nullish coalescing operator syntaxes. r=Standard8
The babel-eslint parser defaults to assuming all scripts are modules. Although
we're moving in that direction it seems reasonable to set the default as regular
scripts for now. All the places that were previously overriding the parser are
already specifying the sourceType in their eslint configs.

I chose to put the babel config in a file that babel itself won't look to avoid
the risk of this config impacting the use of babel elsewhere in the tree.

Differential Revision: https://phabricator.services.mozilla.com/D66127

UltraBlame original commit: 40bfe22bad4cd7d8c86a48ce8514f637f795a278