Bug 1621718 - Use proper email parsing library for parsing patches. r=ahal
This uses the `email` module to parse the .patch file that Github generates,
so that it properly decodes encoded-words in the headers. Also using this
module is better with python3, so this patch also takes the command off the
python2 whitelist and makes it python3-compatible.

Differential Revision: https://phabricator.services.mozilla.com/D66621

UltraBlame original commit: 7b12fa322e5d953f63ca3fea3444e7a18b6cb7b1