Bug 1618198 part 16 - Support calls. r=iain
For now this is the slowest possible implementation.

It reuses the CallInfo infrastructure. That class has some methods we probably
shouldn't use in WarpBuilder, but the class itself is still useful and we can
clean it up more when IonBuilder is gone.

This uncovered an issue with MCheckOverRecursed: IonBuilder gives this
instruction a resume point (it's not clear why) but WarpBuilder doesn't, so
DCE was eliminating it and some jit-test crashed with overrecursion.
Marking the instruction as guard fixes this.

Differential Revision: https://phabricator.services.mozilla.com/D66739

UltraBlame original commit: 45be0faa08c025126f9e979a275e25e38486fc9f