Bug 1622757 - Explicitly release the helper thread lock rather than rely on this happening when we run the first task on the main thread r=sfink
Yoshi pointed out that we don't always unlock the helper thread mutex when using AutoRunParallelWork.  The code will work anyway, but it looks strange and perhaps we should not rely on this.  This patch adds an unlock in the one place it's currently missing.

Differential Revision: https://phabricator.services.mozilla.com/D66964

UltraBlame original commit: 2b586baa631e059ce770753f7f9599034ddb9a55