Bug 1620986 - Introduce run-time switch for multi-value.  r=bbouvier,wingo
This adds a JS shell command line argument to disable multi-value (if
it is enabled at compile time) and an about:config option to do the
same in the browser.

At this time, multi-values are not enabled in the browser, so the
about:config option is not visible.

Differential Revision: https://phabricator.services.mozilla.com/D67465

UltraBlame original commit: 531bdd018faa82d1fead2acabd9528594811b2f8