Bug 1623476 - Fix event targeting in a click_events_on_input.html. r=jgraham
Presumably, the intention of the test is to hit the middle of the target. To
this end, the test specifies an offset of half the target's dimensions
relative to the origin of the target.

However, when pointerMove() is given an element as origin, it already uses
the element's center as the origin point, so we end up actually targeting
the bottom right pixel which is prone to rounding issues.

Differential Revision: https://phabricator.services.mozilla.com/D67518

UltraBlame original commit: 0cd57f00ac5d1aada0c5ae8374c5a25ea543332d