Bug 1612378 - Part 9: Make StorageAccess::InternalStorageAllowedCheck() to aslo checks the document's URI when doing a about URI check. r=dimi,baku
So far, we haven't checked the window's document uri when doing a about
URI check in StorageAccess::InternalStorageAllowedCheck(). This brought
an issue that if the checked page is an about page with the system
principal, for example the 'about:devtools-toolbox' we will miss this
check. Because of the system principal doesn't have a URI. So we need to
check the document URI instead to know if this is a about page.

Differential Revision: https://phabricator.services.mozilla.com/D66481

UltraBlame original commit: 50d282ea77f1381e5bfa7f1297848648084239a8