Bug 1624810 - Add a nsCycleCollectionParticipant flag for JS holders that can have GC things in multiple zones r=mccr8
This adds the flag itself and a means of setting it for a cycle collected class.

The proper way to do this would be to add versions of all the NS_DECL_CYCLE_COLLECTION_CLASS macros that also set this flag but we end up needing to create separate versions for five of these and it's a lot of macro code to add.  Here I added a version of NS_IMPL_CYCLE_COLLECTION_CLASS that sets the flag.  This has the disadvantage that it wouldn't work well if we needed to set the flag on a base class as we'd have to use this macro for every derived class.  However that situation doesn't actually arise (and ideally this flag will bet set on the fewest number of classes possible).

Differential Revision: https://phabricator.services.mozilla.com/D68195

UltraBlame original commit: 4dcdeed2265fe0028fe816097f6cddc62d326977