Bug 1622864 - Avoid sanitizing a name for app-provided engines, and clarify the WebExtension id in the search engine schema. r=daleharvey
makeEngineFromConfig currently initialises SearchEngine with the full name which is then sanitized, only for it to be overriden when calling _initFromMetadata. Being able to pass shortName directly, makes it clearer as to what is happening.

By being explicit about if we're passing shortName or the display name, we can avoid the awkward sanitizeName option.

Also updating the schema to provide validation of the WebExtension Id so that we have the format we expect there (which shortName is based on).

Differential Revision: https://phabricator.services.mozilla.com/D68229

UltraBlame original commit: 69266001850728ca05430fce0cc06ce42c19507c