Bug 1625217: Minor bug fixes to TestCrashHandler service binding; r=geckoview-reviewers,agi
`ServiceConnection.onServiceDisconnected` is not called if
`Context.unbindService` finished cleanly. We should be clearing `mService` in
that case.

Furthermore, if the service did die unexpectedly and
`ServiceConnection.onServiceDisconnected` was called, we need to explicitly
unbind so that Android does not automatically reconnect the binding.

Differential Revision: https://phabricator.services.mozilla.com/D68417

UltraBlame original commit: 5afc157c37555d70287824ea0b490006df28da16