Bug 1618198 part 31 - Turn JSOp whitelist into a blacklist. r=iain
At this point WarpBuilder supports the same JSOps as IonBuilder, except for
JSOp::Eval and JSOp::StrictEval (the IonBuilder code for these ops relies on TI
and we should consider using either a Call IC or MEval calling a VM function).

We now get a compiler warning in WarpOracle and WarpBuilder when a new JSOp is
added but not handled in WarpOracle/WarpBuilder.

WarpOracle is slightly more efficient now because it has one switch-statement
instead of two.

Differential Revision: https://phabricator.services.mozilla.com/D68340

UltraBlame original commit: b085cf5490fe8d7a3dae60a04678b1ccb33cfb31