Bug 1585313 - Fix display of multiple close button on click and drag in network blocking panel r=Honza,bomsy
Click and drag on the closing button activates the `:focus` property
in CSS but doesn't trigger the onclick function. As a result, on a
click and drag event, the close button will remain in focus until
clicked on a vacant space or clicked on it again. When hovering over
other blockied URL, two close buttons appear instead on one (one in
focus, the other from the request on which mouse is hovering)
This commit removes the `:focus` from CSS. A click on the button will
delete the rule in real time. There is no scenario where the button
will need a focus on without removing the blocked URL from the list

Differential Revision: https://phabricator.services.mozilla.com/D68329

UltraBlame original commit: 0617f37534467908d10dc7e284978e1261c74f41