Bug 1616788 - Part 1: Add a hasStoragePermission flag in the LoadInfo. r=dimi,baku
We add a flag 'HasStoragePermission' in the LoadInfo. This flag
represents whether the loading document, for docuemnt loads, or the
loading resource has the storage permission. And this flag would only
get updated in the parent process when opening the channel.

Differential Revision: https://phabricator.services.mozilla.com/D67466

UltraBlame original commit: 739cb622736c88808df633df773ce4a4ca2574ce