Bug 1626137 - Account for tables that are display:block; when looking for the table of a row accessible, r=Jamie
When checking if a parent of a table row counts for a table, take into account that a table might be an ARIAGridAccessible, which remains a generic hyperText, but is still a valid table. Tables with display: block; are such candidates.

Differential Revision: https://phabricator.services.mozilla.com/D68955

UltraBlame original commit: d2b3ff1182052fa4ad2ecdd040eaacafb7d74a89