Bug 1624270 [wpt PR 22388] - [LayoutNG] Add tests for text-overflow: ellipsis with scrolling, a=testonly
Automatic update from web-platform-tests
[LayoutNG] Add tests for text-overflow: ellipsis with scrolling

These set of tests pass in Firefox but not in Chromium.
They check what happens when you scroll an element with
"text-overflow: ellipsis", right now in Chromium
the rest of the content of the element is not visible at all
(once truncated the text ellipsis is not moved).

The solution would be to force a relayout during scrolling
but it seems we prefer to keep current behavior to avoid
performance issues on scrolling.
The behavior is not mandatory according

Both behaviors are allowed in the spec, so the tests are marked
with the "should" flag:
https://drafts.csswg.org/css-overflow/#ellipsis-scrolling

Bug=745905
TEST=external/wpt/css/css-overflow/text-overflow-scroll-001.html
TEST=external/wpt/css/css-overflow/text-overflow-scroll-rtl-001.html
TEST=external/wpt/css/css-overflow/text-overflow-scroll-vertical-lr-001.html
TEST=external/wpt/css/css-overflow/text-overflow-scroll-vertical-lr-rtl-001.html
TEST=external/wpt/css/css-overflow/text-overflow-scroll-vertical-rl-001.html
TEST=external/wpt/css/css-overflow/text-overflow-scroll-vertical-rl-rtl-001.html

Change-Id: I975ca86dbdfedf92875c08736539fb2f33f68c54
Bug: 745905
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2113495
Commit-Queue: Koji Ishii <kojiichromium.org>
Reviewed-by: Koji Ishii <kojiichromium.org>
Cr-Commit-Position: refs/heads/master{#753508}

--

wpt-commits: 86103a7fabf985763c122276a67615da0ced4b3c
wpt-pr: 22388

UltraBlame original commit: c5c828b67e000cfecfbd267cd82a3b82483fd911