Bug 1623159 - Navigate to about:blank before starting pageload tests. r=perftest-reviewers,whimboo,Bebe
This patch fixes an issue in warm pageload tests where the browser doesn't navigate to the page that was requested since the browser was already open on that page. To get around this, we navigate to about:blank, and then to the rquested url.

Differential Revision: https://phabricator.services.mozilla.com/D68428

UltraBlame original commit: 876a51b67aca75334dbe942e425d4eab26919b6b