Bug 1626340 - Fix some mozconfigs where mozconfig.common.override is not at the end r=rstewart
mozconfig.common.override is supposed to be a way for try pushes to change any options that may have been previously set. To do this effectively, it needs to be the last thing in the mozconfig.

Differential Revision: https://phabricator.services.mozilla.com/D69045

UltraBlame original commit: f5447afc575b1ee9c2df80f12cb609c45e40647a