Bug 1616727 - Relax referrer info assertion, r=farre
This assertion would fire incorrectly when the <iframe> element is destroyed,
but the nsDocShell hasn't been destroyed yet, due to it needing to be destroyed
asynchronously.

This isn't an actual problem, as the window will be destroyed either way, so
this relaxes the assertion.

Differential Revision: https://phabricator.services.mozilla.com/D68859

UltraBlame original commit: 32b52f4cb0495b526a89f77503110c6c4131384d