Bug 1624363 - Add GcPrefix opcode space and align encoding with wat update. r=lth
The V8/wabt prototype uses a dedicated prefix opcode for most of the GC
operations. This commit adds this prefix and switches the instructions over
to it.

The encoding for ref/optref is left to its own commit as that implies a larger
change to rename the actual TypeCode.

Differential Revision: https://phabricator.services.mozilla.com/D69407

UltraBlame original commit: 5ee4bec9fb2922757793b3286066da289e41e2b1