Bug 1616353 - Part 5: Always set the opener for chrome windows late, r=kmag
This is OK, as all chrome BrowsingContext instances are always in the same
BrowsingContextGroup, so we don't need to have the initial opener avaliable when
creating the BrowsingContext to determine its BrowsingContextGroup.

This is used in part 7.2 to avoid propagating the opener window around when
creating chrome windows.

Differential Revision: https://phabricator.services.mozilla.com/D67044

UltraBlame original commit: ae60585529692199780c400877001c11e1d7eac9