Bug 1616353 - Part 7.5: Pass nsOpenWindowInfo through Geckoview, r=geckoview-reviewers,esawin,agi
The interface changes which GeckoViewNavigation.jsm is being updated for are
performed in earlier parts. This patch just covers the required changes to
frontend code to propagate this information down to the nsFrameLoader when it is
created.

Differential Revision: https://phabricator.services.mozilla.com/D67054

UltraBlame original commit: bd1706c57d91da8b76a76cf5b25e06c0d6d6ebec