Bug 1616353 - Part 7.4: Pass nsOpenWindowInfo through tabbrowser, r=Gijs
The interface changes which browser.js and tabbrowser.js are being updated for
are performed in earlier parts. This patch just covers the required changes to
frontend code to propagate this information down to the nsFrameLoader when it is
created.

Differential Revision: https://phabricator.services.mozilla.com/D67053

UltraBlame original commit: f3e16dc2311da9d3382b09c6c702f22d2bafa1a1