Bug 1616353 - Part 7.6: Pass nsOpenWindowInfo through the single-site browser, r=Gijs
The interface changes which ssb.jsm is being updated for are performed in
earlier parts. This patch just covers the required changes to frontend code to
propagate this information down to the nsFrameLoader when it is created.

Differential Revision: https://phabricator.services.mozilla.com/D67055

UltraBlame original commit: 39efd9c5c22aa3ce543e515c56cc211948b1cfdf