Bug 1627864 - Fix invalidation of tiles when the shape of the spatial node tree changes. r=nical,Bert,kvark
Previously, primitive dependency checking would invalidate a tile if
the spatial node index for a given primitive changed.

However, if a new display list is sent that changes the shape of
the spatial node tree this may cause unnecessary invalidations.

For example, a new display list that inserts a new spatial node at
the start of the tree could result in spatial node indices being
different, even though the values of the transforms was the same.

This patch changes the invalidation logic for spatial nodes to
compare the transforms by value, rather than index, meaning that
invalidations are avoided if the shape of the spatial tree has
changed, but the values are consistent.

Differential Revision: https://phabricator.services.mozilla.com/D69913

UltraBlame original commit: 498d2204d63739057b626da3239d130c412099e0