Bug 1624552 - P2. Merge reporting channel and tracking channel for ContentBlockingNotifier::OnEvent r=timhuang,baku
The only place is using different the reporting channel and the tracking channel
is ContentBlocking::AllowAccessFor, which use the top-window of the
parent window (reporting) and the parent window (tracking).

In this case, we could remove passing reporting channel because this
will be handled in the parent.

Depends on D68331

Differential Revision: https://phabricator.services.mozilla.com/D69980

UltraBlame original commit: 75beefd331e41a202e652d4427a58a3e3d207762