Bug 1628130: Delete webrender::prim_store::PrimitiveSceneData. r=gw
Most `Internable` implementations give `PrimitiveSceneData` as their
`InternData` associated type, the type of data associated with the handle in the
scene builder thread. However, nothing in the scene builder code, or anywhere in
WebRender, actually uses the contents of `PrimitiveSceneData`, so it can be
replaced with `()` with no effect on the code other than memory savings.

Differential Revision: https://phabricator.services.mozilla.com/D70121

UltraBlame original commit: 6473560b64001edfccf30dacd60db31ad24df3e7