Bug 1627371 - Include shouldSuppressRunOnce in lazy TreatAsRunOnce computation. r=mgaudet
This brings the lazy TreatAsRunOnce flag in sync with the non-lazy one. Note
that although the flag is in-sync when we delazify, a lazy script with a lazy
parent will still be wrong.

Depends on D70287

Differential Revision: https://phabricator.services.mozilla.com/D70288

UltraBlame original commit: 03daacb97b0a5e61cb1757a036f9ad916bd3d3b7