Bug 1628901 - Fix panic caused by calling BeginDraw with empty dirty rect. r=sotaro
Previously, it was possible for a tile that had a valid scroll root
to have an empty valid (and dirty) rect due to the picture cache
clip rect, in some situations.

This could result in the tile not being tagged as off-screen, which
means it is added to the queue of tiles to be updated. On most
platforms this is benign, but the BeginDraw method of DirectComposition
fails if the dirty rect is empty.

This patch fixes the logic so that tiles that meet these conditions
are correctly tagged as not visible, and skipped from update queue.

Differential Revision: https://phabricator.services.mozilla.com/D70616

UltraBlame original commit: 73632227ba00df584e8f6fcc1191e6dfa13fc5fd