Bug 1627520 - Use keydown instead of keypress for autocomplete event listeners. r=mak,Gijs
This is so that we can avoid needing mozSystemGroup (to get keypress for
non-printable keys), which in turn prevents racing with the native key event
listeners, see the second patch in bug 1624657.

This turned out to be a bit tricky, because we need to guarantee the ordering of
the search one-offs handling in the searchbar with the usual autocomplete-input
handling. We could try to move this to the popup subclass but this is already a
bigger patch than what I'd like.

We can also revert the customElements.js change I did in bug 1624657, as the bug
is no longer relevant.

Differential Revision: https://phabricator.services.mozilla.com/D69743

UltraBlame original commit: e46624dd8da6de6d38fc7df213a8a45ea861070c