Bug 1627373 - Use SharedContext source extents instead of passing around line and columns r=tcampbell
Outside of powering BytecodeSection's source notes generation, the line and
column numbers were almost entirely unused, except to power a single assertion
that lineno and column agreed between the stencil (BCE derived) and script.

However, by switching to using the SharedContext extents, we lose the ability
to always assert they're the same, because in the old system dynamically
allocated functions used externally provided column info that disagrees with
the parser result. In this patch the script still gets the same column
number as before, but we can't assert that the value is the same as what you
could get out of the BCE, because the BCE number is actually more accurate
than the script.

Fixing the script column should be followup.

Differential Revision: https://phabricator.services.mozilla.com/D70411

UltraBlame original commit: 17dadcf8617fba804cf7f01d5747ae586526b624