Bug 1627125 Part 6 - Redesign the logic that computes flex container's final size with "box-decoration-break: clone" considered. r=dholbert
By passing unconstrained aAvailableBSizeForContent to DoFlexLayout, we
can get the flex container's desired content-box size as if there's no
pagination. We could then use this content-box size and
maxBlockEndEdgeOfChildren to rewrite the logic to determine flex
container's final size.

I still maintain the compatibility with block frame that if the flex
container is running out of block-size, it can still become incomplete
and request an zero block-size continuation. One comment in block frame
feels this is wrong, but fixing it is not in the scope of this bug.
https://searchfox.org/mozilla-central/rev/4ccefc3181f9d237ef4ca8bd17b4e7c101ddf7b5/layout/generic/nsBlockFrame.cpp#7715-7718

Differential Revision: https://phabricator.services.mozilla.com/D69474

UltraBlame original commit: c473fe50e3fa1d0d8327eb7b715406a13317988b