Bug 1630473 - temporarily stash the client certificate chain so NSS can send it to the server r=kjacobs,bbeurdouche
When sending a client certificate to a server in a TLS handshake, one of the
certificates in the chain should be issued by one of the issuers indicated in
the server's certificate_authorities list in the certificate request message.
The client auth data callback doesn't provide a way to specify this chain
directly - NSS builds it itself. This means that certificates known to gecko
but not NSS won't be included in the chain. This patch stashes the necessary
certificates temporarily so that NSS can find them and send them to the server.

Differential Revision: https://phabricator.services.mozilla.com/D71368

UltraBlame original commit: fc408ad127bc67125d8eb4785cb1c31276957d32