Bug 1628124 - Do not run minidump analysis when we've instructed the browser to shut down after a content process crash r=mconley
When MOZ_CRASHREPORTER_SHUTDOWN is set the main process will shut down as soon
as it detects a content process crash. We usually run the minidump-analyzer
tool for every content process crash but we stop it if we detect we're
shutting down. This is done to prevent hangs because minidump analysis may
take a while.

This patch checks for MOZ_CRASHREPORTER_SHUTDOWN and does not start the
minidump-analyzer at all if it's set. Since we know we'll be shutting down
anyway there's no point in try to run it.

I took the liberty of refactoring some of the surrounding code for better
readability.

Differential Revision: https://phabricator.services.mozilla.com/D71416

UltraBlame original commit: 2ac3889b41cf5790ae4c8093a04a4999fb7ab420