Bug 1622897 - Fix intermitten by waiting for the right time to check the result. r=padenot
First update the previous test,`TestAudioTrackGraph.DifferentDeviceIDs`, to make sure that has been finished before starting the current one. The reason is that when the `cubeb_stream_init` method, of the previous test, is executed (asynchronously) late, the verified test results, of the current test, are wrong (it gets the values of the previous test). Second, on the currently failing test, use the promise returned from the new `MTG::NotifyWhenDeviceStarted()` method, that it is guaranteed to be resolved after the `cubeb_stream_init()` so the correct/expected device id will be found and compared. Finally, create a basic logging mechanism in MockCubeb for easier debugging.

Differential Revision: https://phabricator.services.mozilla.com/D71943

UltraBlame original commit: 2b6cbcc81bdd54d76ed3c1e2a40adb28798be256