Bug 1631512 - Trigger a slice to continue the GC as soon as background sweeping finishes r=sfink
The idea here is to set a flag whenever GC yeilds to the mutator while waiting for a background task (either sweeping or decommiting). Then when the task completes we check the flag and request a GC slice if it's set.

Differential Revision: https://phabricator.services.mozilla.com/D71770

UltraBlame original commit: 3622444e4081c8f8c3a6b21f8f4cfd18b55ff2b4