Bug 1631234 - Part 2: Moving the setting of the 'HasStoragePermission' before the first check of the storage access in nsGlobalWindowOuter::SetNewDocument(). r=baku
Because we would check the storage access permission inside
nsGlobalWindowInner::ExecutionReady(), so we need to set the
'HasStoragePermission' flag before we call it in
nsGlobalWindowOuter::SetNewDocument().

In addition, we check the assertion in the
ClientSource::WindowExecutionReady() back to a diagnostic assertion.

Depends on D71564

Differential Revision: https://phabricator.services.mozilla.com/D71565

UltraBlame original commit: a002677cd9953037c569062b1cbe72cbb773e43e