Bug 1632641 - Fix flakyness in browser_ext_tabs_zoom.js r=aswan
This test was already a bit flaky it seems, see bug 1630399 for example.

But it has become much worse with my Fission zoom changes in bug 1612068, which
made the FullZoomChange events that this API relies on more async.

Properly wait for an event for the right tab instead of assuming that the last
zoom event was for the tab we happen to be waiting on.

Differential Revision: https://phabricator.services.mozilla.com/D72209

UltraBlame original commit: 93842a33d01a3b1ced9c2ba6c13320eb3625b289