Bug 1629893, add graphviz to lint image to fix docs, r=sylvestre
The `sphinx.ext.graphviz` extension requires `dot` to be installed.
That's default on rtd, which is why we use it in the docs.
Adding this to the lint image as we use that for document building,
and it shouldn't hurt.

Differential Revision: https://phabricator.services.mozilla.com/D72395

UltraBlame original commit: 81d4ad686888d8ccf6bd408feaa31a47e59da525