Bug 1631671 - [mozdevice] Handle Scoped storage in Android 10+, r=gbrown.
On a local Pixel 2 running Android 10, launching Geckoview test app via the
shell with the profile on external storage hangs and causes an ADBTimeoutError.
Using a profile in /data/local/tests does not exhibit this behavior.

In addition, attempting to push a directory tree directly to internal
storage can fail with "remote secure_mkdirs failed" errors when pushing
directories containing subdirectories to internal storage on /data.

The subdirectories were created with insufficient permissions for the
contents of the directories to be pushed. We can work around this problem
by catching the failure and falling back to using adb push to push the
source directory tree to a temporary location on the sdcard which does
not exhibit the error and then copying the temporary location to the
desired final destination.

These changes are necessary due to the change in Scoped Storage [1]
in Android 10 and later.

Since storing profiles on the sdcard is no longer supported the choices
for possible test roots has been reduced to using internal storage.
Since geckoview expects to be able to store content under /data/local/tmp
the default test root has been changed to /data/local/tmp/tests.

By default, the test root will be shared by all instances of ADBDevice
unless the share_test_root argument is set to False in the initializer.
This allows the test root to be defined globally with out requiring
each instance of ADBDevice to be modified.

The test_root default initial value should be None, not empty string.
The ADBDevice.test_root property is now read/write to support the new
shared test_root functionality.

Android emulators have exhibited an intermittent failure to clear logcat.
A ride-along change to retry the log clearing is included.

adb pull can fail if not all of the files have permission to be read.
This situation can appear when a profile is located on internal storage
and an app such as geckoview writes files to it. They will be readable
only by the app until the permissions are changed. Work around this
by calling chmod -R on any thing that is to be pulled from the device.


Bump mozdevice version to 3.1.2.

[1] https://developer.android.com/training/data-storage#scoped-storage

Differential Revision: https://phabricator.services.mozilla.com/D71722

UltraBlame original commit: 3b5cbdbdb32eadf3783e757753aebb94dfdd6af3