Bug 1627072 - Runs pip-compile in same directory as temporary requirements file r=rstewart
pip-compile can annotate each requirement with a reason why it is included (e.g.: transitive dependency, or depended-on directly).

When annotating direct dependencies, it notes it as "via -r <direct path to requirements.in file>".
Since we were using a temporary directory, the path of the directory was being included, making the output non-deterministic.

This change ensures that we run pip-compile in the same working directory as the temporary requirements file, enabling
the annotations to be deterministic: "via -r requirements-mach-vendor-python.in".

Differential Revision: https://phabricator.services.mozilla.com/D72181

UltraBlame original commit: 5a93d460a035b076591abf6f45fd2b670f286e05