Bug 1607801 - Pre-define ChromeUtils.import paths; r=julienw
This patch takes the approach from mossop on pre-defining ChromeUtils.import
paths, so that they don't need to be defined where they are used. Perhaps in
the future we could automate this more, but for now this will make the current
approach more ergonomic for consumers of the types.

Differential Revision: https://phabricator.services.mozilla.com/D59206

UltraBlame original commit: 8dfe8cb5a8311dc6062c8d9a421dbbf65bc5050e