Bug 1633437 - Support for test metadata r=acreskey
This patch adds support for tests metadata. A test script parser is added as
well as a new "doc" flavor that can be used to display the script info in the
command line. This parser will be the basis for building automated docs and
scripts verifications if we want to do this.

Differential Revision: https://phabricator.services.mozilla.com/D72800

UltraBlame original commit: c4a72c937045df517cbcb8e02307245bbbbe35b9