Bug 1632820 - Only emit bailAfter instrumentation in debug builds when the testing function is used. r=tcampbell
This mechanism added a lot of machine code to Ion JIT code, making it hard to analyze
codegen. With this change we only emit this when necessary.

Also move the addressOfIonBailAfterCounter method from CompileZone to
CompileRuntime because the field is in JitRuntime.

Differential Revision: https://phabricator.services.mozilla.com/D72376

UltraBlame original commit: 3762f151cab2d54d4db9268cf14d91e0bdd8a1f7