Bug 1612376 - P4. ShouldAllowAccessFor use ContentBlocking::HasStorageAccessGranted r=timhuang,baku
There are two places where we save storage permission:
1. LoadInfo hasStoragePermission attribute
2. mStorageAccessGranted in nsPIDOMWindowInner

For LoadInfo.hasStoragePermission, it is set during channel creation and
its value remains the same even when the storage permission is granted
afterward.

The updated storage permission for a window is saved in
mStorageAccessGranted, which has a different meaning for fission and
non-fission mode.
In non-fission mode, mStorageAccessGranted is saved in the top-level
window and it is an array containing all tracking subframes that
are allowed to access storage.

In fission mode, mStorageAccessGranted is set in individual tracking
windows that we have granted its storage permission. Although it works
like a boolean flag in fission, we still keep using an array to compatible
with the use case in non-fission mode.

Depends on D71984

Differential Revision: https://phabricator.services.mozilla.com/D71985

UltraBlame original commit: e2eb63c3ce5caf55069da424079d9ad293bf5999