Bug 1631799 - Fix order of command and its result message when server and client don't have the same clock time. r=jdescottes.
When remote-debugging a device, it might happen that the device clock
is not exactly at the same time as the client clock. In such case, we
may end up in cases where the result of a command as an older timestamp,
which means we could have in the console output a result that was displayed
before the command message that triggered it.
In order to fix this, we add a `startTime` property on the evaluateJSAsync
response, that we can then check on the client. If the server time is
older than the client time, then we remove the existing command message
and add a new one, with the server timestamp.

We could have waited the result of the command before displaying the
command message (so we don't have to add a message, and then remove it
when the timestamps don't match), but the UI wouldn't feel snappy, and
even worse in case of long execution times (e.g. a top-level await fetch command).

Differential Revision: https://phabricator.services.mozilla.com/D72705

UltraBlame original commit: 6feb116adffe41e0ce0d27cb15bb81c10d764e6e