Bug 1624380: upgrades brew packages if they're out-of-date r=rstewart
`./mach bootstrap` prepares your environment for a build and installs missing "brew" packages".
However, if a package is installed but out-of-date, it was being ignored by the bootstrapping logic.

This change ensures that `brew update` is run, and updates any out-of-date packages that are requested.

Differential Revision: https://phabricator.services.mozilla.com/D72956

UltraBlame original commit: 306fcf2f019450fe89ae62c91c3818b1d2a2c73a