Bug 1632918 - Postpone update check until the user triggers the update intervention. r=harry
As a quick fix to bug 1632918, don't check for app updates until the user
triggers the update intervention. There will still be some jank, but it'll be
postponed until the user types a phrase that triggers the intervention, which is
much better than at the first urlbar use. To avoid the problem where we wouldn't
show an intervention at all the first time the update intervention is triggered
since the updater status in that case will be CHECKING, add an updater listener
to wait for the check to finish and add our result then.

We should look into a better long-term fix, like making the updater's IO async
or not checking for updates at all inside urlbar.

Differential Revision: https://phabricator.services.mozilla.com/D73001

UltraBlame original commit: 86ddde685cfa2a8963fe7a64780800d1377607a5